Struct google_compute1::api::BackendService

pub struct BackendService {
    pub affinity_cookie_ttl_sec: Option<i32>,
    pub backends: Option<Vec<Backend>>,
    pub cdn_policy: Option<BackendServiceCdnPolicy>,
    pub circuit_breakers: Option<CircuitBreakers>,
    pub connection_draining: Option<ConnectionDraining>,
    pub consistent_hash: Option<ConsistentHashLoadBalancerSettings>,
    pub creation_timestamp: Option<String>,
    pub custom_request_headers: Option<Vec<String>>,
    pub custom_response_headers: Option<Vec<String>>,
    pub description: Option<String>,
    pub enable_cdn: Option<bool>,
    pub failover_policy: Option<BackendServiceFailoverPolicy>,
    pub fingerprint: Option<String>,
    pub health_checks: Option<Vec<String>>,
    pub iap: Option<BackendServiceIAP>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub load_balancing_scheme: Option<String>,
    pub locality_lb_policy: Option<String>,
    pub log_config: Option<BackendServiceLogConfig>,
    pub max_stream_duration: Option<Duration>,
    pub name: Option<String>,
    pub network: Option<String>,
    pub outlier_detection: Option<OutlierDetection>,
    pub port: Option<i32>,
    pub port_name: Option<String>,
    pub protocol: Option<String>,
    pub region: Option<String>,
    pub security_policy: Option<String>,
    pub security_settings: Option<SecuritySettings>,
    pub self_link: Option<String>,
    pub session_affinity: Option<String>,
    pub timeout_sec: Option<i32>,
}

Represents a Backend Service resource.

A backend service defines how Google Cloud load balancers distribute traffic. The backend service configuration contains a set of values, such as the protocol used to connect to backends, various distribution and session settings, health checks, and timeouts. These settings provide fine-grained control over how your load balancer behaves. Most of the settings have default values that allow for easy configuration if you need to get started quickly.

Backend services in Google Compute Engine can be either regionally or globally scoped.

Global * Regional

For more information, see Backend Services.

(== resource_for {$api_version}.backendService ==)

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

add signed url key backend services (none)

aggregated list backend services (none)

delete backend services (none)

delete signed url key backend services (none)

get backend services (response)

get health backend services (none)

insert backend services (request)

list backend services (none)

patch backend services (request)

set security policy backend services (none)

update backend services (request)

get region backend services (response)

insert region backend services (request)

patch region backend services (request)

update region backend services (request)

Fields
affinity_cookie_ttl_sec: Option<i32>

Lifetime of cookies in seconds. Only applicable if the loadBalancingScheme is EXTERNAL, INTERNAL_SELF_MANAGED, or INTERNAL_MANAGED, the protocol is HTTP or HTTPS, and the sessionAffinity is GENERATED_COOKIE, or HTTP_COOKIE.

If set to 0, the cookie is non-persistent and lasts only until the end of the browser session (or equivalent). The maximum allowed value is one day (86,400).

Not supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true.

backends: Option<Vec<Backend>>

The list of backends that serve this BackendService.

cdn_policy: Option<BackendServiceCdnPolicy>

Cloud CDN configuration for this BackendService. Only available for external HTTP(S) Load Balancing.

circuit_breakers: Option<CircuitBreakers>

Settings controlling the volume of connections to a backend service. If not set, this feature is considered disabled.

This field is applicable to either:

A regional backend service with the service_protocol set to HTTP, HTTPS, or HTTP2, and load_balancing_scheme set to INTERNAL_MANAGED.

A global backend service with the load_balancing_scheme set to INTERNAL_SELF_MANAGED.

Not supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true.

connection_draining: Option<ConnectionDraining>

no description provided

consistent_hash: Option<ConsistentHashLoadBalancerSettings>

Consistent Hash-based load balancing can be used to provide soft session affinity based on HTTP headers, cookies or other properties. This load balancing policy is applicable only for HTTP connections. The affinity to a particular destination host will be lost when one or more hosts are added/removed from the destination service. This field specifies parameters that control consistent hashing. This field is only applicable when localityLbPolicy is set to MAGLEV or RING_HASH.

This field is applicable to either:

A regional backend service with the service_protocol set to HTTP, HTTPS, or HTTP2, and load_balancing_scheme set to INTERNAL_MANAGED.

A global backend service with the load_balancing_scheme set to INTERNAL_SELF_MANAGED.

Not supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

custom_request_headers: Option<Vec<String>>

Headers that the HTTP/S load balancer should add to proxied requests.

custom_response_headers: Option<Vec<String>>

Headers that the HTTP/S load balancer should add to proxied responses.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

enable_cdn: Option<bool>

If true, enables Cloud CDN for the backend service. Only applicable if the loadBalancingScheme is EXTERNAL and the protocol is HTTP or HTTPS.

failover_policy: Option<BackendServiceFailoverPolicy>

Applicable only to Failover for Internal TCP/UDP Load Balancing and Network Load Balancing. Requires at least one backend instance group to be defined as a backup (failover) backend.

fingerprint: Option<String>

Fingerprint of this resource. A hash of the contents stored in this object. This field is used in optimistic locking. This field will be ignored when inserting a BackendService. An up-to-date fingerprint must be provided in order to update the BackendService, otherwise the request will fail with error 412 conditionNotMet.

To see the latest fingerprint, make a get() request to retrieve a BackendService.

health_checks: Option<Vec<String>>

The list of URLs to the healthChecks, httpHealthChecks (legacy), or httpsHealthChecks (legacy) resource for health checking this backend service. Not all backend services support legacy health checks. See Load balancer guide. Currently, at most one health check can be specified for each backend service. Backend services with instance group or zonal NEG backends must have a health check. Backend services with internet or serverless NEG backends must not have a health check.

iap: Option<BackendServiceIAP>

The configurations for Identity-Aware Proxy on this resource. Not available for Internal TCP/UDP Load Balancing and Network Load Balancing.

id: Option<String>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

kind: Option<String>

[Output Only] Type of resource. Always compute#backendService for backend services.

load_balancing_scheme: Option<String>

Specifies the load balancer type. Choose EXTERNAL for external HTTP(S), SSL Proxy, TCP Proxy and Network Load Balancing. Choose INTERNAL for Internal TCP/UDP Load Balancing. Choose INTERNAL_MANAGED for Internal HTTP(S) Load Balancing. INTERNAL_SELF_MANAGED for Traffic Director. A backend service created for one type of load balancer cannot be used with another. For more information, refer to Choosing a load balancer.

locality_lb_policy: Option<String>

The load balancing algorithm used within the scope of the locality. The possible values are:

ROUND_ROBIN: This is a simple policy in which each healthy backend is selected in round robin order. This is the default.

LEAST_REQUEST: An O(1) algorithm which selects two random healthy hosts and picks the host which has fewer active requests.

RING_HASH: The ring/modulo hash load balancer implements consistent hashing to backends. The algorithm has the property that the addition/removal of a host from a set of N hosts only affects 1/N of the requests.

RANDOM: The load balancer selects a random healthy host.

ORIGINAL_DESTINATION: Backend host is selected based on the client connection metadata, i.e., connections are opened to the same address as the destination address of the incoming connection before the connection was redirected to the load balancer.

MAGLEV: used as a drop in replacement for the ring hash load balancer. Maglev is not as stable as ring hash but has faster table lookup build times and host selection times. For more information about Maglev, see https://ai.google/research/pubs/pub44824

This field is applicable to either:

A regional backend service with the service_protocol set to HTTP, HTTPS, or HTTP2, and load_balancing_scheme set to INTERNAL_MANAGED.

A global backend service with the load_balancing_scheme set to INTERNAL_SELF_MANAGED.

If sessionAffinity is not NONE, and this field is not set to MAGLEV or RING_HASH, session affinity settings will not take effect.

Only the default ROUND_ROBIN policy is supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true.

log_config: Option<BackendServiceLogConfig>

This field denotes the logging options for the load balancer traffic served by this backend service. If logging is enabled, logs will be exported to Stackdriver.

max_stream_duration: Option<Duration>

Specifies the default maximum duration (timeout) for streams to this service. Duration is computed from the beginning of the stream until the response has been completely processed, including all retries. A stream that does not complete in this duration is closed. If not specified, there will be no timeout limit, i.e. the maximum duration is infinite. This field is only allowed when the loadBalancingScheme of the backend service is INTERNAL_SELF_MANAGED.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z]([-a-z0-9]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

network: Option<String>

The URL of the network to which this backend service belongs. This field can only be specified when the load balancing scheme is set to INTERNAL.

outlier_detection: Option<OutlierDetection>

Settings controlling the eviction of unhealthy hosts from the load balancing pool for the backend service. If not set, this feature is considered disabled.

This field is applicable to either:

A regional backend service with the service_protocol set to HTTP, HTTPS, or HTTP2, and load_balancing_scheme set to INTERNAL_MANAGED.

A global backend service with the load_balancing_scheme set to INTERNAL_SELF_MANAGED.

Not supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true.

port: Option<i32>

Deprecated in favor of portName. The TCP port to connect on the backend. The default value is 80.

Backend services for Internal TCP/UDP Load Balancing and Network Load Balancing require you omit port.

port_name: Option<String>

A named port on a backend instance group representing the port for communication to the backend VMs in that group. Required when the loadBalancingScheme is EXTERNAL (except Network Load Balancing), INTERNAL_MANAGED, or INTERNAL_SELF_MANAGED and the backends are instance groups. The named port must be defined on each backend instance group. This parameter has no meaning if the backends are NEGs.

Backend services for Internal TCP/UDP Load Balancing and Network Load Balancing require you omit port_name.

protocol: Option<String>

The protocol this BackendService uses to communicate with backends.

Possible values are HTTP, HTTPS, HTTP2, TCP, SSL, UDP or GRPC. depending on the chosen load balancer or Traffic Director configuration. Refer to the documentation for the load balancer or for Traffic Director for more information.

Must be set to GRPC when the backend service is referenced by a URL map that is bound to target gRPC proxy.

region: Option<String>

[Output Only] URL of the region where the regional backend service resides. This field is not applicable to global backend services. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

security_policy: Option<String>

[Output Only] The resource URL for the security policy associated with this backend service.

security_settings: Option<SecuritySettings>

This field specifies the security policy that applies to this backend service. This field is applicable to either:

A regional backend service with the service_protocol set to HTTP, HTTPS, or HTTP2, and load_balancing_scheme set to INTERNAL_MANAGED.

A global backend service with the load_balancing_scheme set to INTERNAL_SELF_MANAGED.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

session_affinity: Option<String>

Type of session affinity to use. The default is NONE.

When the loadBalancingScheme is EXTERNAL: * For Network Load Balancing, the possible values are NONE, CLIENT_IP, CLIENT_IP_PROTO, or CLIENT_IP_PORT_PROTO. * For all other load balancers that use loadBalancingScheme=EXTERNAL, the possible values are NONE, CLIENT_IP, or GENERATED_COOKIE. * You can use GENERATED_COOKIE if the protocol is HTTP, HTTP2, or HTTPS.

When the loadBalancingScheme is INTERNAL, possible values are NONE, CLIENT_IP, CLIENT_IP_PROTO, or CLIENT_IP_PORT_PROTO.

When the loadBalancingScheme is INTERNAL_SELF_MANAGED, or INTERNAL_MANAGED, possible values are NONE, CLIENT_IP, GENERATED_COOKIE, HEADER_FIELD, or HTTP_COOKIE.

Not supported when the backend service is referenced by a URL map that is bound to target gRPC proxy that has validateForProxyless field set to true.

timeout_sec: Option<i32>

The backend service timeout has a different meaning depending on the type of load balancer. For more information see, Backend service settings The default is 30 seconds. The full range of timeout values allowed is 1 - 2,147,483,647 seconds.