Struct google_compute1::api::Compute

pub struct Compute<C> { /* fields omitted */ }

Central instance to access all Compute related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
use compute1::api::Instance;
use compute1::{Result, Error};
use std::default::Default;
use oauth2;
use compute1::Compute;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: oauth2::ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Compute::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Instance::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.instances().update(req, "project", "zone", "instance")
             .request_id("gubergren")
             .most_disruptive_allowed_action("eos")
             .minimal_action("dolor")
             .doit().await;
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::Io(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Implementations

impl<'a, C> Compute<C> where
C: BorrowMut<Client<HttpsConnector<HttpConnector>, Body>>,

pub fn new(
client: C,
authenticator: Authenticator<HttpsConnector<HttpConnector>>
) -> Compute<C>

pub fn accelerator_types(&'a self) -> AcceleratorTypeMethods<'a, C>

pub fn addresses(&'a self) -> AddresseMethods<'a, C>

pub fn autoscalers(&'a self) -> AutoscalerMethods<'a, C>

pub fn backend_buckets(&'a self) -> BackendBucketMethods<'a, C>

pub fn backend_services(&'a self) -> BackendServiceMethods<'a, C>

pub fn disk_types(&'a self) -> DiskTypeMethods<'a, C>

pub fn disks(&'a self) -> DiskMethods<'a, C>

pub fn external_vpn_gateways(&'a self) -> ExternalVpnGatewayMethods<'a, C>

pub fn firewall_policies(&'a self) -> FirewallPolicyMethods<'a, C>

pub fn firewalls(&'a self) -> FirewallMethods<'a, C>

pub fn forwarding_rules(&'a self) -> ForwardingRuleMethods<'a, C>

pub fn global_addresses(&'a self) -> GlobalAddresseMethods<'a, C>

pub fn global_forwarding_rules(&'a self) -> GlobalForwardingRuleMethods<'a, C>

pub fn global_network_endpoint_groups(
&'a self
) -> GlobalNetworkEndpointGroupMethods<'a, C>

pub fn global_operations(&'a self) -> GlobalOperationMethods<'a, C>

pub fn global_organization_operations(
&'a self
) -> GlobalOrganizationOperationMethods<'a, C>

pub fn global_public_delegated_prefixes(
&'a self
) -> GlobalPublicDelegatedPrefixeMethods<'a, C>

pub fn health_checks(&'a self) -> HealthCheckMethods<'a, C>

pub fn http_health_checks(&'a self) -> HttpHealthCheckMethods<'a, C>

pub fn https_health_checks(&'a self) -> HttpsHealthCheckMethods<'a, C>

pub fn images(&'a self) -> ImageMethods<'a, C>

pub fn instance_group_managers(&'a self) -> InstanceGroupManagerMethods<'a, C>

pub fn instance_groups(&'a self) -> InstanceGroupMethods<'a, C>

pub fn instance_templates(&'a self) -> InstanceTemplateMethods<'a, C>

pub fn instances(&'a self) -> InstanceMethods<'a, C>

pub fn interconnect_attachments(
&'a self
) -> InterconnectAttachmentMethods<'a, C>

pub fn interconnect_locations(&'a self) -> InterconnectLocationMethods<'a, C>

pub fn interconnects(&'a self) -> InterconnectMethods<'a, C>

pub fn license_codes(&'a self) -> LicenseCodeMethods<'a, C>

pub fn licenses(&'a self) -> LicenseMethods<'a, C>

pub fn machine_types(&'a self) -> MachineTypeMethods<'a, C>

pub fn network_endpoint_groups(&'a self) -> NetworkEndpointGroupMethods<'a, C>

pub fn networks(&'a self) -> NetworkMethods<'a, C>

pub fn node_groups(&'a self) -> NodeGroupMethods<'a, C>

pub fn node_templates(&'a self) -> NodeTemplateMethods<'a, C>

pub fn node_types(&'a self) -> NodeTypeMethods<'a, C>

pub fn packet_mirrorings(&'a self) -> PacketMirroringMethods<'a, C>

pub fn projects(&'a self) -> ProjectMethods<'a, C>

pub fn public_advertised_prefixes(
&'a self
) -> PublicAdvertisedPrefixeMethods<'a, C>

pub fn public_delegated_prefixes(
&'a self
) -> PublicDelegatedPrefixeMethods<'a, C>

pub fn region_autoscalers(&'a self) -> RegionAutoscalerMethods<'a, C>

pub fn region_backend_services(&'a self) -> RegionBackendServiceMethods<'a, C>

pub fn region_commitments(&'a self) -> RegionCommitmentMethods<'a, C>

pub fn region_disk_types(&'a self) -> RegionDiskTypeMethods<'a, C>

pub fn region_disks(&'a self) -> RegionDiskMethods<'a, C>

pub fn region_health_check_services(
&'a self
) -> RegionHealthCheckServiceMethods<'a, C>

pub fn region_health_checks(&'a self) -> RegionHealthCheckMethods<'a, C>

pub fn region_instance_group_managers(
&'a self
) -> RegionInstanceGroupManagerMethods<'a, C>

pub fn region_instance_groups(&'a self) -> RegionInstanceGroupMethods<'a, C>

pub fn region_instances(&'a self) -> RegionInstanceMethods<'a, C>

pub fn region_network_endpoint_groups(
&'a self
) -> RegionNetworkEndpointGroupMethods<'a, C>

pub fn region_notification_endpoints(
&'a self
) -> RegionNotificationEndpointMethods<'a, C>

pub fn region_operations(&'a self) -> RegionOperationMethods<'a, C>

pub fn region_ssl_certificates(&'a self) -> RegionSslCertificateMethods<'a, C>

pub fn region_target_http_proxies(
&'a self
) -> RegionTargetHttpProxyMethods<'a, C>

pub fn region_target_https_proxies(
&'a self
) -> RegionTargetHttpsProxyMethods<'a, C>

pub fn region_url_maps(&'a self) -> RegionUrlMapMethods<'a, C>

pub fn regions(&'a self) -> RegionMethods<'a, C>

pub fn reservations(&'a self) -> ReservationMethods<'a, C>

pub fn resource_policies(&'a self) -> ResourcePolicyMethods<'a, C>

pub fn routers(&'a self) -> RouterMethods<'a, C>

pub fn routes(&'a self) -> RouteMethods<'a, C>

pub fn security_policies(&'a self) -> SecurityPolicyMethods<'a, C>

pub fn snapshots(&'a self) -> SnapshotMethods<'a, C>

pub fn ssl_certificates(&'a self) -> SslCertificateMethods<'a, C>

pub fn ssl_policies(&'a self) -> SslPolicyMethods<'a, C>

pub fn subnetworks(&'a self) -> SubnetworkMethods<'a, C>

pub fn target_grpc_proxies(&'a self) -> TargetGrpcProxyMethods<'a, C>

pub fn target_http_proxies(&'a self) -> TargetHttpProxyMethods<'a, C>

pub fn target_https_proxies(&'a self) -> TargetHttpsProxyMethods<'a, C>

pub fn target_instances(&'a self) -> TargetInstanceMethods<'a, C>

pub fn target_pools(&'a self) -> TargetPoolMethods<'a, C>

pub fn target_ssl_proxies(&'a self) -> TargetSslProxyMethods<'a, C>

pub fn target_tcp_proxies(&'a self) -> TargetTcpProxyMethods<'a, C>

pub fn target_vpn_gateways(&'a self) -> TargetVpnGatewayMethods<'a, C>

pub fn url_maps(&'a self) -> UrlMapMethods<'a, C>

pub fn vpn_gateways(&'a self) -> VpnGatewayMethods<'a, C>

pub fn vpn_tunnels(&'a self) -> VpnTunnelMethods<'a, C>

pub fn zone_operations(&'a self) -> ZoneOperationMethods<'a, C>

pub fn zones(&'a self) -> ZoneMethods<'a, C>

pub fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/2.0.0.

Returns the previously set user-agent.

pub fn base_url(&mut self, new_base_url: String) -> String

Set the base url to use in all requests to the server. It defaults to https://compute.googleapis.com/compute/v1/.

Returns the previously set base url.

pub fn root_url(&mut self, new_root_url: String) -> String

Set the root url to use in all requests to the server. It defaults to https://compute.googleapis.com/.

Returns the previously set root url.