Struct google_compute1::api::Firewall

pub struct Firewall {
    pub allowed: Option<Vec<FirewallAllowed>>,
    pub creation_timestamp: Option<String>,
    pub denied: Option<Vec<FirewallDenied>>,
    pub description: Option<String>,
    pub destination_ranges: Option<Vec<String>>,
    pub direction: Option<String>,
    pub disabled: Option<bool>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub log_config: Option<FirewallLogConfig>,
    pub name: Option<String>,
    pub network: Option<String>,
    pub priority: Option<i32>,
    pub self_link: Option<String>,
    pub source_ranges: Option<Vec<String>>,
    pub source_service_accounts: Option<Vec<String>>,
    pub source_tags: Option<Vec<String>>,
    pub target_service_accounts: Option<Vec<String>>,
    pub target_tags: Option<Vec<String>>,
}

Represents a Firewall Rule resource.

Firewall rules allow or deny ingress traffic to, and egress traffic from your instances. For more information, read Firewall rules.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete firewalls (none)

get firewalls (response)

insert firewalls (request)

list firewalls (none)

patch firewalls (request)

update firewalls (request)

Fields
allowed: Option<Vec<FirewallAllowed>>

The list of ALLOW rules specified by this firewall. Each rule specifies a protocol and port-range tuple that describes a permitted connection.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

denied: Option<Vec<FirewallDenied>>

The list of DENY rules specified by this firewall. Each rule specifies a protocol and port-range tuple that describes a denied connection.

description: Option<String>

An optional description of this resource. Provide this field when you create the resource.

destination_ranges: Option<Vec<String>>

If destination ranges are specified, the firewall rule applies only to traffic that has destination IP address in these ranges. These ranges must be expressed in CIDR format. Only IPv4 is supported.

direction: Option<String>

Direction of traffic to which this firewall applies, either INGRESS or EGRESS. The default is INGRESS. For INGRESS traffic, you cannot specify the destinationRanges field, and for EGRESS traffic, you cannot specify the sourceRanges or sourceTags fields.

disabled: Option<bool>

Denotes whether the firewall rule is disabled. When set to true, the firewall rule is not enforced and the network behaves as if it did not exist. If this is unspecified, the firewall rule will be enabled.

id: Option<String>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

kind: Option<String>

[Output Only] Type of the resource. Always compute#firewall for firewall rules.

log_config: Option<FirewallLogConfig>

This field denotes the logging options for a particular firewall rule. If logging is enabled, logs will be exported to Cloud Logging.

name: Option<String>

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression `a-z?. The first character must be a lowercase letter, and all following characters (except for the last character) must be a dash, lowercase letter, or digit. The last character must be a lowercase letter or digit.

network: Option<String>

URL of the network resource for this firewall rule. If not specified when creating a firewall rule, the default network is used: global/networks/default If you choose to specify this field, you can specify the network as a full or partial URL. For example, the following are all valid URLs:

https://www.googleapis.com/compute/v1/projects/myproject/global/networks/my-network

projects/myproject/global/networks/my-network

global/networks/default

priority: Option<i32>

Priority for this rule. This is an integer between 0 and 65535, both inclusive. The default value is 1000. Relative priorities determine which rule takes effect if multiple rules apply. Lower values indicate higher priority. For example, a rule with priority 0 has higher precedence than a rule with priority 1. DENY rules take precedence over ALLOW rules if they have equal priority. Note that VPC networks have implied rules with a priority of 65535. To avoid conflicts with the implied rules, use a priority number less than 65535.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

source_ranges: Option<Vec<String>>

If source ranges are specified, the firewall rule applies only to traffic that has a source IP address in these ranges. These ranges must be expressed in CIDR format. One or both of sourceRanges and sourceTags may be set. If both fields are set, the rule applies to traffic that has a source IP address within sourceRanges OR a source IP from a resource with a matching tag listed in the sourceTags field. The connection does not need to match both fields for the rule to apply. Only IPv4 is supported.

source_service_accounts: Option<Vec<String>>

If source service accounts are specified, the firewall rules apply only to traffic originating from an instance with a service account in this list. Source service accounts cannot be used to control traffic to an instance's external IP address because service accounts are associated with an instance, not an IP address. sourceRanges can be set at the same time as sourceServiceAccounts. If both are set, the firewall applies to traffic that has a source IP address within the sourceRanges OR a source IP that belongs to an instance with service account listed in sourceServiceAccount. The connection does not need to match both fields for the firewall to apply. sourceServiceAccounts cannot be used at the same time as sourceTags or targetTags.

source_tags: Option<Vec<String>>

If source tags are specified, the firewall rule applies only to traffic with source IPs that match the primary network interfaces of VM instances that have the tag and are in the same VPC network. Source tags cannot be used to control traffic to an instance's external IP address, it only applies to traffic between instances in the same virtual network. Because tags are associated with instances, not IP addresses. One or both of sourceRanges and sourceTags may be set. If both fields are set, the firewall applies to traffic that has a source IP address within sourceRanges OR a source IP from a resource with a matching tag listed in the sourceTags field. The connection does not need to match both fields for the firewall to apply.

target_service_accounts: Option<Vec<String>>

A list of service accounts indicating sets of instances located in the network that may make network connections as specified in allowed[]. targetServiceAccounts cannot be used at the same time as targetTags or sourceTags. If neither targetServiceAccounts nor targetTags are specified, the firewall rule applies to all instances on the specified network.

target_tags: Option<Vec<String>>

A list of tags that controls which instances the firewall rule applies to. If targetTags are specified, then the firewall rule applies only to instances in the VPC network that have one of those tags. If no targetTags are specified, the firewall rule applies to all instances on the specified network.