Struct google_compute1::api::Image

pub struct Image {
    pub archive_size_bytes: Option<String>,
    pub creation_timestamp: Option<String>,
    pub deprecated: Option<DeprecationStatus>,
    pub description: Option<String>,
    pub disk_size_gb: Option<String>,
    pub family: Option<String>,
    pub guest_os_features: Option<Vec<GuestOsFeature>>,
    pub id: Option<String>,
    pub image_encryption_key: Option<CustomerEncryptionKey>,
    pub kind: Option<String>,
    pub label_fingerprint: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub license_codes: Option<Vec<String>>,
    pub licenses: Option<Vec<String>>,
    pub name: Option<String>,
    pub raw_disk: Option<ImageRawDisk>,
    pub satisfies_pzs: Option<bool>,
    pub self_link: Option<String>,
    pub shielded_instance_initial_state: Option<InitialStateConfig>,
    pub source_disk: Option<String>,
    pub source_disk_encryption_key: Option<CustomerEncryptionKey>,
    pub source_disk_id: Option<String>,
    pub source_image: Option<String>,
    pub source_image_encryption_key: Option<CustomerEncryptionKey>,
    pub source_image_id: Option<String>,
    pub source_snapshot: Option<String>,
    pub source_snapshot_encryption_key: Option<CustomerEncryptionKey>,
    pub source_snapshot_id: Option<String>,
    pub source_type: Option<String>,
    pub status: Option<String>,
    pub storage_locations: Option<Vec<String>>,
}

Represents an Image resource.

You can use images to create boot disks for your VM instances. For more information, read Images. (== resource_for {$api_version}.images ==)

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete images (none)

deprecate images (none)

get images (response)

get from family images (response)

get iam policy images (none)

insert images (request)

list images (none)

patch images (request)

set iam policy images (none)

set labels images (none)

test iam permissions images (none)

Fields
archive_size_bytes: Option<String>

Size of the image tar.gz archive stored in Google Cloud Storage (in bytes).

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

deprecated: Option<DeprecationStatus>

The deprecation status associated with this image.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

disk_size_gb: Option<String>

Size of the image when restored onto a persistent disk (in GB).

family: Option<String>

The name of the image family to which this image belongs. You can create disks by specifying an image family instead of a specific image name. The image family always returns its latest image that is not deprecated. The name of the image family must comply with RFC1035.

guest_os_features: Option<Vec<GuestOsFeature>>

A list of features to enable on the guest operating system. Applicable only for bootable images. Read Enabling guest operating system features to see a list of available options.

id: Option<String>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

image_encryption_key: Option<CustomerEncryptionKey>

Encrypts the image using a customer-supplied encryption key.

After you encrypt an image with a customer-supplied key, you must provide the same key if you use the image later (e.g. to create a disk from the image).

Customer-supplied encryption keys do not protect access to metadata of the disk.

If you do not provide an encryption key when creating the image, then the disk will be encrypted using an automatically generated key and you do not need to provide a key to use the image later.

kind: Option<String>

[Output Only] Type of the resource. Always compute#image for images.

label_fingerprint: Option<String>

A fingerprint for the labels being applied to this image, which is essentially a hash of the labels used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update labels. You must always provide an up-to-date fingerprint hash in order to update or change labels, otherwise the request will fail with error 412 conditionNotMet.

To see the latest fingerprint, make a get() request to retrieve an image.

labels: Option<HashMap<String, String>>

Labels to apply to this image. These can be later modified by the setLabels method.

license_codes: Option<Vec<String>>

Integer license codes indicating which licenses are attached to this image.

licenses: Option<Vec<String>>

Any applicable license URI.

name: Option<String>

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z]([-a-z0-9]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

raw_disk: Option<ImageRawDisk>

The parameters of the raw disk image.

satisfies_pzs: Option<bool>

[Output Only] Reserved for future use.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

shielded_instance_initial_state: Option<InitialStateConfig>

Set the secure boot keys of shielded instance.

source_disk: Option<String>

URL of the source disk used to create this image. This can be a full or valid partial URL. You must provide either this property or the rawDisk.source property but not both to create an image. For example, the following are valid values:

https://www.googleapis.com/compute/v1/projects/project/zones/zone/disks/disk

projects/project/zones/zone/disks/disk

zones/zone/disks/disk

source_disk_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source disk. Required if the source disk is protected by a customer-supplied encryption key.

source_disk_id: Option<String>

[Output Only] The ID value of the disk used to create this image. This value may be used to determine whether the image was taken from the current or a previous instance of a given disk name.

source_image: Option<String>

URL of the source image used to create this image.

In order to create an image, you must provide the full or partial URL of one of the following:

The selfLink URL

This property

The rawDisk.source URL

The sourceDisk URL

source_image_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source image. Required if the source image is protected by a customer-supplied encryption key.

source_image_id: Option<String>

[Output Only] The ID value of the image used to create this image. This value may be used to determine whether the image was taken from the current or a previous instance of a given image name.

source_snapshot: Option<String>

URL of the source snapshot used to create this image.

In order to create an image, you must provide the full or partial URL of one of the following:

The selfLink URL

This property

The sourceImage URL

The rawDisk.source URL

The sourceDisk URL

source_snapshot_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source snapshot. Required if the source snapshot is protected by a customer-supplied encryption key.

source_snapshot_id: Option<String>

[Output Only] The ID value of the snapshot used to create this image. This value may be used to determine whether the snapshot was taken from the current or a previous instance of a given snapshot name.

source_type: Option<String>

The type of the image used to create this disk. The default and only value is RAW

status: Option<String>

[Output Only] The status of the image. An image can be used to create other resources, such as instances, only after the image has been successfully created and the status is set to READY. Possible values are FAILED, PENDING, or READY.

storage_locations: Option<Vec<String>>

Cloud Storage bucket storage location of the image (regional or multi-regional).