Struct google_compute1::api::InstanceGroupMethods

pub struct InstanceGroupMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on instanceGroup resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2;
use compute1::Compute;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Compute::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_instances(...)`, `aggregated_list(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `list_instances(...)`, `remove_instances(...)` and `set_named_ports(...)`
// to build up your call.
let rb = hub.instance_groups();

Implementations

impl<'a, C> InstanceGroupMethods<'a, C>

pub fn add_instances(
&self,
request: InstanceGroupsAddInstancesRequest,
project: &str,
zone: &str,
instance_group: &str
) -> InstanceGroupAddInstanceCall<'a, C>

Create a builder to help you perform the following task:

Adds a list of instances to the specified instance group. All of the instances in the instance group must be in the same network/subnetwork. Read Adding instances for more information.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group where you are adding instances.

pub fn aggregated_list(
&self,
project: &str
) -> InstanceGroupAggregatedListCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the list of instance groups and sorts them by zone.

Arguments

project - Project ID for this request.

pub fn delete(
&self,
project: &str,
zone: &str,
instance_group: &str
) -> InstanceGroupDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified instance group. The instances in the group are not deleted. Note that instance group must not belong to a backend service. Read Deleting an instance group for more information.

Arguments

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group to delete.

pub fn get(
&self,
project: &str,
zone: &str,
instance_group: &str
) -> InstanceGroupGetCall<'a, C>

Create a builder to help you perform the following task:

Returns the specified zonal instance group. Get a list of available zonal instance groups by making a list() request.

For managed instance groups, use the instanceGroupManagers or regionInstanceGroupManagers methods instead.

Arguments

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group.

pub fn insert(
&self,
request: InstanceGroup,
project: &str,
zone: &str
) -> InstanceGroupInsertCall<'a, C>

Create a builder to help you perform the following task:

Creates an instance group in the specified project using the parameters that are included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where you want to create the instance group.

pub fn list(&self, project: &str, zone: &str) -> InstanceGroupListCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the list of zonal instance group resources contained within the specified zone.

For managed instance groups, use the instanceGroupManagers or regionInstanceGroupManagers methods instead.

Arguments

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

pub fn list_instances(
&self,
request: InstanceGroupsListInstancesRequest,
project: &str,
zone: &str,
instance_group: &str
) -> InstanceGroupListInstanceCall<'a, C>

Create a builder to help you perform the following task:

Lists the instances in the specified instance group. The orderBy query parameter is not supported.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group from which you want to generate a list of included instances.

pub fn remove_instances(
&self,
request: InstanceGroupsRemoveInstancesRequest,
project: &str,
zone: &str,
instance_group: &str
) -> InstanceGroupRemoveInstanceCall<'a, C>

Create a builder to help you perform the following task:

Removes one or more instances from the specified instance group, but does not delete those instances.

If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration before the VM instance is removed or deleted.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group where the specified instances will be removed.

pub fn set_named_ports(
&self,
request: InstanceGroupsSetNamedPortsRequest,
project: &str,
zone: &str,
instance_group: &str
) -> InstanceGroupSetNamedPortCall<'a, C>

Create a builder to help you perform the following task:

Sets the named ports for the specified instance group.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group where the named ports are updated.