Struct google_compute1::api::InstanceListCall

pub struct InstanceListCall<'a, C> where
    C: 'a,  { /* fields omitted */ }

Retrieves the list of instances contained within the specified zone.

A builder for the list method supported by a instance resource. It is not used directly, but through a InstanceMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.instances().list("project", "zone")
             .return_partial_success(false)
             .page_token("dolor")
             .order_by("dolore")
             .max_results(87)
             .filter("et")
             .doit().await;

Implementations

impl<'a, C> InstanceListCall<'a, C> where
C: BorrowMut<Client<HttpsConnector<HttpConnector>, Body>>,

pub async fn doit(self) -> Result<(Response<Body>, InstanceList)>

Perform the operation you have build so far.

pub fn project(self, new_value: &str) -> InstanceListCall<'a, C>

Project ID for this request.

Sets the project path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn zone(self, new_value: &str) -> InstanceListCall<'a, C>

The name of the zone for this request.

Sets the zone path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

pub fn return_partial_success(self, new_value: bool) -> InstanceListCall<'a, C>

Opt-in for partial success behavior which provides partial results in case of failure. The default value is false.

Sets the return partial success query property to the given value.

pub fn page_token(self, new_value: &str) -> InstanceListCall<'a, C>

Specifies a page token to use. Set pageToken to the nextPageToken returned by a previous list request to get the next page of results.

Sets the page token query property to the given value.

pub fn order_by(self, new_value: &str) -> InstanceListCall<'a, C>

Sorts list results by a certain order. By default, results are returned in alphanumerical order based on the resource name.

You can also sort results in descending order based on the creation timestamp using orderBy="creationTimestamp desc". This sorts results based on the creationTimestamp field in reverse chronological order (newest result first). Use this to sort resources like operations so that the newest operation is returned first.

Currently, only sorting by name or creationTimestamp desc is supported.

Sets the order by query property to the given value.

pub fn max_results(self, new_value: u32) -> InstanceListCall<'a, C>

The maximum number of results per page that should be returned. If the number of available results is larger than maxResults, Compute Engine returns a nextPageToken that can be used to get the next page of results in subsequent list requests. Acceptable values are 0 to 500, inclusive. (Default: 500)

Sets the max results query property to the given value.

pub fn filter(self, new_value: &str) -> InstanceListCall<'a, C>

A filter expression that filters resources listed in the response. The expression must specify the field name, a comparison operator, and the value that you want to use for filtering. The value must be a string, a number, or a boolean. The comparison operator must be either =, !=, >, or <.

For example, if you are filtering Compute Engine instances, you can exclude instances named example-instance by specifying name != example-instance.

You can also filter nested fields. For example, you could specify scheduling.automaticRestart = false to include instances only if they are not scheduled for automatic restarts. You can use filtering on nested fields to filter based on resource labels.

To filter on multiple expressions, provide each separate expression within parentheses. For example: (scheduling.automaticRestart = true) (cpuPlatform = "Intel Skylake") By default, each expression is an AND expression. However, you can include AND and OR expressions explicitly. For example: (cpuPlatform = "Intel Skylake") OR (cpuPlatform = "Intel Broadwell") AND (scheduling.automaticRestart = true)

Sets the filter query property to the given value.

pub fn delegate(
self,
new_value: &'a mut dyn Delegate
) -> InstanceListCall<'a, C>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

pub fn param<T>(self, name: T, value: T) -> InstanceListCall<'a, C> where
T: AsRef<str>,

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known parameters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

alt (query-string) - Data format for the response.

fields (query-string) - Selector specifying which fields to include in a partial response.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

oauth_token (query-string) - OAuth 2.0 token for the current user.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

quotaUser (query-string) - An opaque string that represents a user for quota purposes. Must not exceed 40 characters.

userIp (query-string) - Deprecated. Please use quotaUser instead.

pub fn add_scope<T, S>(self, scope: T) -> InstanceListCall<'a, C> where
T: Into<Option<S>>,
S: AsRef<str>,

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Readonly.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope. If None is specified, then all scopes will be removed and no default scope will be used either. In that case, you have to specify your API-key using the key parameter (see the param() function for details).

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.