Struct google_compute1::api::InstanceProperties

pub struct InstanceProperties {
    pub advanced_machine_features: Option<AdvancedMachineFeatures>,
    pub can_ip_forward: Option<bool>,
    pub confidential_instance_config: Option<ConfidentialInstanceConfig>,
    pub description: Option<String>,
    pub disks: Option<Vec<AttachedDisk>>,
    pub guest_accelerators: Option<Vec<AcceleratorConfig>>,
    pub labels: Option<HashMap<String, String>>,
    pub machine_type: Option<String>,
    pub metadata: Option<Metadata>,
    pub min_cpu_platform: Option<String>,
    pub network_interfaces: Option<Vec<NetworkInterface>>,
    pub post_key_revocation_action_type: Option<String>,
    pub private_ipv6_google_access: Option<String>,
    pub reservation_affinity: Option<ReservationAffinity>,
    pub resource_policies: Option<Vec<String>>,
    pub scheduling: Option<Scheduling>,
    pub service_accounts: Option<Vec<ServiceAccount>>,
    pub shielded_instance_config: Option<ShieldedInstanceConfig>,
    pub tags: Option<Tags>,
}

This type is not used in any activity, and only used as part of another schema.

Fields
advanced_machine_features: Option<AdvancedMachineFeatures>

Controls for advanced machine-related behavior features.

can_ip_forward: Option<bool>

Enables instances created based on these properties to send packets with source IP addresses other than their own and receive packets with destination IP addresses other than their own. If these instances will be used as an IP gateway or it will be set as the next-hop in a Route resource, specify true. If unsure, leave this set to false. See the Enable IP forwarding documentation for more information.

confidential_instance_config: Option<ConfidentialInstanceConfig>

Specifies the Confidential Instance options.

description: Option<String>

An optional text description for the instances that are created from these properties.

disks: Option<Vec<AttachedDisk>>

An array of disks that are associated with the instances that are created from these properties.

guest_accelerators: Option<Vec<AcceleratorConfig>>

A list of guest accelerator cards' type and count to use for instances created from these properties.

labels: Option<HashMap<String, String>>

Labels to apply to instances that are created from these properties.

machine_type: Option<String>

The machine type to use for instances that are created from these properties.

metadata: Option<Metadata>

The metadata key/value pairs to assign to instances that are created from these properties. These pairs can consist of custom metadata or predefined keys. See Project and instance metadata for more information.

min_cpu_platform: Option<String>

Minimum cpu/platform to be used by instances. The instance may be scheduled on the specified or newer cpu/platform. Applicable values are the friendly names of CPU platforms, such as minCpuPlatform: "Intel Haswell" or minCpuPlatform: "Intel Sandy Bridge". For more information, read Specifying a Minimum CPU Platform.

network_interfaces: Option<Vec<NetworkInterface>>

An array of network access configurations for this interface.

post_key_revocation_action_type: Option<String>

PostKeyRevocationActionType of the instance.

private_ipv6_google_access: Option<String>

The private IPv6 google access type for VMs. If not specified, use INHERIT_FROM_SUBNETWORK as default.

reservation_affinity: Option<ReservationAffinity>

Specifies the reservations that instances can consume from.

resource_policies: Option<Vec<String>>

Resource policies (names, not ULRs) applied to instances created from these properties.

scheduling: Option<Scheduling>

Specifies the scheduling options for the instances that are created from these properties.

service_accounts: Option<Vec<ServiceAccount>>

A list of service accounts with specified scopes. Access tokens for these service accounts are available to the instances that are created from these properties. Use metadata queries to obtain the access tokens for these instances.

shielded_instance_config: Option<ShieldedInstanceConfig>

no description provided

tags: Option<Tags>

A list of tags to apply to the instances that are created from these properties. The tags identify valid sources or targets for network firewalls. The setTags method can modify this list of tags. Each tag within the list must comply with RFC1035.