Struct google_compute1::api::InterconnectAttachment

pub struct InterconnectAttachment {
    pub admin_enabled: Option<bool>,
    pub bandwidth: Option<String>,
    pub candidate_subnets: Option<Vec<String>>,
    pub cloud_router_ip_address: Option<String>,
    pub creation_timestamp: Option<String>,
    pub customer_router_ip_address: Option<String>,
    pub dataplane_version: Option<i32>,
    pub description: Option<String>,
    pub edge_availability_domain: Option<String>,
    pub encryption: Option<String>,
    pub google_reference_id: Option<String>,
    pub id: Option<String>,
    pub interconnect: Option<String>,
    pub ipsec_internal_addresses: Option<Vec<String>>,
    pub kind: Option<String>,
    pub mtu: Option<i32>,
    pub name: Option<String>,
    pub operational_status: Option<String>,
    pub pairing_key: Option<String>,
    pub partner_asn: Option<String>,
    pub partner_metadata: Option<InterconnectAttachmentPartnerMetadata>,
    pub private_interconnect_info: Option<InterconnectAttachmentPrivateInfo>,
    pub region: Option<String>,
    pub router: Option<String>,
    pub self_link: Option<String>,
    pub state: Option<String>,
    pub type_: Option<String>,
    pub vlan_tag8021q: Option<i32>,
}

Represents an Interconnect Attachment (VLAN) resource.

You can use Interconnect attachments (VLANS) to connect your Virtual Private Cloud networks to your on-premises networks through an Interconnect. For more information, read Creating VLAN Attachments. (== resource_for {$api_version}.interconnectAttachments ==)

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

aggregated list interconnect attachments (none)

delete interconnect attachments (none)

get interconnect attachments (response)

insert interconnect attachments (request)

list interconnect attachments (none)

patch interconnect attachments (request)

Fields
admin_enabled: Option<bool>

Determines whether this Attachment will carry packets. Not present for PARTNER_PROVIDER.

bandwidth: Option<String>

Provisioned bandwidth capacity for the interconnect attachment. For attachments of type DEDICATED, the user can set the bandwidth. For attachments of type PARTNER, the Google Partner that is operating the interconnect must set the bandwidth. Output only for PARTNER type, mutable for PARTNER_PROVIDER and DEDICATED, and can take one of the following values:

BPS_50M: 50 Mbit/s

BPS_100M: 100 Mbit/s

BPS_200M: 200 Mbit/s

BPS_300M: 300 Mbit/s

BPS_400M: 400 Mbit/s

BPS_500M: 500 Mbit/s

BPS_1G: 1 Gbit/s

BPS_2G: 2 Gbit/s

BPS_5G: 5 Gbit/s

BPS_10G: 10 Gbit/s

BPS_20G: 20 Gbit/s

BPS_50G: 50 Gbit/s

candidate_subnets: Option<Vec<String>>

Up to 16 candidate prefixes that can be used to restrict the allocation of cloudRouterIpAddress and customerRouterIpAddress for this attachment. All prefixes must be within link-local address space (169.254.0.0/16) and must be /29 or shorter (/28, /27, etc). Google will attempt to select an unused /29 from the supplied candidate prefix(es). The request will fail if all possible /29s are in use on Google's edge. If not supplied, Google will randomly select an unused /29 from all of link-local space.

cloud_router_ip_address: Option<String>

[Output Only] IPv4 address + prefix length to be configured on Cloud Router Interface for this interconnect attachment.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

customer_router_ip_address: Option<String>

[Output Only] IPv4 address + prefix length to be configured on the customer router subinterface for this interconnect attachment.

dataplane_version: Option<i32>

[Output Only] Dataplane version for this InterconnectAttachment.

description: Option<String>

An optional description of this resource.

edge_availability_domain: Option<String>

Desired availability domain for the attachment. Only available for type PARTNER, at creation time, and can take one of the following values:

AVAILABILITY_DOMAIN_ANY

AVAILABILITY_DOMAIN_1

AVAILABILITY_DOMAIN_2 For improved reliability, customers should configure a pair of attachments, one per availability domain. The selected availability domain will be provided to the Partner via the pairing key, so that the provisioned circuit will lie in the specified domain. If not specified, the value will default to AVAILABILITY_DOMAIN_ANY.

encryption: Option<String>

Indicates the user-supplied encryption option of this interconnect attachment:

NONE is the default value, which means that the attachment carries unencrypted traffic. VMs can send traffic to, or receive traffic from, this type of attachment.

IPSEC indicates that the attachment carries only traffic encrypted by an IPsec device such as an HA VPN gateway. VMs cannot directly send traffic to, or receive traffic from, such an attachment. To use IPsec-encrypted Cloud Interconnect, create the attachment using this option. Not currently available in all Interconnect locations.

google_reference_id: Option<String>

[Output Only] Google reference ID, to be used when raising support tickets with Google or otherwise to debug backend connectivity issues. [Deprecated] This field is not used.

id: Option<String>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

interconnect: Option<String>

URL of the underlying Interconnect object that this attachment's traffic will traverse through.

ipsec_internal_addresses: Option<Vec<String>>

URL of addresses that have been reserved for the interconnect attachment, Used only for interconnect attachment that has the encryption option as IPSEC. The addresses must be RFC 1918 IP address ranges. When creating HA VPN gateway over the interconnect attachment, if the attachment is configured to use an RFC 1918 IP address, then the VPN gateway?s IP address will be allocated from the IP address range specified here. For example, if the HA VPN gateway?s interface 0 is paired to this interconnect attachment, then an RFC 1918 IP address for the VPN gateway interface 0 will be allocated from the IP address specified for this interconnect attachment. If this field is not specified for interconnect attachment that has encryption option as IPSEC, later on when creating HA VPN gateway on this interconnect attachment, the HA VPN gateway's IP address will be allocated from regional external IP address pool. Not currently available in all Interconnect locations.

kind: Option<String>

[Output Only] Type of the resource. Always compute#interconnectAttachment for interconnect attachments.

mtu: Option<i32>

Maximum Transmission Unit (MTU), in bytes, of packets passing through this interconnect attachment. Only 1440 and 1500 are allowed. If not specified, the value will default to 1440.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z]([-a-z0-9]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

operational_status: Option<String>

[Output Only] The current status of whether or not this interconnect attachment is functional, which can take one of the following values:

OS_ACTIVE: The attachment has been turned up and is ready to use.

OS_UNPROVISIONED: The attachment is not ready to use yet, because turnup is not complete.

pairing_key: Option<String>

[Output only for type PARTNER. Input only for PARTNER_PROVIDER. Not present for DEDICATED]. The opaque identifier of an PARTNER attachment used to initiate provisioning with a selected partner. Of the form "XXXXX/region/domain"

partner_asn: Option<String>

Optional BGP ASN for the router supplied by a Layer 3 Partner if they configured BGP on behalf of the customer. Output only for PARTNER type, input only for PARTNER_PROVIDER, not available for DEDICATED.

partner_metadata: Option<InterconnectAttachmentPartnerMetadata>

Informational metadata about Partner attachments from Partners to display to customers. Output only for for PARTNER type, mutable for PARTNER_PROVIDER, not available for DEDICATED.

private_interconnect_info: Option<InterconnectAttachmentPrivateInfo>

[Output Only] Information specific to an InterconnectAttachment. This property is populated if the interconnect that this is attached to is of type DEDICATED.

region: Option<String>

[Output Only] URL of the region where the regional interconnect attachment resides. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

router: Option<String>

URL of the Cloud Router to be used for dynamic routing. This router must be in the same region as this InterconnectAttachment. The InterconnectAttachment will automatically connect the Interconnect to the network & region within which the Cloud Router is configured.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

state: Option<String>

[Output Only] The current state of this attachment's functionality. Enum values ACTIVE and UNPROVISIONED are shared by DEDICATED/PRIVATE, PARTNER, and PARTNER_PROVIDER interconnect attachments, while enum values PENDING_PARTNER, PARTNER_REQUEST_RECEIVED, and PENDING_CUSTOMER are used for only PARTNER and PARTNER_PROVIDER interconnect attachments. This state can take one of the following values:

ACTIVE: The attachment has been turned up and is ready to use.

UNPROVISIONED: The attachment is not ready to use yet, because turnup is not complete.

PENDING_PARTNER: A newly-created PARTNER attachment that has not yet been configured on the Partner side.

PARTNER_REQUEST_RECEIVED: A PARTNER attachment is in the process of provisioning after a PARTNER_PROVIDER attachment was created that references it.

PENDING_CUSTOMER: A PARTNER or PARTNER_PROVIDER attachment that is waiting for a customer to activate it.

DEFUNCT: The attachment was deleted externally and is no longer functional. This could be because the associated Interconnect was removed, or because the other side of a Partner attachment was deleted.

type_: Option<String>

The type of interconnect attachment this is, which can take one of the following values:

DEDICATED: an attachment to a Dedicated Interconnect.

PARTNER: an attachment to a Partner Interconnect, created by the customer.

PARTNER_PROVIDER: an attachment to a Partner Interconnect, created by the partner.

vlan_tag8021q: Option<i32>

The IEEE 802.1Q VLAN tag for this attachment, in the range 2-4094. Only specified at creation time.