Struct google_compute1::api::NodeGroupMethods

pub struct NodeGroupMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on nodeGroup resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2;
use compute1::Compute;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Compute::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_nodes(...)`, `aggregated_list(...)`, `delete(...)`, `delete_nodes(...)`, `get(...)`, `get_iam_policy(...)`, `insert(...)`, `list(...)`, `list_nodes(...)`, `patch(...)`, `set_iam_policy(...)`, `set_node_template(...)` and `test_iam_permissions(...)`
// to build up your call.
let rb = hub.node_groups();

Implementations

impl<'a, C> NodeGroupMethods<'a, C>

pub fn add_nodes(
&self,
request: NodeGroupsAddNodesRequest,
project: &str,
zone: &str,
node_group: &str
) -> NodeGroupAddNodeCall<'a, C>

Create a builder to help you perform the following task:

Adds specified number of nodes to the node group.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

nodeGroup - Name of the NodeGroup resource.

pub fn aggregated_list(
&self,
project: &str
) -> NodeGroupAggregatedListCall<'a, C>

Create a builder to help you perform the following task:

Retrieves an aggregated list of node groups. Note: use nodeGroups.listNodes for more details about each group.

Arguments

project - Project ID for this request.

pub fn delete(
&self,
project: &str,
zone: &str,
node_group: &str
) -> NodeGroupDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified NodeGroup resource.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

nodeGroup - Name of the NodeGroup resource to delete.

pub fn delete_nodes(
&self,
request: NodeGroupsDeleteNodesRequest,
project: &str,
zone: &str,
node_group: &str
) -> NodeGroupDeleteNodeCall<'a, C>

Create a builder to help you perform the following task:

Deletes specified nodes from the node group.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

nodeGroup - Name of the NodeGroup resource whose nodes will be deleted.

pub fn get(
&self,
project: &str,
zone: &str,
node_group: &str
) -> NodeGroupGetCall<'a, C>

Create a builder to help you perform the following task:

Returns the specified NodeGroup. Get a list of available NodeGroups by making a list() request. Note: the "nodes" field should not be used. Use nodeGroups.listNodes instead.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

nodeGroup - Name of the node group to return.

pub fn get_iam_policy(
&self,
project: &str,
zone: &str,
resource: &str
) -> NodeGroupGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. May be empty if no such policy or resource exists.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

resource - Name or id of the resource for this request.

pub fn insert(
&self,
request: NodeGroup,
project: &str,
zone: &str,
initial_node_count: i32
) -> NodeGroupInsertCall<'a, C>

Create a builder to help you perform the following task:

Creates a NodeGroup resource in the specified project using the data included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

initialNodeCount - Initial count of nodes in the node group.

pub fn list(&self, project: &str, zone: &str) -> NodeGroupListCall<'a, C>

Create a builder to help you perform the following task:

Retrieves a list of node groups available to the specified project. Note: use nodeGroups.listNodes for more details about each group.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

pub fn list_nodes(
&self,
project: &str,
zone: &str,
node_group: &str
) -> NodeGroupListNodeCall<'a, C>

Create a builder to help you perform the following task:

Lists nodes in the node group.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

nodeGroup - Name of the NodeGroup resource whose nodes you want to list.

pub fn patch(
&self,
request: NodeGroup,
project: &str,
zone: &str,
node_group: &str
) -> NodeGroupPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the specified node group.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

nodeGroup - Name of the NodeGroup resource to update.

pub fn set_iam_policy(
&self,
request: ZoneSetPolicyRequest,
project: &str,
zone: &str,
resource: &str
) -> NodeGroupSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

resource - Name or id of the resource for this request.

pub fn set_node_template(
&self,
request: NodeGroupsSetNodeTemplateRequest,
project: &str,
zone: &str,
node_group: &str
) -> NodeGroupSetNodeTemplateCall<'a, C>

Create a builder to help you perform the following task:

Updates the node template of the node group.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

nodeGroup - Name of the NodeGroup resource to update.

pub fn test_iam_permissions(
&self,
request: TestPermissionsRequest,
project: &str,
zone: &str,
resource: &str
) -> NodeGroupTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

resource - Name or id of the resource for this request.