Struct google_compute1::api::NodeTemplate

pub struct NodeTemplate {
    pub accelerators: Option<Vec<AcceleratorConfig>>,
    pub cpu_overcommit_type: Option<String>,
    pub creation_timestamp: Option<String>,
    pub description: Option<String>,
    pub disks: Option<Vec<LocalDisk>>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub node_affinity_labels: Option<HashMap<String, String>>,
    pub node_type: Option<String>,
    pub node_type_flexibility: Option<NodeTemplateNodeTypeFlexibility>,
    pub region: Option<String>,
    pub self_link: Option<String>,
    pub server_binding: Option<ServerBinding>,
    pub status: Option<String>,
    pub status_message: Option<String>,
}

Represent a sole-tenant Node Template resource.

You can use a template to define properties for nodes in a node group. For more information, read Creating node groups and instances. (== resource_for {$api_version}.nodeTemplates ==)

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

aggregated list node templates (none)

delete node templates (none)

get node templates (response)

get iam policy node templates (none)

insert node templates (request)

list node templates (none)

set iam policy node templates (none)

test iam permissions node templates (none)

Fields
accelerators: Option<Vec<AcceleratorConfig>>

no description provided

cpu_overcommit_type: Option<String>

CPU overcommit.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

disks: Option<Vec<LocalDisk>>

no description provided

id: Option<String>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

kind: Option<String>

[Output Only] The type of the resource. Always compute#nodeTemplate for node templates.

name: Option<String>

The name of the resource, provided by the client when initially creating the resource. The resource name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z]([-a-z0-9]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

node_affinity_labels: Option<HashMap<String, String>>

Labels to use for node affinity, which will be used in instance scheduling.

node_type: Option<String>

The node type to use for nodes group that are created from this template.

node_type_flexibility: Option<NodeTemplateNodeTypeFlexibility>

The flexible properties of the desired node type. Node groups that use this node template will create nodes of a type that matches these properties.

This field is mutually exclusive with the node_type property; you can only define one or the other, but not both.

region: Option<String>

[Output Only] The name of the region where the node template resides, such as us-central1.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

server_binding: Option<ServerBinding>

Sets the binding properties for the physical server. Valid values include:

Default RESTART_NODE_ON_ANY_SERVER: Restarts VMs on any available physical server

RESTART_NODE_ON_MINIMAL_SERVER: Restarts VMs on the same physical server whenever possible

See Sole-tenant node options for more information.

status: Option<String>

[Output Only] The status of the node template. One of the following values: CREATING, READY, and DELETING.

status_message: Option<String>

[Output Only] An optional, human-readable explanation of the status.