Struct google_compute1::api::Operation

pub struct Operation {
    pub client_operation_id: Option<String>,
    pub creation_timestamp: Option<String>,
    pub description: Option<String>,
    pub end_time: Option<String>,
    pub error: Option<OperationError>,
    pub http_error_message: Option<String>,
    pub http_error_status_code: Option<i32>,
    pub id: Option<String>,
    pub insert_time: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub operation_group_id: Option<String>,
    pub operation_type: Option<String>,
    pub progress: Option<i32>,
    pub region: Option<String>,
    pub self_link: Option<String>,
    pub start_time: Option<String>,
    pub status: Option<String>,
    pub status_message: Option<String>,
    pub target_id: Option<String>,
    pub target_link: Option<String>,
    pub user: Option<String>,
    pub warnings: Option<Vec<OperationWarnings>>,
    pub zone: Option<String>,
}

Represents an Operation resource.

Google Compute Engine has three Operation resources:

Global * Regional * Zonal

You can use an operation resource to manage asynchronous API requests. For more information, read Handling API responses.

Operations can be global, regional or zonal.

For global operations, use the globalOperations resource.

For regional operations, use the regionOperations resource.

For zonal operations, use the zonalOperations resource.

For more information, read Global, Regional, and Zonal Resources. (== resource_for {$api_version}.globalOperations ==) (== resource_for {$api_version}.regionOperations ==) (== resource_for {$api_version}.zoneOperations ==)

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete addresses (response)

insert addresses (response)

delete autoscalers (response)

insert autoscalers (response)

patch autoscalers (response)

update autoscalers (response)

add signed url key backend buckets (response)

delete backend buckets (response)

delete signed url key backend buckets (response)

insert backend buckets (response)

patch backend buckets (response)

update backend buckets (response)

add signed url key backend services (response)

delete backend services (response)

delete signed url key backend services (response)

insert backend services (response)

patch backend services (response)

set security policy backend services (response)

update backend services (response)

add resource policies disks (response)

create snapshot disks (response)

delete disks (response)

insert disks (response)

remove resource policies disks (response)

resize disks (response)

set labels disks (response)

delete external vpn gateways (response)

insert external vpn gateways (response)

set labels external vpn gateways (response)

add association firewall policies (response)

add rule firewall policies (response)

clone rules firewall policies (response)

delete firewall policies (response)

insert firewall policies (response)

move firewall policies (response)

patch firewall policies (response)

patch rule firewall policies (response)

remove association firewall policies (response)

remove rule firewall policies (response)

delete firewalls (response)

insert firewalls (response)

patch firewalls (response)

update firewalls (response)

delete forwarding rules (response)

insert forwarding rules (response)

patch forwarding rules (response)

set labels forwarding rules (response)

set target forwarding rules (response)

delete global addresses (response)

insert global addresses (response)

delete global forwarding rules (response)

insert global forwarding rules (response)

patch global forwarding rules (response)

set labels global forwarding rules (response)

set target global forwarding rules (response)

attach network endpoints global network endpoint groups (response)

delete global network endpoint groups (response)

detach network endpoints global network endpoint groups (response)

insert global network endpoint groups (response)

get global operations (response)

wait global operations (response)

get global organization operations (response)

delete global public delegated prefixes (response)

insert global public delegated prefixes (response)

patch global public delegated prefixes (response)

delete health checks (response)

insert health checks (response)

patch health checks (response)

update health checks (response)

delete http health checks (response)

insert http health checks (response)

patch http health checks (response)

update http health checks (response)

delete https health checks (response)

insert https health checks (response)

patch https health checks (response)

update https health checks (response)

delete images (response)

deprecate images (response)

insert images (response)

patch images (response)

set labels images (response)

abandon instances instance group managers (response)

apply updates to instances instance group managers (response)

create instances instance group managers (response)

delete instance group managers (response)

delete instances instance group managers (response)

delete per instance configs instance group managers (response)

insert instance group managers (response)

patch instance group managers (response)

patch per instance configs instance group managers (response)

recreate instances instance group managers (response)

resize instance group managers (response)

set instance template instance group managers (response)

set target pools instance group managers (response)

update per instance configs instance group managers (response)

add instances instance groups (response)

delete instance groups (response)

insert instance groups (response)

remove instances instance groups (response)

set named ports instance groups (response)

delete instance templates (response)

insert instance templates (response)

add access config instances (response)

add resource policies instances (response)

attach disk instances (response)

bulk insert instances (response)

delete instances (response)

delete access config instances (response)

detach disk instances (response)

insert instances (response)

remove resource policies instances (response)

reset instances (response)

set deletion protection instances (response)

set disk auto delete instances (response)

set labels instances (response)

set machine resources instances (response)

set machine type instances (response)

set metadata instances (response)

set min cpu platform instances (response)

set scheduling instances (response)

set service account instances (response)

set shielded instance integrity policy instances (response)

set tags instances (response)

simulate maintenance event instances (response)

start instances (response)

start with encryption key instances (response)

stop instances (response)

update instances (response)

update access config instances (response)

update display device instances (response)

update network interface instances (response)

update shielded instance config instances (response)

delete interconnect attachments (response)

insert interconnect attachments (response)

patch interconnect attachments (response)

delete interconnects (response)

insert interconnects (response)

patch interconnects (response)

delete licenses (response)

insert licenses (response)

attach network endpoints network endpoint groups (response)

delete network endpoint groups (response)

detach network endpoints network endpoint groups (response)

insert network endpoint groups (response)

add peering networks (response)

delete networks (response)

insert networks (response)

patch networks (response)

remove peering networks (response)

switch to custom mode networks (response)

update peering networks (response)

add nodes node groups (response)

delete node groups (response)

delete nodes node groups (response)

insert node groups (response)

patch node groups (response)

set node template node groups (response)

delete node templates (response)

insert node templates (response)

delete packet mirrorings (response)

insert packet mirrorings (response)

patch packet mirrorings (response)

disable xpn host projects (response)

disable xpn resource projects (response)

enable xpn host projects (response)

enable xpn resource projects (response)

move disk projects (response)

move instance projects (response)

set common instance metadata projects (response)

set default network tier projects (response)

set usage export bucket projects (response)

delete public advertised prefixes (response)

insert public advertised prefixes (response)

patch public advertised prefixes (response)

delete public delegated prefixes (response)

insert public delegated prefixes (response)

patch public delegated prefixes (response)

delete region autoscalers (response)

insert region autoscalers (response)

patch region autoscalers (response)

update region autoscalers (response)

delete region backend services (response)

insert region backend services (response)

patch region backend services (response)

update region backend services (response)

insert region commitments (response)

add resource policies region disks (response)

create snapshot region disks (response)

delete region disks (response)

insert region disks (response)

remove resource policies region disks (response)

resize region disks (response)

set labels region disks (response)

delete region health check services (response)

insert region health check services (response)

patch region health check services (response)

delete region health checks (response)

insert region health checks (response)

patch region health checks (response)

update region health checks (response)

abandon instances region instance group managers (response)

apply updates to instances region instance group managers (response)

create instances region instance group managers (response)

delete region instance group managers (response)

delete instances region instance group managers (response)

delete per instance configs region instance group managers (response)

insert region instance group managers (response)

patch region instance group managers (response)

patch per instance configs region instance group managers (response)

recreate instances region instance group managers (response)

resize region instance group managers (response)

set instance template region instance group managers (response)

set target pools region instance group managers (response)

update per instance configs region instance group managers (response)

set named ports region instance groups (response)

bulk insert region instances (response)

delete region network endpoint groups (response)

insert region network endpoint groups (response)

delete region notification endpoints (response)

insert region notification endpoints (response)

get region operations (response)

wait region operations (response)

delete region ssl certificates (response)

insert region ssl certificates (response)

delete region target http proxies (response)

insert region target http proxies (response)

set url map region target http proxies (response)

delete region target https proxies (response)

insert region target https proxies (response)

set ssl certificates region target https proxies (response)

set url map region target https proxies (response)

delete region url maps (response)

insert region url maps (response)

patch region url maps (response)

update region url maps (response)

delete reservations (response)

insert reservations (response)

resize reservations (response)

delete resource policies (response)

insert resource policies (response)

delete routers (response)

insert routers (response)

patch routers (response)

update routers (response)

delete routes (response)

insert routes (response)

add rule security policies (response)

delete security policies (response)

insert security policies (response)

patch security policies (response)

patch rule security policies (response)

remove rule security policies (response)

delete snapshots (response)

set labels snapshots (response)

delete ssl certificates (response)

insert ssl certificates (response)

delete ssl policies (response)

insert ssl policies (response)

patch ssl policies (response)

delete subnetworks (response)

expand ip cidr range subnetworks (response)

insert subnetworks (response)

patch subnetworks (response)

set private ip google access subnetworks (response)

delete target grpc proxies (response)

insert target grpc proxies (response)

patch target grpc proxies (response)

delete target http proxies (response)

insert target http proxies (response)

patch target http proxies (response)

set url map target http proxies (response)

delete target https proxies (response)

insert target https proxies (response)

patch target https proxies (response)

set quic override target https proxies (response)

set ssl certificates target https proxies (response)

set ssl policy target https proxies (response)

set url map target https proxies (response)

delete target instances (response)

insert target instances (response)

add health check target pools (response)

add instance target pools (response)

delete target pools (response)

insert target pools (response)

remove health check target pools (response)

remove instance target pools (response)

set backup target pools (response)

delete target ssl proxies (response)

insert target ssl proxies (response)

set backend service target ssl proxies (response)

set proxy header target ssl proxies (response)

set ssl certificates target ssl proxies (response)

set ssl policy target ssl proxies (response)

delete target tcp proxies (response)

insert target tcp proxies (response)

set backend service target tcp proxies (response)

set proxy header target tcp proxies (response)

delete target vpn gateways (response)

insert target vpn gateways (response)

delete url maps (response)

insert url maps (response)

invalidate cache url maps (response)

patch url maps (response)

update url maps (response)

delete vpn gateways (response)

insert vpn gateways (response)

set labels vpn gateways (response)

delete vpn tunnels (response)

insert vpn tunnels (response)

get zone operations (response)

wait zone operations (response)

Fields
client_operation_id: Option<String>

[Output Only] The value of requestId if you provided it in the request. Not present otherwise.

creation_timestamp: Option<String>

[Deprecated] This field is deprecated.

description: Option<String>

[Output Only] A textual description of the operation, which is set when the operation is created.

end_time: Option<String>

[Output Only] The time that this operation was completed. This value is in RFC3339 text format.

error: Option<OperationError>

[Output Only] If errors are generated during processing of the operation, this field will be populated.

http_error_message: Option<String>

[Output Only] If the operation fails, this field contains the HTTP error message that was returned, such as NOT FOUND.

http_error_status_code: Option<i32>

[Output Only] If the operation fails, this field contains the HTTP error status code that was returned. For example, a 404 means the resource was not found.

id: Option<String>

[Output Only] The unique identifier for the operation. This identifier is defined by the server.

insert_time: Option<String>

[Output Only] The time that this operation was requested. This value is in RFC3339 text format.

kind: Option<String>

[Output Only] Type of the resource. Always compute#operation for Operation resources.

name: Option<String>

[Output Only] Name of the operation.

operation_group_id: Option<String>

[Output Only] An ID that represents a group of operations, such as when a group of operations results from a bulkInsert API request.

operation_type: Option<String>

[Output Only] The type of operation, such as insert, update, or delete, and so on.

progress: Option<i32>

[Output Only] An optional progress indicator that ranges from 0 to 100. There is no requirement that this be linear or support any granularity of operations. This should not be used to guess when the operation will be complete. This number should monotonically increase as the operation progresses.

region: Option<String>

[Output Only] The URL of the region where the operation resides. Only applicable when performing regional operations.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

start_time: Option<String>

[Output Only] The time that this operation was started by the server. This value is in RFC3339 text format.

status: Option<String>

[Output Only] The status of the operation, which can be one of the following: PENDING, RUNNING, or DONE.

status_message: Option<String>

[Output Only] An optional textual description of the current status of the operation.

target_id: Option<String>

[Output Only] The unique target ID, which identifies a specific incarnation of the target resource.

target_link: Option<String>

[Output Only] The URL of the resource that the operation modifies. For operations related to creating a snapshot, this points to the persistent disk that the snapshot was created from.

user: Option<String>

[Output Only] User who requested the operation, for example: user@example.com.

warnings: Option<Vec<OperationWarnings>>

[Output Only] If warning messages are generated during processing of the operation, this field will be populated.

zone: Option<String>

[Output Only] The URL of the zone where the operation resides. Only applicable when performing per-zone operations.