Struct google_compute1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2;
use compute1::Compute;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Compute::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `disable_xpn_host(...)`, `disable_xpn_resource(...)`, `enable_xpn_host(...)`, `enable_xpn_resource(...)`, `get(...)`, `get_xpn_host(...)`, `get_xpn_resources(...)`, `list_xpn_hosts(...)`, `move_disk(...)`, `move_instance(...)`, `set_common_instance_metadata(...)`, `set_default_network_tier(...)` and `set_usage_export_bucket(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn disable_xpn_host(
&self,
project: &str
) -> ProjectDisableXpnHostCall<'a, C>

Create a builder to help you perform the following task:

Disable this project as a shared VPC host project.

Arguments

project - Project ID for this request.

pub fn disable_xpn_resource(
&self,
request: ProjectsDisableXpnResourceRequest,
project: &str
) -> ProjectDisableXpnResourceCall<'a, C>

Create a builder to help you perform the following task:

Disable a service resource (also known as service project) associated with this host project.

Arguments

request - No description provided.

project - Project ID for this request.

pub fn enable_xpn_host(&self, project: &str) -> ProjectEnableXpnHostCall<'a, C>

Create a builder to help you perform the following task:

Enable this project as a shared VPC host project.

Arguments

project - Project ID for this request.

pub fn enable_xpn_resource(
&self,
request: ProjectsEnableXpnResourceRequest,
project: &str
) -> ProjectEnableXpnResourceCall<'a, C>

Create a builder to help you perform the following task:

Enable service resource (a.k.a service project) for a host project, so that subnets in the host project can be used by instances in the service project.

Arguments

request - No description provided.

project - Project ID for this request.

pub fn get(&self, project: &str) -> ProjectGetCall<'a, C>

Create a builder to help you perform the following task:

Returns the specified Project resource.

Arguments

project - Project ID for this request.

pub fn get_xpn_host(&self, project: &str) -> ProjectGetXpnHostCall<'a, C>

Create a builder to help you perform the following task:

Gets the shared VPC host project that this project links to. May be empty if no link exists.

Arguments

project - Project ID for this request.

pub fn get_xpn_resources(
&self,
project: &str
) -> ProjectGetXpnResourceCall<'a, C>

Create a builder to help you perform the following task:

Gets service resources (a.k.a service project) associated with this host project.

Arguments

project - Project ID for this request.

pub fn list_xpn_hosts(
&self,
request: ProjectsListXpnHostsRequest,
project: &str
) -> ProjectListXpnHostCall<'a, C>

Create a builder to help you perform the following task:

Lists all shared VPC host projects visible to the user in an organization.

Arguments

request - No description provided.

project - Project ID for this request.

pub fn move_disk(
&self,
request: DiskMoveRequest,
project: &str
) -> ProjectMoveDiskCall<'a, C>

Create a builder to help you perform the following task:

Moves a persistent disk from one zone to another.

Arguments

request - No description provided.

project - Project ID for this request.

pub fn move_instance(
&self,
request: InstanceMoveRequest,
project: &str
) -> ProjectMoveInstanceCall<'a, C>

Create a builder to help you perform the following task:

Moves an instance and its attached persistent disks from one zone to another.

Arguments

request - No description provided.

project - Project ID for this request.

pub fn set_common_instance_metadata(
&self,
request: Metadata,
project: &str
) -> ProjectSetCommonInstanceMetadataCall<'a, C>

Create a builder to help you perform the following task:

Sets metadata common to all instances within the specified project using the data included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

pub fn set_default_network_tier(
&self,
request: ProjectsSetDefaultNetworkTierRequest,
project: &str
) -> ProjectSetDefaultNetworkTierCall<'a, C>

Create a builder to help you perform the following task:

Sets the default network tier of the project. The default network tier is used when an address/forwardingRule/instance is created without specifying the network tier field.

Arguments

request - No description provided.

project - Project ID for this request.

pub fn set_usage_export_bucket(
&self,
request: UsageExportLocation,
project: &str
) -> ProjectSetUsageExportBucketCall<'a, C>

Create a builder to help you perform the following task:

Enables the usage export feature and sets the usage export bucket where reports are stored. If you provide an empty request body using this method, the usage export feature will be disabled.

Arguments

request - No description provided.

project - Project ID for this request.