Struct google_compute1::api::RegionInstanceGroupManagerMethods

pub struct RegionInstanceGroupManagerMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on regionInstanceGroupManager resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2;
use compute1::Compute;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Compute::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `abandon_instances(...)`, `apply_updates_to_instances(...)`, `create_instances(...)`, `delete(...)`, `delete_instances(...)`, `delete_per_instance_configs(...)`, `get(...)`, `insert(...)`, `list(...)`, `list_errors(...)`, `list_managed_instances(...)`, `list_per_instance_configs(...)`, `patch(...)`, `patch_per_instance_configs(...)`, `recreate_instances(...)`, `resize(...)`, `set_instance_template(...)`, `set_target_pools(...)` and `update_per_instance_configs(...)`
// to build up your call.
let rb = hub.region_instance_group_managers();

Implementations

impl<'a, C> RegionInstanceGroupManagerMethods<'a, C>

pub fn abandon_instances(
&self,
request: RegionInstanceGroupManagersAbandonInstancesRequest,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerAbandonInstanceCall<'a, C>

Create a builder to help you perform the following task:

Flags the specified instances to be immediately removed from the managed instance group. Abandoning an instance does not delete the instance, but it does remove the instance from any target pools that are applied by the managed instance group. This method reduces the targetSize of the managed instance group by the number of instances that you abandon. This operation is marked as DONE when the action is scheduled even if the instances have not yet been removed from the group. You must separately verify the status of the abandoning action with the listmanagedinstances method.

If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted.

You can specify a maximum of 1000 instances with this method per request.

Arguments

request - No description provided.

project - Project ID for this request.

region - Name of the region scoping this request.

instanceGroupManager - Name of the managed instance group.

pub fn apply_updates_to_instances(
&self,
request: RegionInstanceGroupManagersApplyUpdatesRequest,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerApplyUpdatesToInstanceCall<'a, C>

Create a builder to help you perform the following task:

Apply updates to selected instances the managed instance group.

Arguments

request - No description provided.

project - Project ID for this request.

region - Name of the region scoping this request, should conform to RFC1035.

instanceGroupManager - The name of the managed instance group, should conform to RFC1035.

pub fn create_instances(
&self,
request: RegionInstanceGroupManagersCreateInstancesRequest,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerCreateInstanceCall<'a, C>

Create a builder to help you perform the following task:

Creates instances with per-instance configs in this regional managed instance group. Instances are created using the current instance template. The create instances operation is marked DONE if the createInstances request is successful. The underlying actions take additional time. You must separately verify the status of the creating or actions with the listmanagedinstances method.

Arguments

request - No description provided.

project - Project ID for this request.

region - The name of the region where the managed instance group is located. It should conform to RFC1035.

instanceGroupManager - The name of the managed instance group. It should conform to RFC1035.

pub fn delete(
&self,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes the specified managed instance group and all of the instances in that group.

Arguments

project - Project ID for this request.

region - Name of the region scoping this request.

instanceGroupManager - Name of the managed instance group to delete.

pub fn delete_instances(
&self,
request: RegionInstanceGroupManagersDeleteInstancesRequest,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerDeleteInstanceCall<'a, C>

Create a builder to help you perform the following task:

Flags the specified instances in the managed instance group to be immediately deleted. The instances are also removed from any target pools of which they were a member. This method reduces the targetSize of the managed instance group by the number of instances that you delete. The deleteInstances operation is marked DONE if the deleteInstances request is successful. The underlying actions take additional time. You must separately verify the status of the deleting action with the listmanagedinstances method.

If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted.

You can specify a maximum of 1000 instances with this method per request.

Arguments

request - No description provided.

project - Project ID for this request.

region - Name of the region scoping this request.

instanceGroupManager - Name of the managed instance group.

pub fn delete_per_instance_configs(
&self,
request: RegionInstanceGroupManagerDeleteInstanceConfigReq,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerDeletePerInstanceConfigCall<'a, C>

Create a builder to help you perform the following task:

Deletes selected per-instance configs for the managed instance group.

Arguments

request - No description provided.

project - Project ID for this request.

region - Name of the region scoping this request, should conform to RFC1035.

instanceGroupManager - The name of the managed instance group. It should conform to RFC1035.

pub fn get(
&self,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerGetCall<'a, C>

Create a builder to help you perform the following task:

Returns all of the details about the specified managed instance group.

Arguments

project - Project ID for this request.

region - Name of the region scoping this request.

instanceGroupManager - Name of the managed instance group to return.

pub fn insert(
&self,
request: InstanceGroupManager,
project: &str,
region: &str
) -> RegionInstanceGroupManagerInsertCall<'a, C>

Create a builder to help you perform the following task:

Creates a managed instance group using the information that you specify in the request. After the group is created, instances in the group are created using the specified instance template. This operation is marked as DONE when the group is created even if the instances in the group have not yet been created. You must separately verify the status of the individual instances with the listmanagedinstances method.

A regional managed instance group can contain up to 2000 instances.

Arguments

request - No description provided.

project - Project ID for this request.

region - Name of the region scoping this request.

pub fn list(
&self,
project: &str,
region: &str
) -> RegionInstanceGroupManagerListCall<'a, C>

Create a builder to help you perform the following task:

Retrieves the list of managed instance groups that are contained within the specified region.

Arguments

project - Project ID for this request.

region - Name of the region scoping this request.

pub fn list_errors(
&self,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerListErrorCall<'a, C>

Create a builder to help you perform the following task:

Lists all errors thrown by actions on instances for a given regional managed instance group. The filter and orderBy query parameters are not supported.

Arguments

project - Project ID for this request.

region - Name of the region scoping this request. This should conform to RFC1035.

instanceGroupManager - The name of the managed instance group. It must be a string that meets the requirements in RFC1035, or an unsigned long integer: must match regexp pattern: (?:a-z?)|[1-9][0-9]{0,19}.

pub fn list_managed_instances(
&self,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerListManagedInstanceCall<'a, C>

Create a builder to help you perform the following task:

Lists the instances in the managed instance group and instances that are scheduled to be created. The list includes any current actions that the group has scheduled for its instances. The orderBy query parameter is not supported.

Arguments

project - Project ID for this request.

region - Name of the region scoping this request.

instanceGroupManager - The name of the managed instance group.

pub fn list_per_instance_configs(
&self,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerListPerInstanceConfigCall<'a, C>

Create a builder to help you perform the following task:

Lists all of the per-instance configs defined for the managed instance group. The orderBy query parameter is not supported.

Arguments

project - Project ID for this request.

region - Name of the region scoping this request, should conform to RFC1035.

instanceGroupManager - The name of the managed instance group. It should conform to RFC1035.

pub fn patch(
&self,
request: InstanceGroupManager,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a managed instance group using the information that you specify in the request. This operation is marked as DONE when the group is patched even if the instances in the group are still in the process of being patched. You must separately verify the status of the individual instances with the listmanagedinstances method. This method supports PATCH semantics and uses the JSON merge patch format and processing rules.

Arguments

request - No description provided.

project - Project ID for this request.

region - Name of the region scoping this request.

instanceGroupManager - The name of the instance group manager.

pub fn patch_per_instance_configs(
&self,
request: RegionInstanceGroupManagerPatchInstanceConfigReq,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerPatchPerInstanceConfigCall<'a, C>

Create a builder to help you perform the following task:

Inserts or patches per-instance configs for the managed instance group. perInstanceConfig.name serves as a key used to distinguish whether to perform insert or patch.

Arguments

request - No description provided.

project - Project ID for this request.

region - Name of the region scoping this request, should conform to RFC1035.

instanceGroupManager - The name of the managed instance group. It should conform to RFC1035.

pub fn recreate_instances(
&self,
request: RegionInstanceGroupManagersRecreateRequest,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerRecreateInstanceCall<'a, C>

Create a builder to help you perform the following task:

Flags the specified VM instances in the managed instance group to be immediately recreated. Each instance is recreated using the group's current configuration. This operation is marked as DONE when the flag is set even if the instances have not yet been recreated. You must separately verify the status of each instance by checking its currentAction field; for more information, see Checking the status of managed instances.

If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted.

You can specify a maximum of 1000 instances with this method per request.

Arguments

request - No description provided.

project - Project ID for this request.

region - Name of the region scoping this request.

instanceGroupManager - Name of the managed instance group.

pub fn resize(
&self,
project: &str,
region: &str,
instance_group_manager: &str,
size: i32
) -> RegionInstanceGroupManagerResizeCall<'a, C>

Create a builder to help you perform the following task:

Changes the intended size of the managed instance group. If you increase the size, the group creates new instances using the current instance template. If you decrease the size, the group deletes one or more instances.

The resize operation is marked DONE if the resize request is successful. The underlying actions take additional time. You must separately verify the status of the creating or deleting actions with the listmanagedinstances method.

If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted.

Arguments

project - Project ID for this request.

region - Name of the region scoping this request.

instanceGroupManager - Name of the managed instance group.

size - Number of instances that should exist in this instance group manager.

pub fn set_instance_template(
&self,
request: RegionInstanceGroupManagersSetTemplateRequest,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerSetInstanceTemplateCall<'a, C>

Create a builder to help you perform the following task:

Sets the instance template to use when creating new instances or recreating instances in this group. Existing instances are not affected.

Arguments

request - No description provided.

project - Project ID for this request.

region - Name of the region scoping this request.

instanceGroupManager - The name of the managed instance group.

pub fn set_target_pools(
&self,
request: RegionInstanceGroupManagersSetTargetPoolsRequest,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerSetTargetPoolCall<'a, C>

Create a builder to help you perform the following task:

Modifies the target pools to which all new instances in this group are assigned. Existing instances in the group are not affected.

Arguments

request - No description provided.

project - Project ID for this request.

region - Name of the region scoping this request.

instanceGroupManager - Name of the managed instance group.

pub fn update_per_instance_configs(
&self,
request: RegionInstanceGroupManagerUpdateInstanceConfigReq,
project: &str,
region: &str,
instance_group_manager: &str
) -> RegionInstanceGroupManagerUpdatePerInstanceConfigCall<'a, C>

Create a builder to help you perform the following task:

Inserts or updates per-instance configs for the managed instance group. perInstanceConfig.name serves as a key used to distinguish whether to perform insert or patch.

Arguments

request - No description provided.

project - Project ID for this request.

region - Name of the region scoping this request, should conform to RFC1035.

instanceGroupManager - The name of the managed instance group. It should conform to RFC1035.