Struct google_compute1::api::UrlMap

pub struct UrlMap {
    pub creation_timestamp: Option<String>,
    pub default_route_action: Option<HttpRouteAction>,
    pub default_service: Option<String>,
    pub default_url_redirect: Option<HttpRedirectAction>,
    pub description: Option<String>,
    pub fingerprint: Option<String>,
    pub header_action: Option<HttpHeaderAction>,
    pub host_rules: Option<Vec<HostRule>>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub path_matchers: Option<Vec<PathMatcher>>,
    pub region: Option<String>,
    pub self_link: Option<String>,
    pub tests: Option<Vec<UrlMapTest>>,
}

Represents a URL Map resource.

Google Compute Engine has two URL Map resources:

Global * Regional

A URL map resource is a component of certain types of GCP load balancers and Traffic Director.

urlMaps are used by external HTTP(S) load balancers and Traffic Director. * regionUrlMaps are used by internal HTTP(S) load balancers.

For a list of supported URL map features by load balancer type, see the Load balancing features: Routing and traffic management table.

For a list of supported URL map features for Traffic Director, see the Traffic Director features: Routing and traffic management table.

This resource defines mappings from host names and URL paths to either a backend service or a backend bucket.

To use the global urlMaps resource, the backend service must have a loadBalancingScheme of either EXTERNAL or INTERNAL_SELF_MANAGED. To use the regionUrlMaps resource, the backend service must have a loadBalancingScheme of INTERNAL_MANAGED. For more information, read URL Map Concepts.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get region url maps (response)

insert region url maps (request)

patch region url maps (request)

update region url maps (request)

aggregated list url maps (none)

delete url maps (none)

get url maps (response)

insert url maps (request)

invalidate cache url maps (none)

list url maps (none)

patch url maps (request)

update url maps (request)

validate url maps (none)

Fields
creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

default_route_action: Option<HttpRouteAction>

defaultRouteAction takes effect when none of the hostRules match. The load balancer performs advanced routing actions like URL rewrites, header transformations, etc. prior to forwarding the request to the selected backend. If defaultRouteAction specifies any weightedBackendServices, defaultService must not be set. Conversely if defaultService is set, defaultRouteAction cannot contain any weightedBackendServices. Only one of defaultRouteAction or defaultUrlRedirect must be set. UrlMaps for external HTTP(S) load balancers support only the urlRewrite action within defaultRouteAction. defaultRouteAction has no effect when the URL map is bound to target gRPC proxy that has validateForProxyless field set to true.

default_service: Option<String>

The full or partial URL of the defaultService resource to which traffic is directed if none of the hostRules match. If defaultRouteAction is additionally specified, advanced routing actions like URL Rewrites, etc. take effect prior to sending the request to the backend. However, if defaultService is specified, defaultRouteAction cannot contain any weightedBackendServices. Conversely, if routeAction specifies any weightedBackendServices, service must not be specified. Only one of defaultService, defaultUrlRedirect or defaultRouteAction.weightedBackendService must be set. defaultService has no effect when the URL map is bound to target gRPC proxy that has validateForProxyless field set to true.

default_url_redirect: Option<HttpRedirectAction>

When none of the specified hostRules match, the request is redirected to a URL specified by defaultUrlRedirect. If defaultUrlRedirect is specified, defaultService or defaultRouteAction must not be set. Not supported when the URL map is bound to target gRPC proxy.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

fingerprint: Option<String>

Fingerprint of this resource. A hash of the contents stored in this object. This field is used in optimistic locking. This field will be ignored when inserting a UrlMap. An up-to-date fingerprint must be provided in order to update the UrlMap, otherwise the request will fail with error 412 conditionNotMet.

To see the latest fingerprint, make a get() request to retrieve a UrlMap.

header_action: Option<HttpHeaderAction>

Specifies changes to request and response headers that need to take effect for the selected backendService. The headerAction specified here take effect after headerAction specified under pathMatcher. Note that headerAction is not supported for Loadbalancers that have their loadBalancingScheme set to EXTERNAL. Not supported when the URL map is bound to target gRPC proxy that has validateForProxyless field set to true.

host_rules: Option<Vec<HostRule>>

The list of HostRules to use against the URL.

id: Option<String>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

kind: Option<String>

[Output Only] Type of the resource. Always compute#urlMaps for url maps.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z]([-a-z0-9]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

path_matchers: Option<Vec<PathMatcher>>

The list of named PathMatchers to use against the URL.

region: Option<String>

[Output Only] URL of the region where the regional URL map resides. This field is not applicable to global URL maps. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

tests: Option<Vec<UrlMapTest>>

The list of expected URL mapping tests. Request to update this UrlMap will succeed only if all of the test cases pass. You can specify a maximum of 100 tests per UrlMap. Not supported when the URL map is bound to target gRPC proxy that has validateForProxyless field set to true.