Struct google_compute1::Address

pub struct Address {
    pub status: Option<String>,
    pub address_type: Option<String>,
    pub users: Option<Vec<String>>,
    pub ip_version: Option<String>,
    pub purpose: Option<String>,
    pub address: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub network: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub prefix_length: Option<i32>,
    pub region: Option<String>,
    pub description: Option<String>,
    pub network_tier: Option<String>,
    pub subnetwork: Option<String>,
    pub self_link: Option<String>,
}

Represents an IP Address resource.

An address resource represents a regional internal IP address. Regional internal IP addresses are RFC 1918 addresses that come from either a primary or secondary IP range of a subnet in a VPC network. Regional external IP addresses can be assigned to GCP VM instances, Cloud VPN gateways, regional external forwarding rules for network load balancers (in either Standard or Premium Tier), and regional external forwarding rules for HTTP(S), SSL Proxy, and TCP Proxy load balancers in Standard Tier. For more information, read IP addresses.

A globalAddresses resource represent a global external IP address. Global external IP addresses are IPv4 or IPv6 addresses. They can only be assigned to global forwarding rules for HTTP(S), SSL Proxy, or TCP Proxy load balancers in Premium Tier. For more information, read Global resources. (== resource_for beta.addresses ==) (== resource_for v1.addresses ==) (== resource_for beta.globalAddresses ==) (== resource_for v1.globalAddresses ==)

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert global addresses (request)

insert addresses (request)

get addresses (response)

get global addresses (response)

Fields
status: Option<String>

[Output Only] The status of the address, which can be one of RESERVING, RESERVED, or IN_USE. An address that is RESERVING is currently in the process of being reserved. A RESERVED address is currently reserved and available to use. An IN_USE address is currently being used by another resource and is not available.

address_type: Option<String>

The type of address to reserve, either INTERNAL or EXTERNAL. If unspecified, defaults to EXTERNAL.

users: Option<Vec<String>>

[Output Only] The URLs of the resources that are using this address.

ip_version: Option<String>

The IP version that will be used by this address. Valid options are IPV4 or IPV6. This can only be specified for a global address.

purpose: Option<String>

The purpose of this resource, which can be one of the following values:

GCE_ENDPOINT for addresses that are used by VM instances, alias IP ranges, internal load balancers, and similar resources.

DNS_RESOLVER for a DNS resolver address in a subnetwork

VPC_PEERING for addresses that are reserved for VPC peer networks.

NAT_AUTO for addresses that are external IP addresses automatically reserved for Cloud NAT.

address: Option<String>

The static IP address represented by this resource.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

id: Option<String>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

network: Option<String>

The URL of the network in which to reserve the address. This field can only be used with INTERNAL type with the VPC_PEERING purpose.

kind: Option<String>

[Output Only] Type of the resource. Always compute#address for addresses.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z]([-a-z0-9]*[a-z0-9])?. The first character must be a lowercase letter, and all following characters (except for the last character) must be a dash, lowercase letter, or digit. The last character must be a lowercase letter or digit.

prefix_length: Option<i32>

The prefix length if the resource reprensents an IP range.

region: Option<String>

[Output Only] The URL of the region where the regional address resides. This field is not applicable to global addresses. You must specify this field as part of the HTTP request URL.

description: Option<String>

An optional description of this resource. Provide this field when you create the resource.

network_tier: Option<String>

This signifies the networking tier used for configuring this address and can only take the following values: PREMIUM or STANDARD. Global forwarding rules can only be Premium Tier. Regional forwarding rules can be either Premium or Standard Tier. Standard Tier addresses applied to regional forwarding rules can be used with any external load balancer. Regional forwarding rules in Premium Tier can only be used with a network load balancer.

If this field is not specified, it is assumed to be PREMIUM.

subnetwork: Option<String>

The URL of the subnetwork in which to reserve the address. If an IP address is specified, it must be within the subnetwork's IP range. This field can only be used with INTERNAL type with a GCE_ENDPOINT or DNS_RESOLVER purpose.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.