Struct google_compute1::AddresseMethods

pub struct AddresseMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on addresse resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `aggregated_list(...)`, `delete(...)`, `get(...)`, `insert(...)` and `list(...)`
// to build up your call.
let rb = hub.addresses();

Methods

impl<'a, C, A> AddresseMethods<'a, C, A>

fn list(&self, project: &str, region: &str) -> AddresseListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of address resources contained within the specified region.

Arguments

project - Project ID for this request.

region - The name of the region for this request.

fn aggregated_list(&self, project: &str) -> AddresseAggregatedListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves an aggregated list of addresses.

Arguments

project - Project ID for this request.

fn get(&self, project: &str, region: &str, address: &str) -> AddresseGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the specified address resource.

Arguments

project - Project ID for this request.

region - The name of the region for this request.

address - Name of the address resource to return.

fn insert(&self, request: Address, project: &str, region: &str) -> AddresseInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an address resource in the specified project using the data included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

region - The name of the region for this request.

fn delete(&self, project: &str, region: &str, address: &str) -> AddresseDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified address resource.

Arguments

project - Project ID for this request.

region - The name of the region for this request.

address - Name of the address resource to delete.