Struct google_compute1::AttachedDiskInitializeParams

pub struct AttachedDiskInitializeParams {
    pub disk_size_gb: Option<String>,
    pub disk_name: Option<String>,
    pub source_image: Option<String>,
    pub disk_type: Option<String>,
}

[Input Only] Specifies the parameters for a new disk that will be created alongside the new instance. Use initialization parameters to create boot disks or local SSDs attached to the new instance.

This property is mutually exclusive with the source property; you can only define one or the other, but not both.

This type is not used in any activity, and only used as part of another schema.

Fields

disk_size_gb

Specifies the size of the disk in base-2 GB.

disk_name

Specifies the disk name. If not specified, the default is to use the name of the instance.

source_image

A source image used to create the disk. You can provide a private (custom) image, and Compute Engine will use the corresponding image from your project. For example:

global/images/my-private-image

Or you can provide an image from a publicly-available project. For example, to use a Debian image from the debian-cloud project, make sure to include the project in the URL:

projects/debian-cloud/global/images/debian-7-wheezy-vYYYYMMDD

where vYYYYMMDD is the image version. The fully-qualified URL will also work in both cases.

disk_type

Specifies the disk type to use to create the instance. If not specified, the default is pd-standard, specified using the full URL. For example:

https://www.googleapis.com/compute/v1/projects/project/zones/zone/diskTypes/pd-standard

Other values include pd-ssd and local-ssd. If you define this field, you can provide either the full or partial URL. For example, the following are valid values:
- https://www.googleapis.com/compute/v1/projects/project/zones/zone/diskTypes/diskType - projects/project/zones/zone/diskTypes/diskType - zones/zone/diskTypes/diskType