Struct google_compute1::AutoscalerMethods

pub struct AutoscalerMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on autoscaler resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `aggregated_list(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.autoscalers();

Methods

impl<'a, C, A> AutoscalerMethods<'a, C, A>

fn delete(&self, project: &str, zone: &str, autoscaler: &str) -> AutoscalerDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified autoscaler resource.

Arguments

project - Project ID for this request.

zone - Name of the zone scoping this request.

autoscaler - Name of the persistent autoscaler resource to delete.

fn get(&self, project: &str, zone: &str, autoscaler: &str) -> AutoscalerGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the specified autoscaler resource.

Arguments

project - Project ID for this request.

zone - Name of the zone scoping this request.

autoscaler - Name of the persistent autoscaler resource to return.

fn list(&self, project: &str, zone: &str) -> AutoscalerListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of autoscaler resources contained within the specified zone.

Arguments

project - Project ID for this request.

zone - Name of the zone scoping this request.

fn patch(&self, request: Autoscaler, project: &str, zone: &str, autoscaler: &str) -> AutoscalerPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an autoscaler resource in the specified project using the data included in the request. This method supports patch semantics.

Arguments

request - No description provided.

project - Project ID for this request.

zone - Name of the zone scoping this request.

autoscaler - Name of the autoscaler resource to update.

fn insert(&self, request: Autoscaler, project: &str, zone: &str) -> AutoscalerInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an autoscaler resource in the specified project using the data included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - Name of the zone scoping this request.

fn aggregated_list(&self, project: &str) -> AutoscalerAggregatedListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves an aggregated list of autoscalers.

Arguments

project - Project ID for this request.

fn update(&self, request: Autoscaler, project: &str, zone: &str) -> AutoscalerUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an autoscaler resource in the specified project using the data included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - Name of the zone scoping this request.