Struct google_compute1::Backend

pub struct Backend {
    pub capacity_scaler: Option<f32>,
    pub max_rate: Option<i32>,
    pub group: Option<String>,
    pub description: Option<String>,
    pub max_utilization: Option<f32>,
    pub max_rate_per_instance: Option<f32>,
    pub balancing_mode: Option<String>,
}

Message containing information of one individual backend.

This type is not used in any activity, and only used as part of another schema.

Fields

capacity_scaler

A multiplier applied to the group's maximum servicing capacity (either UTILIZATION or RATE). Default value is 1, which means the group will serve up to 100% of its configured CPU or RPS (depending on balancingMode). A setting of 0 means the group is completely drained, offering 0% of its available CPU or RPS. Valid range is [0.0,1.0].

max_rate

The max requests per second (RPS) of the group. Can be used with either RATE or UTILIZATION balancing modes, but required if RATE mode. For RATE mode, either maxRate or maxRatePerInstance must be set.

group

The fully-qualified URL of a zonal Instance Group resource. This instance group defines the list of instances that serve traffic. Member virtual machine instances from each instance group must live in the same zone as the instance group itself. No two backends in a backend service are allowed to use same Instance Group resource.

Note that you must specify an Instance Group resource using the fully-qualified URL, rather than a partial URL.

description

An optional description of this resource. Provide this property when you create the resource.

max_utilization

Used when balancingMode is UTILIZATION. This ratio defines the CPU utilization target for the group. The default is 0.8. Valid range is [0.0, 1.0].

max_rate_per_instance

The max requests per second (RPS) that a single backend instance can handle.This is used to calculate the capacity of the group. Can be used in either balancing mode. For RATE mode, either maxRate or maxRatePerInstance must be set.

balancing_mode

Specifies the balancing mode for this backend. For global HTTP(S) load balancing, the default is UTILIZATION. Valid values are UTILIZATION and RATE.