Struct google-compute1::Compute

pub struct Compute<C, NC, A> {
    // some fields omitted
}

Central instance to access all Compute related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate serde;
extern crate "google-compute1" as compute1;
use compute1::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.instances().set_disk_auto_delete("project", "zone", "instance", true, "deviceName")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::JsonDecodeError(err) => println!("Json failed to decode: {:?}", err),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Methods

impl<'a, C, NC, A> Compute<C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>>, A: GetToken

fn new(client: C, authenticator: A) -> Compute<C, NC, A>

fn addresses(&'a self) -> AddresseMethodsBuilder<'a, C, NC, A>

fn backend_services(&'a self) -> BackendServiceMethodsBuilder<'a, C, NC, A>

fn disk_types(&'a self) -> DiskTypeMethodsBuilder<'a, C, NC, A>

fn disks(&'a self) -> DiskMethodsBuilder<'a, C, NC, A>

fn firewalls(&'a self) -> FirewallMethodsBuilder<'a, C, NC, A>

fn forwarding_rules(&'a self) -> ForwardingRuleMethodsBuilder<'a, C, NC, A>

fn global_addresses(&'a self) -> GlobalAddresseMethodsBuilder<'a, C, NC, A>

fn global_forwarding_rules(&'a self) -> GlobalForwardingRuleMethodsBuilder<'a, C, NC, A>

fn global_operations(&'a self) -> GlobalOperationMethodsBuilder<'a, C, NC, A>

fn http_health_checks(&'a self) -> HttpHealthCheckMethodsBuilder<'a, C, NC, A>

fn images(&'a self) -> ImageMethodsBuilder<'a, C, NC, A>

fn instance_templates(&'a self) -> InstanceTemplateMethodsBuilder<'a, C, NC, A>

fn instances(&'a self) -> InstanceMethodsBuilder<'a, C, NC, A>

fn licenses(&'a self) -> LicenseMethodsBuilder<'a, C, NC, A>

fn machine_types(&'a self) -> MachineTypeMethodsBuilder<'a, C, NC, A>

fn networks(&'a self) -> NetworkMethodsBuilder<'a, C, NC, A>

fn projects(&'a self) -> ProjectMethodsBuilder<'a, C, NC, A>

fn region_operations(&'a self) -> RegionOperationMethodsBuilder<'a, C, NC, A>

fn regions(&'a self) -> RegionMethodsBuilder<'a, C, NC, A>

fn routes(&'a self) -> RouteMethodsBuilder<'a, C, NC, A>

fn snapshots(&'a self) -> SnapshotMethodsBuilder<'a, C, NC, A>

fn target_http_proxies(&'a self) -> TargetHttpProxyMethodsBuilder<'a, C, NC, A>

fn target_instances(&'a self) -> TargetInstanceMethodsBuilder<'a, C, NC, A>

fn target_pools(&'a self) -> TargetPoolMethodsBuilder<'a, C, NC, A>

fn url_maps(&'a self) -> UrlMapMethodsBuilder<'a, C, NC, A>

fn zone_operations(&'a self) -> ZoneOperationMethodsBuilder<'a, C, NC, A>

fn zones(&'a self) -> ZoneMethodsBuilder<'a, C, NC, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.0.1.

Returns the previously set user-agent.