Struct google_compute1::Disk

pub struct Disk {
    pub status: Option<String>,
    pub source_snapshot: Option<String>,
    pub last_detach_timestamp: Option<String>,
    pub description: Option<String>,
    pub source_image_id: Option<String>,
    pub licenses: Option<Vec<String>>,
    pub source_image: Option<String>,
    pub source_snapshot_id: Option<String>,
    pub users: Option<Vec<String>>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub name: Option<String>,
    pub kind: Option<String>,
    pub size_gb: Option<String>,
    pub zone: Option<String>,
    pub last_attach_timestamp: Option<String>,
    pub type_: Option<String>,
    pub options: Option<String>,
    pub self_link: Option<String>,
}

A Disk resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert disks (request)

delete disks (none)

resize disks (none)

get disks (response)

create snapshot disks (none)

aggregated list disks (none)

list disks (none)

Fields

status

[Output Only] The status of disk creation. Applicable statuses includes: CREATING, FAILED, READY, RESTORING.

source_snapshot

The source snapshot used to create this disk. You can provide this as a partial or full URL to the resource. For example, the following are valid values: - https://www.googleapis.com/compute/v1/projects/project/global/snapshots/snapshot - projects/project/global/snapshots/snapshot - global/snapshots/snapshot

last_detach_timestamp

[Output Only] Last detach timestamp in RFC3339 text format.

description

An optional description of this resource. Provide this property when you create the resource.

source_image_id

[Output Only] The ID value of the image used to create this disk. This value identifies the exact image that was used to create this persistent disk. For example, if you created the persistent disk from an image that was later deleted and recreated under the same name, the source image ID would identify the exact version of the image that was used.

licenses

[Output Only] Any applicable publicly visible licenses.

source_image

The source image used to create this disk. If the source image is deleted, this field will not be set.

To create a disk with one of the public operating system images, specify the image by its family name. For example, specify family/debian-8 to use the latest Debian 8 image:

projects/debian-cloud/global/images/family/debian-8

Alternatively, use a specific version of a public operating system image:

projects/debian-cloud/global/images/debian-8-jessie-vYYYYMMDD

To create a disk with a private image that you created, specify the image name in the following format:

global/images/my-private-image

You can also specify a private image by its image family, which returns the latest version of the image in that family. Replace the image name with family/family-name:

global/images/family/my-private-family

source_snapshot_id

[Output Only] The unique ID of the snapshot used to create this disk. This value identifies the exact snapshot that was used to create this persistent disk. For example, if you created the persistent disk from a snapshot that was later deleted and recreated under the same name, the source snapshot ID would identify the exact version of the snapshot that was used.

users

[Output Only] Links to the users of the disk (attached instances) in form: project/zones/zone/instances/instance

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

name

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

kind

[Output Only] Type of the resource. Always compute#disk for disks.

size_gb

Size of the persistent disk, specified in GB. You can specify this field when creating a persistent disk using the sourceImage or sourceSnapshot parameter, or specify it alone to create an empty persistent disk.

If you specify this field along with sourceImage or sourceSnapshot, the value of sizeGb must not be less than the size of the sourceImage or the size of the snapshot.

zone

[Output Only] URL of the zone where the disk resides.

last_attach_timestamp

[Output Only] Last attach timestamp in RFC3339 text format.

type_

URL of the disk type resource describing which disk type to use to create the disk. Provide this when creating the disk.

options

Internal use only.

self_link

[Output Only] Server-defined fully-qualified URL for this resource.