Struct google_compute1::DiskMethods

pub struct DiskMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on disk resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `aggregated_list(...)`, `create_snapshot(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)` and `resize(...)`
// to build up your call.
let rb = hub.disks();

Methods

impl<'a, C, A> DiskMethods<'a, C, A>

fn delete(&self, project: &str, zone: &str, disk: &str) -> DiskDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified persistent disk. Deleting a disk removes its data permanently and is irreversible. However, deleting a disk does not delete any snapshots previously made from the disk. You must separately delete snapshots.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

disk - Name of the persistent disk to delete.

fn aggregated_list(&self, project: &str) -> DiskAggregatedListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves an aggregated list of persistent disks.

Arguments

project - Project ID for this request.

fn get(&self, project: &str, zone: &str, disk: &str) -> DiskGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns a specified persistent disk. Get a list of available persistent disks by making a list() request.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

disk - Name of the persistent disk to return.

fn resize(&self, request: DisksResizeRequest, project: &str, zone: &str, disk: &str) -> DiskResizeCall<'a, C, A>

Create a builder to help you perform the following task:

Resizes the specified persistent disk.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

disk - The name of the persistent disk.

fn list(&self, project: &str, zone: &str) -> DiskListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of persistent disks contained within the specified zone.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

fn insert(&self, request: Disk, project: &str, zone: &str) -> DiskInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a persistent disk in the specified project using the data in the request. You can create a disk with a sourceImage, a sourceSnapshot, or create an empty 500 GB data disk by omitting all properties. You can also create a disk that is larger than the default size by specifying the sizeGb property.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

fn create_snapshot(&self, request: Snapshot, project: &str, zone: &str, disk: &str) -> DiskCreateSnapshotCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a snapshot of a specified persistent disk.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

disk - Name of the persistent disk to snapshot.