Struct google-compute1::DiskType

pub struct DiskType {
    pub kind: String,
    pub description: String,
    pub zone: String,
    pub valid_disk_size: String,
    pub deprecated: DeprecationStatus,
    pub default_disk_size_gb: String,
    pub creation_timestamp: String,
    pub id: String,
    pub self_link: String,
    pub name: String,
}

A disk type resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

aggregated list disk types (none)

list disk types (none)

get disk types (response)

Fields

kind

[Output Only] Type of the resource. Always compute#diskType for disk types.

description

[Output Only] An optional textual description of the resource.

zone

[Output Only] URL of the zone where the disk type resides.

valid_disk_size

[Output Only] An optional textual description of the valid disk size, such as "10GB-10TB".

deprecated

[Output Only] The deprecation status associated with this disk type.

default_disk_size_gb

[Output Only] Server defined default disk size in GB.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] Unique identifier for the resource; defined by the server.

self_link

[Output Only] Server defined URL for the resource.

name

[Output Only] Name of the resource.