Struct google_compute1::Firewall

pub struct Firewall {
    pub kind: Option<String>,
    pub description: Option<String>,
    pub source_tags: Option<Vec<String>>,
    pub source_ranges: Option<Vec<String>>,
    pub name: Option<String>,
    pub target_tags: Option<Vec<String>>,
    pub allowed: Option<Vec<FirewallAllowed>>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub network: Option<String>,
}

A Firewall resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch firewalls (request)

list firewalls (none)

get firewalls (response)

update firewalls (request)

delete firewalls (none)

insert firewalls (request)

Fields

kind

[Output Ony] Type of the resource. Always compute#firewall for firewall rules.

description

An optional textual description of the resource; provided by the client when the resource is created.

source_tags

A list of instance tags which this rule applies to. One or both of sourceRanges and sourceTags may be set.

If both properties are set, an inbound connection is allowed if the range or the tag of the source matches the sourceRanges OR matches the sourceTags property; the connection does not need to match both properties.

source_ranges

The IP address blocks that this rule applies to, expressed in CIDR format. One or both of sourceRanges and sourceTags may be set.

If both properties are set, an inbound connection is allowed if the range or the tag of the source matches the sourceRanges OR matches the sourceTags property; the connection does not need to match both properties.

name

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

target_tags

A list of instance tags indicating sets of instances located on network which may make network connections as specified in allowed[]. If no targetTags are specified, the firewall rule applies to all instances on the specified network.

allowed

The list of rules specified by this firewall. Each rule specifies a protocol and port-range tuple that describes a permitted connection.

creation_timestamp

[Output Only] Creation timestamp in RFC3339text format.

id

[Output Only] Unique identifier for the resource; defined by the server.

self_link

[Output Only] Server defined URL for the resource.

network

URL of the network resource for this firewall rule. This field is required for creating an instance but optional when creating a firewall rule. If not specified when creating a firewall rule, the default network is used: global/networks/default If you choose to specify this property, you can specify the network as a full or partial URL. For example, the following are all valid URLs:
- https://www.googleapis.com/compute/v1/projects/myproject/global/networks/my-network - projects/myproject/global/networks/my-network - global/networks/default