Struct google_compute1::ForwardingRule

pub struct ForwardingRule {
    pub kind: Option<String>,
    pub description: Option<String>,
    pub port_range: Option<String>,
    pub region: Option<String>,
    pub name: Option<String>,
    pub creation_timestamp: Option<String>,
    pub ip_protocol: Option<String>,
    pub ip_address: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub target: Option<String>,
}

A ForwardingRule resource. A ForwardingRule resource specifies which pool of target virtual machines to forward a packet to if it matches the given [IPAddress, IPProtocol, portRange] tuple.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

aggregated list forwarding rules (none)

get global forwarding rules (response)

insert global forwarding rules (request)

insert forwarding rules (request)

delete forwarding rules (none)

set target forwarding rules (none)

get forwarding rules (response)

list forwarding rules (none)

Fields

kind

[Output Only] Type of the resource. Always compute#forwardingRule for Forwarding Rule resources.

description

An optional description of this resource. Provide this property when you create the resource.

port_range

Applicable only when IPProtocol is TCP, UDP, or SCTP, only packets addressed to ports in the specified range will be forwarded to target. Forwarding rules with the same [IPAddress, IPProtocol] pair must have disjoint port ranges.

region

[Output Only] URL of the region where the regional forwarding rule resides. This field is not applicable to global forwarding rules.

name

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

ip_protocol

The IP protocol to which this rule applies. Valid options are TCP, UDP, ESP, AH or SCTP.

ip_address

Value of the reserved IP address that this forwarding rule is serving on behalf of. For global forwarding rules, the address must be a global IP; for regional forwarding rules, the address must live in the same region as the forwarding rule. If left empty (default value), an ephemeral IP from the same scope (global or regional) will be assigned.

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

self_link

[Output Only] Server-defined URL for the resource.

target

The URL of the target resource to receive the matched traffic. For regional forwarding rules, this target must live in the same region as the forwarding rule. For global forwarding rules, this target must be a global TargetHttpProxy or TargetHttpsProxy resource. The forwarded traffic must be of a type appropriate to the target object. For example, TargetHttpProxy requires HTTP traffic, and TargetHttpsProxy requires HTTPS traffic.