Struct google_compute1::HttpsHealthCheck

pub struct HttpsHealthCheck {
    pub kind: Option<String>,
    pub description: Option<String>,
    pub timeout_sec: Option<i32>,
    pub name: Option<String>,
    pub check_interval_sec: Option<i32>,
    pub unhealthy_threshold: Option<i32>,
    pub healthy_threshold: Option<i32>,
    pub host: Option<String>,
    pub id: Option<String>,
    pub creation_timestamp: Option<String>,
    pub port: Option<i32>,
    pub self_link: Option<String>,
    pub request_path: Option<String>,
}

An HttpsHealthCheck resource. This resource defines a template for how individual instances should be checked for health, via HTTPS.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get https health checks (response)

list https health checks (none)

insert https health checks (request)

patch https health checks (request)

update https health checks (request)

delete https health checks (none)

Fields

kind

Type of the resource.

description

An optional description of this resource. Provide this property when you create the resource.

timeout_sec

How long (in seconds) to wait before claiming failure. The default value is 5 seconds. It is invalid for timeoutSec to have a greater value than checkIntervalSec.

name

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

check_interval_sec

How often (in seconds) to send a health check. The default value is 5 seconds.

unhealthy_threshold

A so-far healthy instance will be marked unhealthy after this many consecutive failures. The default value is 2.

healthy_threshold

A so-far unhealthy instance will be marked healthy after this many consecutive successes. The default value is 2.

host

The value of the host header in the HTTPS health check request. If left empty (default value), the public IP on behalf of which this health check is performed will be used.

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

port

The TCP port number for the HTTPS health check request. The default value is 443.

self_link

[Output Only] Server-defined URL for the resource.

request_path

The request path of the HTTPS health check request. The default value is "/".