Struct google_compute1::InstanceGroupManager

pub struct InstanceGroupManager {
    pub current_actions: Option<InstanceGroupManagerActionsSummary>,
    pub description: Option<String>,
    pub zone: Option<String>,
    pub name: Option<String>,
    pub fingerprint: Option<String>,
    pub instance_group: Option<String>,
    pub target_size: Option<i32>,
    pub id: Option<String>,
    pub base_instance_name: Option<String>,
    pub target_pools: Option<Vec<String>>,
    pub instance_template: Option<String>,
    pub creation_timestamp: Option<String>,
    pub kind: Option<String>,
    pub self_link: Option<String>,
    pub named_ports: Option<Vec<NamedPort>>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list managed instances instance group managers (none)

resize instance group managers (none)

aggregated list instance group managers (none)

delete instances instance group managers (none)

delete instance group managers (none)

abandon instances instance group managers (none)

get instance group managers (response)

set instance template instance group managers (none)

list instance group managers (none)

insert instance group managers (request)

set target pools instance group managers (none)

recreate instances instance group managers (none)

Fields

current_actions

[Output Only] The list of instance actions and the number of instances in this managed instance group that are scheduled for each of those actions.

description

An optional description of this resource. Provide this property when you create the resource.

zone

The name of the zone where the managed instance group is located.

name

The name of the managed instance group. The name must be 1-63 characters long, and comply with RFC1035.

fingerprint

[Output Only] The fingerprint of the target pools information. You can use this optional field for optimistic locking when you update the target pool entries.

instance_group

[Output Only] The URL of the Instance Group resource.

target_size

The target number of running instances for this managed instance group. Deleting or abandoning instances reduces this number. Resizing the group changes this number.

id

[Output Only] A unique identifier for this resource type. The server generates this identifier.

base_instance_name

The base instance name to use for instances in this group. The value must be 1-58 characters long. Instances are named by appending a hyphen and a random four-character string to the base instance name. The base instance name must comply with RFC1035.

target_pools

The URLs for all TargetPool resources to which instances in the instanceGroup field are added. The target pools automatically apply to all of the instances in the managed instance group.

instance_template

The URL of the instance template that is specified for this managed instance group. The group uses this template to create all new instances in the managed instance group.

creation_timestamp

[Output Only] The creation timestamp for this managed instance group in RFC3339 text format.

kind

[Output Only] The resource type, which is always compute#instanceGroupManager for managed instance groups.

self_link

[Output Only] The URL for this managed instance group. The server defines this URL.

named_ports

Named ports configured for the Instance Groups complementary to this Instance Group Manager.