Struct google_compute1::InstanceGroupManagerMethods

pub struct InstanceGroupManagerMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on instanceGroupManager resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `abandon_instances(...)`, `aggregated_list(...)`, `delete(...)`, `delete_instances(...)`, `get(...)`, `insert(...)`, `list(...)`, `list_managed_instances(...)`, `patch(...)`, `recreate_instances(...)`, `resize(...)`, `set_instance_template(...)` and `set_target_pools(...)`
// to build up your call.
let rb = hub.instance_group_managers();

Methods

impl<'a, C, A> InstanceGroupManagerMethods<'a, C, A>

pub fn list_managed_instances(
&self,
project: &str,
zone: &str,
instance_group_manager: &str
) -> InstanceGroupManagerListManagedInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all of the instances in the managed instance group. Each instance in the list has a currentAction, which indicates the action that the managed instance group is performing on the instance. For example, if the group is still creating an instance, the currentAction is CREATING. If a previous action failed, the list displays the errors for that failed action.

Arguments

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

pub fn insert(
&self,
request: InstanceGroupManager,
project: &str,
zone: &str
) -> InstanceGroupManagerInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a managed instance group using the information that you specify in the request. After the group is created, instances in the group are created using the specified instance template. This operation is marked as DONE when the group is created even if the instances in the group have not yet been created. You must separately verify the status of the individual instances with the listmanagedinstances method.

A managed instance group can have up to 1000 VM instances per group. Please contact Cloud Support if you need an increase in this limit.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where you want to create the managed instance group.

pub fn list(
&self,
project: &str,
zone: &str
) -> InstanceGroupManagerListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of managed instance groups that are contained within the specified project and zone.

Arguments

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

pub fn aggregated_list(
&self,
project: &str
) -> InstanceGroupManagerAggregatedListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the list of managed instance groups and groups them by zone.

Arguments

project - Project ID for this request.

pub fn get(
&self,
project: &str,
zone: &str,
instance_group_manager: &str
) -> InstanceGroupManagerGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns all of the details about the specified managed instance group. Gets a list of available managed instance groups by making a list() request.

Arguments

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

pub fn set_instance_template(
&self,
request: InstanceGroupManagersSetInstanceTemplateRequest,
project: &str,
zone: &str,
instance_group_manager: &str
) -> InstanceGroupManagerSetInstanceTemplateCall<'a, C, A>

Create a builder to help you perform the following task:

Specifies the instance template to use when creating new instances in this group. The templates for existing instances in the group do not change unless you recreate them.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

pub fn set_target_pools(
&self,
request: InstanceGroupManagersSetTargetPoolsRequest,
project: &str,
zone: &str,
instance_group_manager: &str
) -> InstanceGroupManagerSetTargetPoolCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies the target pools to which all instances in this managed instance group are assigned. The target pools automatically apply to all of the instances in the managed instance group. This operation is marked DONE when you make the request even if the instances have not yet been added to their target pools. The change might take some time to apply to all of the instances in the group depending on the size of the group.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

pub fn abandon_instances(
&self,
request: InstanceGroupManagersAbandonInstancesRequest,
project: &str,
zone: &str,
instance_group_manager: &str
) -> InstanceGroupManagerAbandonInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Flags the specified instances to be removed from the managed instance group. Abandoning an instance does not delete the instance, but it does remove the instance from any target pools that are applied by the managed instance group. This method reduces the targetSize of the managed instance group by the number of instances that you abandon. This operation is marked as DONE when the action is scheduled even if the instances have not yet been removed from the group. You must separately verify the status of the abandoning action with the listmanagedinstances method.

If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted.

You can specify a maximum of 1000 instances with this method per request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

pub fn patch(
&self,
request: InstanceGroupManager,
project: &str,
zone: &str,
instance_group_manager: &str
) -> InstanceGroupManagerPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a managed instance group using the information that you specify in the request. This operation is marked as DONE when the group is patched even if the instances in the group are still in the process of being patched. You must separately verify the status of the individual instances with the listManagedInstances method. This method supports PATCH semantics and uses the JSON merge patch format and processing rules.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where you want to create the managed instance group.

instanceGroupManager - The name of the instance group manager.

pub fn resize(
&self,
project: &str,
zone: &str,
instance_group_manager: &str,
size: i32
) -> InstanceGroupManagerResizeCall<'a, C, A>

Create a builder to help you perform the following task:

Resizes the managed instance group. If you increase the size, the group creates new instances using the current instance template. If you decrease the size, the group deletes instances. The resize operation is marked DONE when the resize actions are scheduled even if the group has not yet added or deleted any instances. You must separately verify the status of the creating or deleting actions with the listmanagedinstances method.

When resizing down, the instance group arbitrarily chooses the order in which VMs are deleted. The group takes into account some VM attributes when making the selection including:

The status of the VM instance. + The health of the VM instance. + The instance template version the VM is based on. + For regional managed instance groups, the location of the VM instance.

This list is subject to change.

If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted.

Arguments

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

size - The number of running instances that the managed instance group should maintain at any given time. The group automatically adds or removes instances to maintain the number of instances specified by this parameter.

pub fn delete_instances(
&self,
request: InstanceGroupManagersDeleteInstancesRequest,
project: &str,
zone: &str,
instance_group_manager: &str
) -> InstanceGroupManagerDeleteInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Flags the specified instances in the managed instance group for immediate deletion. The instances are also removed from any target pools of which they were a member. This method reduces the targetSize of the managed instance group by the number of instances that you delete. This operation is marked as DONE when the action is scheduled even if the instances are still being deleted. You must separately verify the status of the deleting action with the listmanagedinstances method.

If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted.

You can specify a maximum of 1000 instances with this method per request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

pub fn delete(
&self,
project: &str,
zone: &str,
instance_group_manager: &str
) -> InstanceGroupManagerDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified managed instance group and all of the instances in that group. Note that the instance group must not belong to a backend service. Read Deleting an instance group for more information.

Arguments

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group to delete.

pub fn recreate_instances(
&self,
request: InstanceGroupManagersRecreateInstancesRequest,
project: &str,
zone: &str,
instance_group_manager: &str
) -> InstanceGroupManagerRecreateInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Flags the specified instances in the managed instance group to be immediately recreated. The instances are deleted and recreated using the current instance template for the managed instance group. This operation is marked as DONE when the flag is set even if the instances have not yet been recreated. You must separately verify the status of the recreating action with the listmanagedinstances method.

If the group is part of a backend service that has enabled connection draining, it can take up to 60 seconds after the connection draining duration has elapsed before the VM instance is removed or deleted.

You can specify a maximum of 1000 instances with this method per request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.