Struct google_compute1::InstanceGroupMethods

pub struct InstanceGroupMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on instanceGroup resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_instances(...)`, `aggregated_list(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `list_instances(...)`, `remove_instances(...)` and `set_named_ports(...)`
// to build up your call.
let rb = hub.instance_groups();

Methods

impl<'a, C, A> InstanceGroupMethods<'a, C, A>

fn delete(&self, project: &str, zone: &str, instance_group: &str) -> InstanceGroupDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified instance group. The instances in the group are not deleted. Note that instance group must not belong to a backend service. Read Deleting an instance group for more information.

Arguments

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group to delete.

fn list_instances(&self, request: InstanceGroupsListInstancesRequest, project: &str, zone: &str, instance_group: &str) -> InstanceGroupListInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the instances in the specified instance group.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group from which you want to generate a list of included instances.

fn insert(&self, request: InstanceGroup, project: &str, zone: &str) -> InstanceGroupInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an instance group in the specified project using the parameters that are included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where you want to create the instance group.

fn get(&self, project: &str, zone: &str, instance_group: &str) -> InstanceGroupGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the specified instance group. Get a list of available instance groups by making a list() request.

Arguments

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group.

fn remove_instances(&self, request: InstanceGroupsRemoveInstancesRequest, project: &str, zone: &str, instance_group: &str) -> InstanceGroupRemoveInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Removes one or more instances from the specified instance group, but does not delete those instances.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group where the specified instances will be removed.

fn add_instances(&self, request: InstanceGroupsAddInstancesRequest, project: &str, zone: &str, instance_group: &str) -> InstanceGroupAddInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Adds a list of instances to the specified instance group. All of the instances in the instance group must be in the same network/subnetwork. Read Adding instances for more information.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group where you are adding instances.

fn set_named_ports(&self, request: InstanceGroupsSetNamedPortsRequest, project: &str, zone: &str, instance_group: &str) -> InstanceGroupSetNamedPortCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the named ports for the specified instance group.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the instance group is located.

instanceGroup - The name of the instance group where the named ports are updated.

fn aggregated_list(&self, project: &str) -> InstanceGroupAggregatedListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the list of instance groups and sorts them by zone.

Arguments

project - Project ID for this request.

fn list(&self, project: &str, zone: &str) -> InstanceGroupListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the list of instance groups that are located in the specified project and zone.

Arguments

project - Project ID for this request.

zone - The name of the zone where the instance group is located.