Struct google-compute1::InstanceMethodsBuilder

pub struct InstanceMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on instance resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-compute1" as compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `add_access_config(...)`, `aggregated_list(...)`, `attach_disk(...)`, `delete(...)`, `delete_access_config(...)`, `detach_disk(...)`, `get(...)`, `get_serial_port_output(...)`, `insert(...)`, `list(...)`, `reset(...)`, `set_disk_auto_delete(...)`, `set_metadata(...)`, `set_scheduling(...)`, `set_tags(...)`, `start(...)` and `stop(...)`
// to build up your call.
let rb = hub.instances();

Methods

impl<'a, C, NC, A> InstanceMethodsBuilder<'a, C, NC, A>

fn set_scheduling(&self, request: &Scheduling, project: &str, zone: &str, instance: &str) -> InstanceSetSchedulingMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Sets an instance's scheduling options.

fn delete(&self, project: &str, zone: &str, instance: &str) -> InstanceDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes the specified Instance resource. For more information, see Shutting down an instance.

fn get_serial_port_output(&self, project: &str, zone: &str, instance: &str) -> InstanceGetSerialPortOutputMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns the specified instance's serial port output.

fn set_disk_auto_delete(&self, project: &str, zone: &str, instance: &str, auto_delete: bool, device_name: &str) -> InstanceSetDiskAutoDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Sets the auto-delete flag for a disk attached to an instance.

fn add_access_config(&self, request: &AccessConfig, project: &str, zone: &str, instance: &str, network_interface: &str) -> InstanceAddAccessConfigMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Adds an access config to an instance's network interface.

fn start(&self, project: &str, zone: &str, instance: &str) -> InstanceStartMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

This method starts an instance that was stopped using the using the instances().stop method. For more information, see Restart an instance.

fn get(&self, project: &str, zone: &str, instance: &str) -> InstanceGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns the specified instance resource.

fn set_tags(&self, request: &Tags, project: &str, zone: &str, instance: &str) -> InstanceSetTagMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Sets tags for the specified instance to the data included in the request.

fn set_metadata(&self, request: &Metadata, project: &str, zone: &str, instance: &str) -> InstanceSetMetadataMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Sets metadata for the specified instance to the data included in the request.

fn detach_disk(&self, project: &str, zone: &str, instance: &str, device_name: &str) -> InstanceDetachDiskMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Detaches a disk from an instance.

fn stop(&self, project: &str, zone: &str, instance: &str) -> InstanceStopMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

This method stops a running instance, shutting it down cleanly, and allows you to restart the instance at a later time. Stopped instances do not incur per-minute, virtual machine usage charges while they are stopped, but any resources that the virtual machine is using, such as persistent disks and static IP addresses,will continue to be charged until they are deleted. For more information, see Stopping an instance.

fn insert(&self, request: &Instance, project: &str, zone: &str) -> InstanceInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates an instance resource in the specified project using the data included in the request.

fn reset(&self, project: &str, zone: &str, instance: &str) -> InstanceResetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Performs a hard reset on the instance.

fn delete_access_config(&self, project: &str, zone: &str, instance: &str, access_config: &str, network_interface: &str) -> InstanceDeleteAccessConfigMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes an access config from an instance's network interface.

fn attach_disk(&self, request: &AttachedDisk, project: &str, zone: &str, instance: &str) -> InstanceAttachDiskMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Attaches a Disk resource to an instance.

fn list(&self, project: &str, zone: &str) -> InstanceListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves the list of instance resources contained within the specified zone.

fn aggregated_list(&self, project: &str) -> InstanceAggregatedListMethodBuilder<'a, C, NC, A>