Struct google_compute1::InstanceProperties

pub struct InstanceProperties {
    pub can_ip_forward: Option<bool>,
    pub description: Option<String>,
    pub tags: Option<Tags>,
    pub disks: Option<Vec<AttachedDisk>>,
    pub scheduling: Option<Scheduling>,
    pub machine_type: Option<String>,
    pub service_accounts: Option<Vec<ServiceAccount>>,
    pub network_interfaces: Option<Vec<NetworkInterface>>,
    pub metadata: Option<Metadata>,
}

This type is not used in any activity, and only used as part of another schema.

Fields

can_ip_forward

Allows instances created based on this template to send packets with source IP addresses other than their own and receive packets with destination IP addresses other than their own. If these instances will be used as an IP gateway or it will be set as the next-hop in a Route resource, say true. If unsure, leave this set to false.

description

An optional textual description for the instances created based on the instance template resource; provided by the client when the template is created.

tags

A list of tags to be applied to the instances created based on this template used to identify valid sources or targets for network firewalls. Provided by the client on instance creation. The tags can be later modified by the setTags method. Each tag within the list must comply with RFC1035.

disks

Array of disks associated with instance created based on this template.

scheduling

Scheduling options for the instances created based on this template.

machine_type

Name of the machine type resource describing which machine type to use to host the instances created based on this template; provided by the client when the instance template is created.

service_accounts

A list of service accounts each with specified scopes, for which access tokens are to be made available to the instances created based on this template, through metadata queries.

network_interfaces

Array of configurations for this interface. This specifies how this interface is configured to interact with other network services, such as connecting to the internet. Currently, ONE_TO_ONE_NAT is the only access config supported. If there are no accessConfigs specified, then this instances created based based on this template will have no external internet access.

metadata

Metadata key/value pairs assigned to instances created based on this template. Consists of custom metadata or predefined keys; see Instance documentation for more information.