Struct google_compute1::OperationWarnings

pub struct OperationWarnings {
    pub message: Option<String>,
    pub code: Option<String>,
    pub data: Option<Vec<OperationWarningsData>>,
}

[Output Only] If warning messages are generated during processing of the operation, this field will be populated.

This type is not used in any activity, and only used as part of another schema.

Fields

message

[Output Only] A human-readable description of the warning code.

code

[Output Only] A warning code, if applicable. For example, Compute Engine returns NO_RESULTS_ON_PAGE if there are no results in the response.

data

[Output Only] Metadata about this warning in key: value format. For example: "data": [ { "key": "scope", "value": "zones/us-east1-d" }