Struct google_compute1::Route

pub struct Route {
    pub next_hop_gateway: Option<String>,
    pub description: Option<String>,
    pub tags: Option<Vec<String>>,
    pub next_hop_network: Option<String>,
    pub dest_range: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub next_hop_ip: Option<String>,
    pub network: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub next_hop_instance: Option<String>,
    pub warnings: Option<Vec<RouteWarnings>>,
    pub priority: Option<u32>,
    pub next_hop_vpn_tunnel: Option<String>,
    pub self_link: Option<String>,
}

The route resource. A Route is a rule that specifies how certain packets should be handled by the virtual network. Routes are associated with VMs by tag and the set of Routes for a particular VM is called its routing table. For each packet leaving a VM, the system searches that VM's routing table for a single best matching Route. Routes match packets by destination IP address, preferring smaller or more specific ranges over larger ones. If there is a tie, the system selects the Route with the smallest priority value. If there is still a tie, it uses the layer three and four packet headers to select just one of the remaining matching Routes. The packet is then forwarded as specified by the next_hop field of the winning Route -- either to another VM destination, a VM gateway or a GCE operated gateway. Packets that do not match any Route in the sending VM's routing table will be dropped.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list routes (none)

insert routes (request)

delete routes (none)

get routes (response)

Fields

next_hop_gateway

The URL to a gateway that should handle matching packets.

description

An optional textual description of the resource; provided by the client when the resource is created.

tags

A list of instance tags to which this route applies.

next_hop_network

The URL of the local network if it should handle matching packets.

dest_range

Which packets does this route apply to?

creation_timestamp

Creation timestamp in RFC3339 text format (output only).

id

Unique identifier for the resource; defined by the server (output only).

next_hop_ip

The network IP address of an instance that should handle matching packets.

network

URL of the network to which this route is applied; provided by the client when the route is created.

kind

Type of the resource.

name

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035.

next_hop_instance

The URL to an instance that should handle matching packets.

warnings

If potential misconfigurations are detected for this route, this field will be populated with warning messages.

priority

Breaks ties between Routes of equal specificity. Routes with smaller values win when tied with routes with larger values.

next_hop_vpn_tunnel

The URL to a VpnTunnel that should handle matching packets.

self_link

Server defined URL for the resource (output only).