Struct google_compute1::Snapshot

pub struct Snapshot {
    pub status: Option<String>,
    pub kind: Option<String>,
    pub storage_bytes: Option<String>,
    pub description: Option<String>,
    pub source_disk: Option<String>,
    pub creation_timestamp: Option<String>,
    pub source_disk_id: Option<String>,
    pub disk_size_gb: Option<String>,
    pub licenses: Option<Vec<String>>,
    pub storage_bytes_status: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

A persistent disk snapshot resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list snapshots (none)

create snapshot disks (request)

delete snapshots (none)

get snapshots (response)

Fields

status

The status of the persistent disk snapshot (output only).

kind

Type of the resource.

storage_bytes

A size of the the storage used by the snapshot. As snapshots share storage this number is expected to change with snapshot creation/deletion.

description

An optional textual description of the resource; provided by the client when the resource is created.

source_disk

The source disk used to create this snapshot.

creation_timestamp

Creation timestamp in RFC3339 text format (output only).

source_disk_id

The 'id' value of the disk used to create this snapshot. This value may be used to determine whether the snapshot was taken from the current or a previous instance of a given disk name.

disk_size_gb

Size of the persistent disk snapshot, specified in GB (output only).

licenses

Public visible licenses.

storage_bytes_status

An indicator whether storageBytes is in a stable state, or it is being adjusted as a result of shared storage reallocation.

id

Unique identifier for the resource; defined by the server (output only).

self_link

Server defined URL for the resource (output only).

name

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035.