Struct google_compute1::TargetHttpsProxy

pub struct TargetHttpsProxy {
    pub kind: Option<String>,
    pub description: Option<String>,
    pub ssl_certificates: Option<Vec<String>>,
    pub ssl_policy: Option<String>,
    pub quic_override: Option<String>,
    pub url_map: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

A TargetHttpsProxy resource. This resource defines an HTTPS proxy. (== resource_for beta.targetHttpsProxies ==) (== resource_for v1.targetHttpsProxies ==)

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get target https proxies (response)

insert target https proxies (request)

Fields
kind: Option<String>

[Output Only] Type of resource. Always compute#targetHttpsProxy for target HTTPS proxies.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

ssl_certificates: Option<Vec<String>>

URLs to SslCertificate resources that are used to authenticate connections between users and the load balancer. At least one SSL certificate must be specified. Currently, you may specify up to 15 SSL certificates.

ssl_policy: Option<String>

URL of SslPolicy resource that will be associated with the TargetHttpsProxy resource. If not set, the TargetHttpsProxy resource will not have any SSL policy configured.

quic_override: Option<String>

Specifies the QUIC override policy for this TargetHttpsProxy resource. This determines whether the load balancer will attempt to negotiate QUIC with clients or not. Can specify one of NONE, ENABLE, or DISABLE. Specify ENABLE to always enable QUIC, Enables QUIC when set to ENABLE, and disables QUIC when set to DISABLE. If NONE is specified, uses the QUIC policy with no user overrides, which is equivalent to DISABLE. Not specifying this field is equivalent to specifying NONE.

url_map: Option<String>

A fully-qualified or valid partial URL to the UrlMap resource that defines the mapping from URL to the BackendService. For example, the following are all valid URLs for specifying a URL map:

https://www.googleapis.compute/v1/projects/project/global/urlMaps/url-map

projects/project/global/urlMaps/url-map

global/urlMaps/url-map

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

id: Option<String>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z]([-a-z0-9]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.