Struct google_compute1::TargetHttpsProxyListCall

pub struct TargetHttpsProxyListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Retrieves the list of TargetHttpsProxy resources available to the specified project.

A builder for the list method supported by a targetHttpsProxy resource. It is not used directly, but through a TargetHttpsProxyMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.target_https_proxies().list("project")
             .page_token("diam")
             .max_results(79)
             .filter("dolores")
             .doit();

Methods

impl<'a, C, A> TargetHttpsProxyListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, TargetHttpsProxyList)>

Perform the operation you have build so far.

fn project(self, new_value: &str) -> TargetHttpsProxyListCall<'a, C, A>

Project ID for this request.

Sets the project path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn page_token(self, new_value: &str) -> TargetHttpsProxyListCall<'a, C, A>

Specifies a page token to use. Set pageToken to the nextPageToken returned by a previous list request to get the next page of results.

Sets the page token query property to the given value.

fn max_results(self, new_value: u32) -> TargetHttpsProxyListCall<'a, C, A>

The maximum number of results per page that should be returned. If the number of available results is larger than maxResults, Compute Engine returns a nextPageToken that can be used to get the next page of results in subsequent list requests.

Sets the max results query property to the given value.

fn filter(self, new_value: &str) -> TargetHttpsProxyListCall<'a, C, A>

Sets a filter expression for filtering listed resources, in the form filter={expression}. Your {expression} must be in the format: field_name comparison_string literal_string.

The field_name is the name of the field you want to compare. Only atomic field types are supported (string, number, boolean). The comparison_string must be either eq (equals) or ne (not equals). The literal_string is the string value to filter to. The literal value must be valid for the type of field you are filtering by (string, number, boolean). For string fields, the literal value is interpreted as a regular expression using RE2 syntax. The literal value must match the entire field.

For example, to filter for instances that do not have a name of example-instance, you would use filter=name ne example-instance.

Compute Engine Beta API Only: If you use filtering in the Beta API, you can also filter on nested fields. For example, you could filter on instances that have set the scheduling.automaticRestart field to true. In particular, use filtering on nested fields to take advantage of instance labels to organize and filter results based on label values.

The Beta API also supports filtering on multiple expressions by providing each separate expression within parentheses. For example, (scheduling.automaticRestart eq true) (zone eq us-central1-f). Multiple expressions are treated as AND expressions, meaning that resources must match all expressions to pass the filters.

Sets the filter query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> TargetHttpsProxyListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> TargetHttpsProxyListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> TargetHttpsProxyListCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Readonly.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.