Struct google_compute1::TargetPoolMethods

pub struct TargetPoolMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on targetPool resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_health_check(...)`, `add_instance(...)`, `aggregated_list(...)`, `delete(...)`, `get(...)`, `get_health(...)`, `insert(...)`, `list(...)`, `remove_health_check(...)`, `remove_instance(...)` and `set_backup(...)`
// to build up your call.
let rb = hub.target_pools();

Methods

impl<'a, C, A> TargetPoolMethods<'a, C, A>

fn list(&self, project: &str, region: &str) -> TargetPoolListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the list of TargetPool resources available to the specified project and region.

Arguments

project - Name of the project scoping this request.

region - Name of the region scoping this request.

fn add_health_check(&self, request: TargetPoolsAddHealthCheckRequest, project: &str, region: &str, target_pool: &str) -> TargetPoolAddHealthCheckCall<'a, C, A>

Create a builder to help you perform the following task:

Adds health check URL to targetPool.

Arguments

request - No description provided.

project - No description provided.

region - Name of the region scoping this request.

targetPool - Name of the TargetPool resource to which health_check_url is to be added.

fn insert(&self, request: TargetPool, project: &str, region: &str) -> TargetPoolInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a TargetPool resource in the specified project and region using the data included in the request.

Arguments

request - No description provided.

project - Name of the project scoping this request.

region - Name of the region scoping this request.

fn remove_health_check(&self, request: TargetPoolsRemoveHealthCheckRequest, project: &str, region: &str, target_pool: &str) -> TargetPoolRemoveHealthCheckCall<'a, C, A>

Create a builder to help you perform the following task:

Removes health check URL from targetPool.

Arguments

request - No description provided.

project - No description provided.

region - Name of the region scoping this request.

targetPool - Name of the TargetPool resource to which health_check_url is to be removed.

fn get_health(&self, request: InstanceReference, project: &str, region: &str, target_pool: &str) -> TargetPoolGetHealthCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the most recent health check results for each IP for the given instance that is referenced by given TargetPool.

Arguments

request - No description provided.

project - No description provided.

region - Name of the region scoping this request.

targetPool - Name of the TargetPool resource to which the queried instance belongs.

fn set_backup(&self, request: TargetReference, project: &str, region: &str, target_pool: &str) -> TargetPoolSetBackupCall<'a, C, A>

Create a builder to help you perform the following task:

Changes backup pool configurations.

Arguments

request - No description provided.

project - Name of the project scoping this request.

region - Name of the region scoping this request.

targetPool - Name of the TargetPool resource for which the backup is to be set.

fn aggregated_list(&self, project: &str) -> TargetPoolAggregatedListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the list of target pools grouped by scope.

Arguments

project - Name of the project scoping this request.

fn get(&self, project: &str, region: &str, target_pool: &str) -> TargetPoolGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the specified TargetPool resource.

Arguments

project - Name of the project scoping this request.

region - Name of the region scoping this request.

targetPool - Name of the TargetPool resource to return.

fn add_instance(&self, request: TargetPoolsAddInstanceRequest, project: &str, region: &str, target_pool: &str) -> TargetPoolAddInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Adds instance url to targetPool.

Arguments

request - No description provided.

project - No description provided.

region - Name of the region scoping this request.

targetPool - Name of the TargetPool resource to which instance_url is to be added.

fn remove_instance(&self, request: TargetPoolsRemoveInstanceRequest, project: &str, region: &str, target_pool: &str) -> TargetPoolRemoveInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Removes instance URL from targetPool.

Arguments

request - No description provided.

project - No description provided.

region - Name of the region scoping this request.

targetPool - Name of the TargetPool resource to which instance_url is to be removed.

fn delete(&self, project: &str, region: &str, target_pool: &str) -> TargetPoolDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified TargetPool resource.

Arguments

project - Name of the project scoping this request.

region - Name of the region scoping this request.

targetPool - Name of the TargetPool resource to delete.