Struct google-compute1::TargetPoolMethodsBuilder

pub struct TargetPoolMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on targetPool resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-compute1" as compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `add_health_check(...)`, `add_instance(...)`, `aggregated_list(...)`, `delete(...)`, `get(...)`, `get_health(...)`, `insert(...)`, `list(...)`, `remove_health_check(...)`, `remove_instance(...)` and `set_backup(...)`
// to build up your call.
let rb = hub.target_pools();

Methods

impl<'a, C, NC, A> TargetPoolMethodsBuilder<'a, C, NC, A>

fn list(&self, project: &str, region: &str) -> TargetPoolListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves the list of TargetPool resources available to the specified project and region.

fn add_health_check(&self, request: &TargetPoolsAddHealthCheckRequest, project: &str, region: &str, target_pool: &str) -> TargetPoolAddHealthCheckMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Adds health check URL to targetPool.

fn insert(&self, request: &TargetPool, project: &str, region: &str) -> TargetPoolInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a TargetPool resource in the specified project and region using the data included in the request.

fn remove_health_check(&self, request: &TargetPoolsRemoveHealthCheckRequest, project: &str, region: &str, target_pool: &str) -> TargetPoolRemoveHealthCheckMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Removes health check URL from targetPool.

fn get_health(&self, request: &InstanceReference, project: &str, region: &str, target_pool: &str) -> TargetPoolGetHealthMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets the most recent health check results for each IP for the given instance that is referenced by given TargetPool.

fn set_backup(&self, request: &TargetReference, project: &str, region: &str, target_pool: &str) -> TargetPoolSetBackupMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Changes backup pool configurations.

fn aggregated_list(&self, project: &str) -> TargetPoolAggregatedListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves the list of target pools grouped by scope.

fn get(&self, project: &str, region: &str, target_pool: &str) -> TargetPoolGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns the specified TargetPool resource.

fn add_instance(&self, request: &TargetPoolsAddInstanceRequest, project: &str, region: &str, target_pool: &str) -> TargetPoolAddInstanceMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Adds instance url to targetPool.

fn remove_instance(&self, request: &TargetPoolsRemoveInstanceRequest, project: &str, region: &str, target_pool: &str) -> TargetPoolRemoveInstanceMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Removes instance URL from targetPool.

fn delete(&self, project: &str, region: &str, target_pool: &str) -> TargetPoolDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes the specified TargetPool resource.