Struct google-compute1::UrlMap

pub struct UrlMap {
    pub kind: Option<String>,
    pub description: Option<String>,
    pub default_service: Option<String>,
    pub tests: Vec<UrlMapTest>,
    pub host_rules: Vec<HostRule>,
    pub fingerprint: Option<String>,
    pub path_matchers: Vec<PathMatcher>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

A UrlMap resource. This resource defines the mapping from URL to the BackendService resource, based on the "longest-match" of the URL's host and path.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update url maps (request)

insert url maps (request)

list url maps (none)

validate url maps (none)

delete url maps (none)

get url maps (response)

patch url maps (request)

Fields

kind

Type of the resource.

description

An optional textual description of the resource; provided by the client when the resource is created.

default_service

The URL of the BackendService resource if none of the hostRules match.

tests

The list of expected URL mappings. Request to update this UrlMap will succeed only all of the test cases pass.

host_rules

The list of HostRules to use against the URL.

fingerprint

Fingerprint of this resource. A hash of the contents stored in this object. This field is used in optimistic locking. This field will be ignored when inserting a UrlMap. An up-to-date fingerprint must be provided in order to update the UrlMap.

path_matchers

The list of named PathMatchers to use against the URL.

creation_timestamp

Creation timestamp in RFC3339 text format (output only).

id

Unique identifier for the resource; defined by the server (output only).

self_link

Server defined URL for the resource (output only).

name

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035.