Struct google_compute1::UrlMapValidationResult

pub struct UrlMapValidationResult {
    pub test_failures: Option<Vec<TestFailure>>,
    pub load_errors: Option<Vec<String>>,
    pub load_succeeded: Option<bool>,
    pub test_passed: Option<bool>,
}

Message representing the validation result for a UrlMap.

This type is not used in any activity, and only used as part of another schema.

Fields

test_failures

no description provided

load_errors

no description provided

load_succeeded

Whether the given UrlMap can be successfully loaded. If false, 'loadErrors' indicates the reasons.

test_passed

If successfully loaded, this field indicates whether the test passed. If false, 'testFailures's indicate the reason of failure.