Struct google_compute1::VpnTunnel

pub struct VpnTunnel {
    pub detailed_status: Option<String>,
    pub ike_version: Option<i32>,
    pub kind: Option<String>,
    pub description: Option<String>,
    pub peer_ip: Option<String>,
    pub region: Option<String>,
    pub local_traffic_selector: Option<Vec<String>>,
    pub name: Option<String>,
    pub status: Option<String>,
    pub shared_secret_hash: Option<String>,
    pub target_vpn_gateway: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub shared_secret: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete vpn tunnels (none)

aggregated list vpn tunnels (none)

get vpn tunnels (response)

insert vpn tunnels (request)

list vpn tunnels (none)

Fields

detailed_status

[Output Only] Detailed status message for the VPN tunnel.

ike_version

IKE protocol version to use when establishing the VPN tunnel with peer VPN gateway. Acceptable IKE versions are 1 or 2. Default version is 2.

kind

[Output Only] Type of resource. Always compute#vpnTunnel for VPN tunnels.

description

An optional description of this resource. Provide this property when you create the resource.

peer_ip

IP address of the peer VPN gateway.

region

[Output Only] URL of the region where the VPN tunnel resides.

local_traffic_selector

Local traffic selector to use when establishing the VPN tunnel with peer VPN gateway. The value should be a CIDR formatted string, for example: 192.168.0.0/16. The ranges should be disjoint.

name

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

status

[Output Only] The status of the VPN tunnel.

shared_secret_hash

Hash of the shared secret.

target_vpn_gateway

URL of the VPN gateway with which this VPN tunnel is associated. Provided by the client when the VPN tunnel is created.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

self_link

[Output Only] Server-defined URL for the resource.

shared_secret

Shared secret used to set the secure session between the Cloud VPN gateway and the peer VPN gateway.