Struct google_compute1::VpnTunnelList

pub struct VpnTunnelList {
    pub next_page_token: Option<String>,
    pub items: Option<Vec<VpnTunnel>>,
    pub kind: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
}

Contains a list of VpnTunnel resources.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list vpn tunnels (response)

Fields

next_page_token

[Output Only] This token allows you to get the next page of results for list requests. If the number of results is larger than maxResults, use the nextPageToken as a value for the query parameter pageToken in the next list request. Subsequent list requests will have their own nextPageToken to continue paging through the results.

items

[Output Only] A list of VpnTunnel resources.

kind

[Output Only] Type of resource. Always compute#vpnTunnel for VPN tunnels.

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

self_link

[Output Only] Server-defined URL for the resource.