Crate google_compute1

This documentation was generated from compute crate version 0.1.8+20150326, where 20150326 is the exact revision of the compute:v1 schema built by the mako code generator v0.1.8.

Everything else about the compute v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

addresses

aggregated list, delete, get, insert and list

backend services

delete, get, get health, insert, list, patch and update

disk types

aggregated list, get and list

disks

aggregated list, create snapshot, delete, get, insert and list

firewalls

delete, get, insert, list, patch and update

forwarding rules

aggregated list, delete, get, insert, list and set target

global addresses

delete, get, insert and list

global forwarding rules

delete, get, insert, list and set target

global operations

aggregated list, delete, get and list

http health checks

delete, get, insert, list, patch and update

images

delete, deprecate, get, insert and list

instance templates

delete, get, insert and list

instances

add access config, aggregated list, attach disk, delete, delete access config, detach disk, get, get serial port output, insert, list, reset, set disk auto delete, set metadata, set scheduling, set tags, start and stop

licenses

get

machine types

aggregated list, get and list

networks

delete, get, insert and list

projects

get, move disk, move instance, set common instance metadata and set usage export bucket

region operations

delete, get and list

regions

get and list

routes

delete, get, insert and list

snapshots

delete, get and list

target http proxies

delete, get, insert, list and set url map

target instances

aggregated list, delete, get, insert and list

target pools

add health check, add instance, aggregated list, delete, get, get health, insert, list, remove health check, remove instance and set backup

target vpn gateways

aggregated list, delete, get, insert and list

url maps

delete, get, insert, list, patch, update and validate

vpn tunnels

aggregated list, delete, get, insert and list

zone operations

delete, get and list

zones

get and list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.target_instances().delete(...).doit()
let r = hub.instances().stop(...).doit()
let r = hub.target_pools().add_instance(...).doit()
let r = hub.target_pools().add_health_check(...).doit()
let r = hub.target_pools().remove_health_check(...).doit()
let r = hub.backend_services().update(...).doit()
let r = hub.target_pools().delete(...).doit()
let r = hub.firewalls().update(...).doit()
let r = hub.networks().insert(...).doit()
let r = hub.addresses().insert(...).doit()
let r = hub.backend_services().patch(...).doit()
let r = hub.instances().set_tags(...).doit()
let r = hub.instances().set_disk_auto_delete(...).doit()
let r = hub.target_http_proxies().insert(...).doit()
let r = hub.projects().set_usage_export_bucket(...).doit()
let r = hub.url_maps().update(...).doit()
let r = hub.instance_templates().insert(...).doit()
let r = hub.firewalls().patch(...).doit()
let r = hub.vpn_tunnels().delete(...).doit()
let r = hub.firewalls().delete(...).doit()
let r = hub.instances().detach_disk(...).doit()
let r = hub.instances().delete_access_config(...).doit()
let r = hub.backend_services().delete(...).doit()
let r = hub.instances().insert(...).doit()
let r = hub.global_forwarding_rules().set_target(...).doit()
let r = hub.url_maps().insert(...).doit()
let r = hub.instances().add_access_config(...).doit()
let r = hub.target_vpn_gateways().delete(...).doit()
let r = hub.region_operations().get(...).doit()
let r = hub.backend_services().insert(...).doit()
let r = hub.global_operations().get(...).doit()
let r = hub.disks().insert(...).doit()
let r = hub.instance_templates().delete(...).doit()
let r = hub.global_forwarding_rules().delete(...).doit()
let r = hub.global_forwarding_rules().insert(...).doit()
let r = hub.disks().delete(...).doit()
let r = hub.global_addresses().delete(...).doit()
let r = hub.projects().set_common_instance_metadata(...).doit()
let r = hub.target_pools().insert(...).doit()
let r = hub.target_http_proxies().delete(...).doit()
let r = hub.networks().delete(...).doit()
let r = hub.target_vpn_gateways().insert(...).doit()
let r = hub.url_maps().delete(...).doit()
let r = hub.instances().set_scheduling(...).doit()
let r = hub.instances().attach_disk(...).doit()
let r = hub.addresses().delete(...).doit()
let r = hub.instances().reset(...).doit()
let r = hub.target_http_proxies().set_url_map(...).doit()
let r = hub.disks().create_snapshot(...).doit()
let r = hub.routes().insert(...).doit()
let r = hub.global_addresses().insert(...).doit()
let r = hub.images().deprecate(...).doit()
let r = hub.http_health_checks().patch(...).doit()
let r = hub.http_health_checks().delete(...).doit()
let r = hub.instances().set_metadata(...).doit()
let r = hub.http_health_checks().insert(...).doit()
let r = hub.images().delete(...).doit()
let r = hub.snapshots().delete(...).doit()
let r = hub.forwarding_rules().insert(...).doit()
let r = hub.images().insert(...).doit()
let r = hub.instances().start(...).doit()
let r = hub.target_pools().set_backup(...).doit()
let r = hub.target_instances().insert(...).doit()
let r = hub.forwarding_rules().delete(...).doit()
let r = hub.projects().move_instance(...).doit()
let r = hub.forwarding_rules().set_target(...).doit()
let r = hub.projects().move_disk(...).doit()
let r = hub.vpn_tunnels().insert(...).doit()
let r = hub.firewalls().insert(...).doit()
let r = hub.target_pools().remove_instance(...).doit()
let r = hub.instances().delete(...).doit()
let r = hub.zone_operations().get(...).doit()
let r = hub.http_health_checks().update(...).doit()
let r = hub.routes().delete(...).doit()
let r = hub.url_maps().patch(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-compute1 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
use compute1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.instances().set_disk_auto_delete("project", "zone", "instance", true, "deviceName")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AccessConfig
An access configuration attached to an instance's network interface.

Address
A reserved address resource.

AddressAggregatedList
There is no detailed description.

AddressList
Contains a list of address resources.

AddresseAggregatedListCall
Retrieves the list of addresses grouped by scope.

AddresseDeleteCall
Deletes the specified address resource.

AddresseGetCall
Returns the specified address resource.

AddresseInsertCall
Creates an address resource in the specified project using the data included in the request.

AddresseListCall
Retrieves the list of address resources contained within the specified region.

AddresseMethods
A builder providing access to all methods supported on addresse resources. It is not used directly, but through the Compute hub.

AddressesScopedList
There is no detailed description.

AddressesScopedListWarning
[Output Only] Informational warning which replaces the list of addresses when the list is empty.

AddressesScopedListWarningData
[Output Only] Metadata for this warning in key: value format.

AttachedDisk
An instance-attached disk resource.

AttachedDiskInitializeParams
[Input Only] Specifies the parameters for a new disk that will be created alongside the new instance. Use initialization parameters to create boot disks or local SSDs attached to the new instance.

Backend
Message containing information of one individual backend.

BackendService
A BackendService resource. This resource defines a group of backend VMs together with their serving capacity.

BackendServiceDeleteCall
Deletes the specified BackendService resource.

BackendServiceGetCall
Returns the specified BackendService resource.

BackendServiceGetHealthCall
Gets the most recent health check results for this BackendService.

BackendServiceGroupHealth
There is no detailed description.

BackendServiceInsertCall
Creates a BackendService resource in the specified project using the data included in the request.

BackendServiceList
Contains a list of BackendService resources.

BackendServiceListCall
Retrieves the list of BackendService resources available to the specified project.

BackendServiceMethods
A builder providing access to all methods supported on backendService resources. It is not used directly, but through the Compute hub.

BackendServicePatchCall
Update the entire content of the BackendService resource. This method supports patch semantics.

BackendServiceUpdateCall
Update the entire content of the BackendService resource.

Compute
Central instance to access all Compute related resource activities

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DeprecationStatus
Deprecation status for a public resource.

Disk
A Disk resource.

DiskAggregatedList
There is no detailed description.

DiskAggregatedListCall
Retrieves the list of disks grouped by scope.

DiskCreateSnapshotCall
Creates a snapshot of this disk.

DiskDeleteCall
Deletes the specified persistent disk.

DiskGetCall
Returns a specified persistent disk.

DiskInsertCall
Creates a persistent disk in the specified project using the data included in the request.

DiskList
A list of Disk resources.

DiskListCall
Retrieves the list of persistent disks contained within the specified zone.

DiskMethods
A builder providing access to all methods supported on disk resources. It is not used directly, but through the Compute hub.

DiskMoveRequest
There is no detailed description.

DiskType
A disk type resource.

DiskTypeAggregatedList
There is no detailed description.

DiskTypeAggregatedListCall
Retrieves the list of disk type resources grouped by scope.

DiskTypeGetCall
Returns the specified disk type resource.

DiskTypeList
Contains a list of disk type resources.

DiskTypeListCall
Retrieves the list of disk type resources available to the specified project.

DiskTypeMethods
A builder providing access to all methods supported on diskType resources. It is not used directly, but through the Compute hub.

DiskTypesScopedList
There is no detailed description.

DiskTypesScopedListWarning
[Output Only] Informational warning which replaces the list of disk types when the list is empty.

DiskTypesScopedListWarningData
[Output Only] Metadata for this warning in key: value format.

DisksScopedList
There is no detailed description.

DisksScopedListWarning
[Output Only] Informational warning which replaces the list of disks when the list is empty.

DisksScopedListWarningData
[Output Only] Metadata for this warning in key: value format.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

Firewall
A Firewall resource.

FirewallAllowed
The list of rules specified by this firewall. Each rule specifies a protocol and port-range tuple that describes a permitted connection.

FirewallDeleteCall
Deletes the specified firewall resource.

FirewallGetCall
Returns the specified firewall resource.

FirewallInsertCall
Creates a firewall resource in the specified project using the data included in the request.

FirewallList
Contains a list of Firewall resources.

FirewallListCall
Retrieves the list of firewall resources available to the specified project.

FirewallMethods
A builder providing access to all methods supported on firewall resources. It is not used directly, but through the Compute hub.

FirewallPatchCall
Updates the specified firewall resource with the data included in the request. This method supports patch semantics.

FirewallUpdateCall
Updates the specified firewall resource with the data included in the request.

ForwardingRule
A ForwardingRule resource. A ForwardingRule resource specifies which pool of target VMs to forward a packet to if it matches the given [IPAddress, IPProtocol, portRange] tuple.

ForwardingRuleAggregatedList
There is no detailed description.

ForwardingRuleAggregatedListCall
Retrieves the list of forwarding rules grouped by scope.

ForwardingRuleDeleteCall
Deletes the specified ForwardingRule resource.

ForwardingRuleGetCall
Returns the specified ForwardingRule resource.

ForwardingRuleInsertCall
Creates a ForwardingRule resource in the specified project and region using the data included in the request.

ForwardingRuleList
Contains a list of ForwardingRule resources.

ForwardingRuleListCall
Retrieves the list of ForwardingRule resources available to the specified project and region.

ForwardingRuleMethods
A builder providing access to all methods supported on forwardingRule resources. It is not used directly, but through the Compute hub.

ForwardingRuleSetTargetCall
Changes target url for forwarding rule.

ForwardingRulesScopedList
There is no detailed description.

ForwardingRulesScopedListWarning
Informational warning which replaces the list of forwarding rules when the list is empty.

ForwardingRulesScopedListWarningData
[Output Only] Metadata for this warning in key: value format.

GlobalAddresseDeleteCall
Deletes the specified address resource.

GlobalAddresseGetCall
Returns the specified address resource.

GlobalAddresseInsertCall
Creates an address resource in the specified project using the data included in the request.

GlobalAddresseListCall
Retrieves the list of global address resources.

GlobalAddresseMethods
A builder providing access to all methods supported on globalAddresse resources. It is not used directly, but through the Compute hub.

GlobalForwardingRuleDeleteCall
Deletes the specified ForwardingRule resource.

GlobalForwardingRuleGetCall
Returns the specified ForwardingRule resource.

GlobalForwardingRuleInsertCall
Creates a ForwardingRule resource in the specified project and region using the data included in the request.

GlobalForwardingRuleListCall
Retrieves the list of ForwardingRule resources available to the specified project.

GlobalForwardingRuleMethods
A builder providing access to all methods supported on globalForwardingRule resources. It is not used directly, but through the Compute hub.

GlobalForwardingRuleSetTargetCall
Changes target url for forwarding rule.

GlobalOperationAggregatedListCall
Retrieves the list of all operations grouped by scope.

GlobalOperationDeleteCall
Deletes the specified operation resource.

GlobalOperationGetCall
Retrieves the specified operation resource.

GlobalOperationListCall
Retrieves the list of operation resources contained within the specified project.

GlobalOperationMethods
A builder providing access to all methods supported on globalOperation resources. It is not used directly, but through the Compute hub.

HealthCheckReference
There is no detailed description.

HealthStatus
There is no detailed description.

HostRule
A host-matching rule for a URL. If matched, will use the named PathMatcher to select the BackendService.

HttpHealthCheck
An HttpHealthCheck resource. This resource defines a template for how individual VMs should be checked for health, via HTTP.

HttpHealthCheckDeleteCall
Deletes the specified HttpHealthCheck resource.

HttpHealthCheckGetCall
Returns the specified HttpHealthCheck resource.

HttpHealthCheckInsertCall
Creates a HttpHealthCheck resource in the specified project using the data included in the request.

HttpHealthCheckList
Contains a list of HttpHealthCheck resources.

HttpHealthCheckListCall
Retrieves the list of HttpHealthCheck resources available to the specified project.

HttpHealthCheckMethods
A builder providing access to all methods supported on httpHealthCheck resources. It is not used directly, but through the Compute hub.

HttpHealthCheckPatchCall
Updates a HttpHealthCheck resource in the specified project using the data included in the request. This method supports patch semantics.

HttpHealthCheckUpdateCall
Updates a HttpHealthCheck resource in the specified project using the data included in the request.

Image
An Image resource.

ImageDeleteCall
Deletes the specified image resource.

ImageDeprecateCall
Sets the deprecation status of an image.

ImageGetCall
Returns the specified image resource.

ImageInsertCall
Creates an image resource in the specified project using the data included in the request.

ImageList
Contains a list of Image resources.

ImageListCall
Retrieves the list of image resources available to the specified project.

ImageMethods
A builder providing access to all methods supported on image resources. It is not used directly, but through the Compute hub.

ImageRawDisk
The parameters of the raw disk image.

Instance
An Instance resource.

InstanceAddAccessConfigCall
Adds an access config to an instance's network interface.

InstanceAggregatedList
There is no detailed description.

InstanceAggregatedListCall
A builder for the aggregatedList method supported by a instance resource. It is not used directly, but through a InstanceMethods instance.

InstanceAttachDiskCall
Attaches a Disk resource to an instance.

InstanceDeleteAccessConfigCall
Deletes an access config from an instance's network interface.

InstanceDeleteCall
Deletes the specified Instance resource. For more information, see Shutting down an instance.

InstanceDetachDiskCall
Detaches a disk from an instance.

InstanceGetCall
Returns the specified instance resource.

InstanceGetSerialPortOutputCall
Returns the specified instance's serial port output.

InstanceInsertCall
Creates an instance resource in the specified project using the data included in the request.

InstanceList
Contains a list of instance resources.

InstanceListCall
Retrieves the list of instance resources contained within the specified zone.

InstanceMethods
A builder providing access to all methods supported on instance resources. It is not used directly, but through the Compute hub.

InstanceMoveRequest
There is no detailed description.

InstanceProperties

InstanceReference
There is no detailed description.

InstanceResetCall
Performs a hard reset on the instance.

InstanceSetDiskAutoDeleteCall
Sets the auto-delete flag for a disk attached to an instance.

InstanceSetMetadataCall
Sets metadata for the specified instance to the data included in the request.

InstanceSetSchedulingCall
Sets an instance's scheduling options.

InstanceSetTagCall
Sets tags for the specified instance to the data included in the request.

InstanceStartCall
This method starts an instance that was stopped using the using the instances().stop method. For more information, see Restart an instance.

InstanceStopCall
This method stops a running instance, shutting it down cleanly, and allows you to restart the instance at a later time. Stopped instances do not incur per-minute, virtual machine usage charges while they are stopped, but any resources that the virtual machine is using, such as persistent disks and static IP addresses,will continue to be charged until they are deleted. For more information, see Stopping an instance.

InstanceTemplate
An Instance Template resource.

InstanceTemplateDeleteCall
Deletes the specified instance template resource.

InstanceTemplateGetCall
Returns the specified instance template resource.

InstanceTemplateInsertCall
Creates an instance template resource in the specified project using the data included in the request.

InstanceTemplateList
Contains a list of instance template resources.

InstanceTemplateListCall
Retrieves the list of instance template resources contained within the specified project.

InstanceTemplateMethods
A builder providing access to all methods supported on instanceTemplate resources. It is not used directly, but through the Compute hub.

InstancesScopedList
There is no detailed description.

InstancesScopedListWarning
[Output Only] Informational warning which replaces the list of instances when the list is empty.

InstancesScopedListWarningData
[Output Only] Metadata for this warning in key: value format.

License
A license resource.

LicenseGetCall
Returns the specified license resource.

LicenseMethods
A builder providing access to all methods supported on license resources. It is not used directly, but through the Compute hub.

MachineType
A Machine Type resource.

MachineTypeAggregatedList
There is no detailed description.

MachineTypeAggregatedListCall
Retrieves the list of machine type resources grouped by scope.

MachineTypeGetCall
Returns the specified machine type resource.

MachineTypeList
Contains a list of Machine Type resources.

MachineTypeListCall
Retrieves the list of machine type resources available to the specified project.

MachineTypeMethods
A builder providing access to all methods supported on machineType resources. It is not used directly, but through the Compute hub.

MachineTypeScratchDisks
[Output Only] List of extended scratch disks assigned to the instance.

MachineTypesScopedList
There is no detailed description.

MachineTypesScopedListWarning
[Output Only] An informational warning that appears when the machine types list is empty.

MachineTypesScopedListWarningData
[Output Only] Metadata for this warning in key: value format.

Metadata
A metadata key/value entry.

MetadataItems
Array of key/value pairs. The total size of all keys and values must be less than 512 KB.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Network
A network resource.

NetworkDeleteCall
Deletes the specified network resource.

NetworkGetCall
Returns the specified network resource.

NetworkInsertCall
Creates a network resource in the specified project using the data included in the request.

NetworkInterface
A network interface resource attached to an instance.

NetworkList
Contains a list of Network resources.

NetworkListCall
Retrieves the list of network resources available to the specified project.

NetworkMethods
A builder providing access to all methods supported on network resources. It is not used directly, but through the Compute hub.

Operation
An operation resource, used to manage asynchronous API requests.

OperationAggregatedList
There is no detailed description.

OperationError
[Output Only] If errors are generated during processing of the operation, this field will be populated.

OperationErrorErrors
[Output Only] The array of errors encountered while processing this operation.

OperationList
Contains a list of operation resources.

OperationWarnings
[Output Only] If warning messages are generated during processing of the operation, this field will be populated.

OperationWarningsData
[Output Only] Metadata for this warning in key: value format.

OperationsScopedList
There is no detailed description.

OperationsScopedListWarning
[Output Only] Informational warning which replaces the list of operations when the list is empty.

OperationsScopedListWarningData
[Output Only] Metadata for this warning in key: value format.

PathMatcher
A matcher for the path portion of the URL. The BackendService from the longest-matched rule will serve the URL. If no rule was matched, the default_service will be used.

PathRule
A path-matching rule for a URL. If matched, will use the specified BackendService to handle the traffic arriving at this URL.

Project
A Project resource. Projects can only be created in the Google Developers Console. Unless marked otherwise, values can only be modified in the console.

ProjectGetCall
Returns the specified project resource.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Compute hub.

ProjectMoveDiskCall
Moves a persistent disk from one zone to another.

ProjectMoveInstanceCall
Moves an instance and its attached persistent disks from one zone to another.

ProjectSetCommonInstanceMetadataCall
Sets metadata common to all instances within the specified project using the data included in the request.

ProjectSetUsageExportBucketCall
Enables the usage export feature and sets the usage export bucket where reports are stored. If you provide an empty request body using this method, the usage export feature will be disabled.

Quota
A quotas entry.

Region
Region resource.

RegionGetCall
Returns the specified region resource.

RegionList
Contains a list of region resources.

RegionListCall
Retrieves the list of region resources available to the specified project.

RegionMethods
A builder providing access to all methods supported on region resources. It is not used directly, but through the Compute hub.

RegionOperationDeleteCall
Deletes the specified region-specific operation resource.

RegionOperationGetCall
Retrieves the specified region-specific operation resource.

RegionOperationListCall
Retrieves the list of operation resources contained within the specified region.

RegionOperationMethods
A builder providing access to all methods supported on regionOperation resources. It is not used directly, but through the Compute hub.

ResourceGroupReference
There is no detailed description.

Route
The route resource. A Route is a rule that specifies how certain packets should be handled by the virtual network. Routes are associated with VMs by tag and the set of Routes for a particular VM is called its routing table. For each packet leaving a VM, the system searches that VM's routing table for a single best matching Route. Routes match packets by destination IP address, preferring smaller or more specific ranges over larger ones. If there is a tie, the system selects the Route with the smallest priority value. If there is still a tie, it uses the layer three and four packet headers to select just one of the remaining matching Routes. The packet is then forwarded as specified by the next_hop field of the winning Route -- either to another VM destination, a VM gateway or a GCE operated gateway. Packets that do not match any Route in the sending VM's routing table will be dropped.

RouteDeleteCall
Deletes the specified route resource.

RouteGetCall
Returns the specified route resource.

RouteInsertCall
Creates a route resource in the specified project using the data included in the request.

RouteList
Contains a list of route resources.

RouteListCall
Retrieves the list of route resources available to the specified project.

RouteMethods
A builder providing access to all methods supported on route resources. It is not used directly, but through the Compute hub.

RouteWarnings
If potential misconfigurations are detected for this route, this field will be populated with warning messages.

RouteWarningsData
[Output Only] Metadata for this warning in key: value format.

Scheduling
Sets the scheduling options for an Instance.

SerialPortOutput
An instance's serial console output.

ServiceAccount
A service account.

Snapshot
A persistent disk snapshot resource.

SnapshotDeleteCall
Deletes the specified persistent disk snapshot resource.

SnapshotGetCall
Returns the specified persistent disk snapshot resource.

SnapshotList
Contains a list of persistent disk snapshot resources.

SnapshotListCall
Retrieves the list of persistent disk snapshot resources contained within the specified project.

SnapshotMethods
A builder providing access to all methods supported on snapshot resources. It is not used directly, but through the Compute hub.

Tags
A set of instance tags.

TargetHttpProxy
A TargetHttpProxy resource. This resource defines an HTTP proxy.

TargetHttpProxyDeleteCall
Deletes the specified TargetHttpProxy resource.

TargetHttpProxyGetCall
Returns the specified TargetHttpProxy resource.

TargetHttpProxyInsertCall
Creates a TargetHttpProxy resource in the specified project using the data included in the request.

TargetHttpProxyList
Contains a list of TargetHttpProxy resources.

TargetHttpProxyListCall
Retrieves the list of TargetHttpProxy resources available to the specified project.

TargetHttpProxyMethods
A builder providing access to all methods supported on targetHttpProxy resources. It is not used directly, but through the Compute hub.

TargetHttpProxySetUrlMapCall
Changes the URL map for TargetHttpProxy.

TargetInstance
A TargetInstance resource. This resource defines an endpoint VM that terminates traffic of certain protocols.

TargetInstanceAggregatedList
There is no detailed description.

TargetInstanceAggregatedListCall
Retrieves the list of target instances grouped by scope.

TargetInstanceDeleteCall
Deletes the specified TargetInstance resource.

TargetInstanceGetCall
Returns the specified TargetInstance resource.

TargetInstanceInsertCall
Creates a TargetInstance resource in the specified project and zone using the data included in the request.

TargetInstanceList
Contains a list of TargetInstance resources.

TargetInstanceListCall
Retrieves the list of TargetInstance resources available to the specified project and zone.

TargetInstanceMethods
A builder providing access to all methods supported on targetInstance resources. It is not used directly, but through the Compute hub.

TargetInstancesScopedList
There is no detailed description.

TargetInstancesScopedListWarning
Informational warning which replaces the list of addresses when the list is empty.

TargetInstancesScopedListWarningData
[Output Only] Metadata for this warning in key: value format.

TargetPool
A TargetPool resource. This resource defines a pool of VMs, associated HttpHealthCheck resources, and the fallback TargetPool.

TargetPoolAddHealthCheckCall
Adds health check URL to targetPool.

TargetPoolAddInstanceCall
Adds instance url to targetPool.

TargetPoolAggregatedList
There is no detailed description.

TargetPoolAggregatedListCall
Retrieves the list of target pools grouped by scope.

TargetPoolDeleteCall
Deletes the specified TargetPool resource.

TargetPoolGetCall
Returns the specified TargetPool resource.

TargetPoolGetHealthCall
Gets the most recent health check results for each IP for the given instance that is referenced by given TargetPool.

TargetPoolInsertCall
Creates a TargetPool resource in the specified project and region using the data included in the request.

TargetPoolInstanceHealth
There is no detailed description.

TargetPoolList
Contains a list of TargetPool resources.

TargetPoolListCall
Retrieves the list of TargetPool resources available to the specified project and region.

TargetPoolMethods
A builder providing access to all methods supported on targetPool resources. It is not used directly, but through the Compute hub.

TargetPoolRemoveHealthCheckCall
Removes health check URL from targetPool.

TargetPoolRemoveInstanceCall
Removes instance URL from targetPool.

TargetPoolSetBackupCall
Changes backup pool configurations.

TargetPoolsAddHealthCheckRequest
There is no detailed description.

TargetPoolsAddInstanceRequest
There is no detailed description.

TargetPoolsRemoveHealthCheckRequest
There is no detailed description.

TargetPoolsRemoveInstanceRequest
There is no detailed description.

TargetPoolsScopedList
There is no detailed description.

TargetPoolsScopedListWarning
Informational warning which replaces the list of addresses when the list is empty.

TargetPoolsScopedListWarningData
[Output Only] Metadata for this warning in key: value format.

TargetReference
There is no detailed description.

TargetVpnGateway
There is no detailed description.

TargetVpnGatewayAggregatedList
There is no detailed description.

TargetVpnGatewayAggregatedListCall
Retrieves the list of target VPN gateways grouped by scope.

TargetVpnGatewayDeleteCall
Deletes the specified TargetVpnGateway resource.

TargetVpnGatewayGetCall
Returns the specified TargetVpnGateway resource.

TargetVpnGatewayInsertCall
Creates a TargetVpnGateway resource in the specified project and region using the data included in the request.

TargetVpnGatewayList
Contains a list of TargetVpnGateway resources.

TargetVpnGatewayListCall
Retrieves the list of TargetVpnGateway resources available to the specified project and region.

TargetVpnGatewayMethods
A builder providing access to all methods supported on targetVpnGateway resources. It is not used directly, but through the Compute hub.

TargetVpnGatewaysScopedList
There is no detailed description.

TargetVpnGatewaysScopedListWarning
[Output Only] Informational warning which replaces the list of addresses when the list is empty.

TargetVpnGatewaysScopedListWarningData
[Output Only] Metadata for this warning in key: value format.

TestFailure
There is no detailed description.

UrlMap
A UrlMap resource. This resource defines the mapping from URL to the BackendService resource, based on the "longest-match" of the URL's host and path.

UrlMapDeleteCall
Deletes the specified UrlMap resource.

UrlMapGetCall
Returns the specified UrlMap resource.

UrlMapInsertCall
Creates a UrlMap resource in the specified project using the data included in the request.

UrlMapList
Contains a list of UrlMap resources.

UrlMapListCall
Retrieves the list of UrlMap resources available to the specified project.

UrlMapMethods
A builder providing access to all methods supported on urlMap resources. It is not used directly, but through the Compute hub.

UrlMapPatchCall
Update the entire content of the UrlMap resource. This method supports patch semantics.

UrlMapReference
There is no detailed description.

UrlMapTest
Message for the expected URL mappings.

UrlMapUpdateCall
Update the entire content of the UrlMap resource.

UrlMapValidateCall
Run static validation for the UrlMap. In particular, the tests of the provided UrlMap will be run. Calling this method does NOT create the UrlMap.

UrlMapValidationResult
Message representing the validation result for a UrlMap.

UrlMapsValidateRequest
There is no detailed description.

UrlMapsValidateResponse
There is no detailed description.

UsageExportLocation
The location in Cloud Storage and naming method of the daily usage report. Contains bucket_name and report_name prefix.

VpnTunnel
There is no detailed description.

VpnTunnelAggregatedList
There is no detailed description.

VpnTunnelAggregatedListCall
Retrieves the list of VPN tunnels grouped by scope.

VpnTunnelDeleteCall
Deletes the specified VpnTunnel resource.

VpnTunnelGetCall
Returns the specified VpnTunnel resource.

VpnTunnelInsertCall
Creates a VpnTunnel resource in the specified project and region using the data included in the request.

VpnTunnelList
Contains a list of VpnTunnel resources.

VpnTunnelListCall
Retrieves the list of VpnTunnel resources contained in the specified project and region.

VpnTunnelMethods
A builder providing access to all methods supported on vpnTunnel resources. It is not used directly, but through the Compute hub.

VpnTunnelsScopedList
There is no detailed description.

VpnTunnelsScopedListWarning
Informational warning which replaces the list of addresses when the list is empty.

VpnTunnelsScopedListWarningData
[Output Only] Metadata for this warning in key: value format.

Zone
A Zone resource.

ZoneGetCall
Returns the specified zone resource.

ZoneList
Contains a list of zone resources.

ZoneListCall
Retrieves the list of zone resources available to the specified project.

ZoneMaintenanceWindows
[Output Only] Any scheduled maintenance windows for this zone. When the zone is in a maintenance window, all resources which reside in the zone will be unavailable. For more information, see Maintenance Windows

ZoneMethods
A builder providing access to all methods supported on zone resources. It is not used directly, but through the Compute hub.

ZoneOperationDeleteCall
Deletes the specified zone-specific operation resource.

ZoneOperationGetCall
Retrieves the specified zone-specific operation resource.

ZoneOperationListCall
Retrieves the list of operation resources contained within the specified zone.

ZoneOperationMethods
A builder providing access to all methods supported on zoneOperation resources. It is not used directly, but through the Compute hub.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.