Struct google_compute1::api::Instance

pub struct Instance {
    pub advanced_machine_features: Option<AdvancedMachineFeatures>,
    pub can_ip_forward: Option<bool>,
    pub confidential_instance_config: Option<ConfidentialInstanceConfig>,
    pub cpu_platform: Option<String>,
    pub creation_timestamp: Option<String>,
    pub deletion_protection: Option<bool>,
    pub description: Option<String>,
    pub disks: Option<Vec<AttachedDisk>>,
    pub display_device: Option<DisplayDevice>,
    pub fingerprint: Option<String>,
    pub guest_accelerators: Option<Vec<AcceleratorConfig>>,
    pub hostname: Option<String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub label_fingerprint: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub last_start_timestamp: Option<String>,
    pub last_stop_timestamp: Option<String>,
    pub last_suspended_timestamp: Option<String>,
    pub machine_type: Option<String>,
    pub metadata: Option<Metadata>,
    pub min_cpu_platform: Option<String>,
    pub name: Option<String>,
    pub network_interfaces: Option<Vec<NetworkInterface>>,
    pub post_key_revocation_action_type: Option<String>,
    pub private_ipv6_google_access: Option<String>,
    pub reservation_affinity: Option<ReservationAffinity>,
    pub resource_policies: Option<Vec<String>>,
    pub satisfies_pzs: Option<bool>,
    pub scheduling: Option<Scheduling>,
    pub self_link: Option<String>,
    pub service_accounts: Option<Vec<ServiceAccount>>,
    pub shielded_instance_config: Option<ShieldedInstanceConfig>,
    pub shielded_instance_integrity_policy: Option<ShieldedInstanceIntegrityPolicy>,
    pub start_restricted: Option<bool>,
    pub status: Option<String>,
    pub status_message: Option<String>,
    pub tags: Option<Tags>,
    pub zone: Option<String>,
}

Represents an Instance resource.

An instance is a virtual machine that is hosted on Google Cloud Platform. For more information, read Virtual Machine Instances. (== resource_for {$api_version}.instances ==)

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

add access config instances (none)

add resource policies instances (none)

aggregated list instances (none)

attach disk instances (none)

bulk insert instances (none)

delete instances (none)

delete access config instances (none)

detach disk instances (none)

get instances (response)

get effective firewalls instances (none)

get guest attributes instances (none)

get iam policy instances (none)

get screenshot instances (none)

get serial port output instances (none)

get shielded instance identity instances (none)

insert instances (request)

list instances (none)

list referrers instances (none)

remove resource policies instances (none)

reset instances (none)

set deletion protection instances (none)

set disk auto delete instances (none)

set iam policy instances (none)

set labels instances (none)

set machine resources instances (none)

set machine type instances (none)

set metadata instances (none)

set min cpu platform instances (none)

set scheduling instances (none)

set service account instances (none)

set shielded instance integrity policy instances (none)

set tags instances (none)

simulate maintenance event instances (none)

start instances (none)

start with encryption key instances (none)

stop instances (none)

test iam permissions instances (none)

update instances (request)

update access config instances (none)

update display device instances (none)

update network interface instances (none)

update shielded instance config instances (none)

Fields
advanced_machine_features: Option<AdvancedMachineFeatures>

Controls for advanced machine-related behavior features.

can_ip_forward: Option<bool>

Allows this instance to send and receive packets with non-matching destination or source IPs. This is required if you plan to use this instance to forward routes. For more information, see Enabling IP Forwarding.

confidential_instance_config: Option<ConfidentialInstanceConfig>

no description provided

cpu_platform: Option<String>

[Output Only] The CPU platform used by this instance.

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

deletion_protection: Option<bool>

Whether the resource should be protected against deletion.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

disks: Option<Vec<AttachedDisk>>

Array of disks associated with this instance. Persistent disks must be created before you can assign them.

display_device: Option<DisplayDevice>

Enables display device for the instance.

fingerprint: Option<String>

Specifies a fingerprint for this resource, which is essentially a hash of the instance's contents and used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update the instance. You must always provide an up-to-date fingerprint hash in order to update the instance.

To see the latest fingerprint, make get() request to the instance.

guest_accelerators: Option<Vec<AcceleratorConfig>>

A list of the type and count of accelerator cards attached to the instance.

hostname: Option<String>

Specifies the hostname of the instance. The specified hostname must be RFC1035 compliant. If hostname is not specified, the default hostname is [INSTANCE_NAME].c.[PROJECT_ID].internal when using the global DNS, and [INSTANCE_NAME].[ZONE].c.[PROJECT_ID].internal when using zonal DNS.

id: Option<String>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

kind: Option<String>

[Output Only] Type of the resource. Always compute#instance for instances.

label_fingerprint: Option<String>

A fingerprint for this request, which is essentially a hash of the label's contents and used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update labels. You must always provide an up-to-date fingerprint hash in order to update or change labels.

To see the latest fingerprint, make get() request to the instance.

labels: Option<HashMap<String, String>>

Labels to apply to this instance. These can be later modified by the setLabels method.

last_start_timestamp: Option<String>

[Output Only] Last start timestamp in RFC3339 text format.

last_stop_timestamp: Option<String>

[Output Only] Last stop timestamp in RFC3339 text format.

last_suspended_timestamp: Option<String>

[Output Only] Last suspended timestamp in RFC3339 text format.

machine_type: Option<String>

Full or partial URL of the machine type resource to use for this instance, in the format: zones/zone/machineTypes/machine-type. This is provided by the client when the instance is created. For example, the following is a valid partial url to a predefined machine type: zones/us-central1-f/machineTypes/n1-standard-1

To create a custom machine type, provide a URL to a machine type in the following format, where CPUS is 1 or an even number up to 32 (2, 4, 6, ... 24, etc), and MEMORY is the total memory for this instance. Memory must be a multiple of 256 MB and must be supplied in MB (e.g. 5 GB of memory is 5120 MB): zones/zone/machineTypes/custom-CPUS-MEMORY

For example: zones/us-central1-f/machineTypes/custom-4-5120

For a full list of restrictions, read the Specifications for custom machine types.

metadata: Option<Metadata>

The metadata key/value pairs assigned to this instance. This includes custom metadata and predefined keys.

min_cpu_platform: Option<String>

Specifies a minimum CPU platform for the VM instance. Applicable values are the friendly names of CPU platforms, such as minCpuPlatform: "Intel Haswell" or minCpuPlatform: "Intel Sandy Bridge".

name: Option<String>

The name of the resource, provided by the client when initially creating the resource. The resource name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z]([-a-z0-9]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

network_interfaces: Option<Vec<NetworkInterface>>

An array of network configurations for this instance. These specify how interfaces are configured to interact with other network services, such as connecting to the internet. Multiple interfaces are supported per instance.

post_key_revocation_action_type: Option<String>

PostKeyRevocationActionType of the instance.

private_ipv6_google_access: Option<String>

The private IPv6 google access type for the VM. If not specified, use INHERIT_FROM_SUBNETWORK as default.

reservation_affinity: Option<ReservationAffinity>

Specifies the reservations that this instance can consume from.

resource_policies: Option<Vec<String>>

Resource policies applied to this instance.

satisfies_pzs: Option<bool>

[Output Only] Reserved for future use.

scheduling: Option<Scheduling>

Sets the scheduling options for this instance.

self_link: Option<String>

[Output Only] Server-defined URL for this resource.

service_accounts: Option<Vec<ServiceAccount>>

A list of service accounts, with their specified scopes, authorized for this instance. Only one service account per VM instance is supported.

Service accounts generate access tokens that can be accessed through the metadata server and used to authenticate applications on the instance. See Service Accounts for more information.

shielded_instance_config: Option<ShieldedInstanceConfig>

no description provided

shielded_instance_integrity_policy: Option<ShieldedInstanceIntegrityPolicy>

no description provided

start_restricted: Option<bool>

[Output Only] Whether a VM has been restricted for start because Compute Engine has detected suspicious activity.

status: Option<String>

[Output Only] The status of the instance. One of the following values: PROVISIONING, STAGING, RUNNING, STOPPING, SUSPENDING, SUSPENDED, REPAIRING, and TERMINATED. For more information about the status of the instance, see Instance life cycle.

status_message: Option<String>

[Output Only] An optional, human-readable explanation of the status.

tags: Option<Tags>

Tags to apply to this instance. Tags are used to identify valid sources or targets for network firewalls and are specified by the client during instance creation. The tags can be later modified by the setTags method. Each tag within the list must comply with RFC1035. Multiple tags can be specified via the 'tags.items' field.

zone: Option<String>

[Output Only] URL of the zone where the instance resides. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.