Struct google_compute1::BackendService

pub struct BackendService {
    pub kind: Option<String>,
    pub protocol: Option<String>,
    pub description: Option<String>,
    pub timeout_sec: Option<i32>,
    pub id: Option<String>,
    pub backends: Option<Vec<Backend>>,
    pub fingerprint: Option<String>,
    pub port_name: Option<String>,
    pub health_checks: Option<Vec<String>>,
    pub creation_timestamp: Option<String>,
    pub port: Option<i32>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

A BackendService resource. This resource defines a group of backend virtual machines and their serving capacity.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get backend services (response)

insert backend services (request)

list backend services (none)

update backend services (request)

get health backend services (none)

delete backend services (none)

patch backend services (request)

Fields

kind

[Output Only] Type of resource. Always compute#backendService for backend services.

protocol

no description provided

description

An optional description of this resource. Provide this property when you create the resource.

timeout_sec

How many seconds to wait for the backend before considering it a failed request. Default is 30 seconds. Valid range is [1, 86400].

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

backends

The list of backends that serve this BackendService.

fingerprint

Fingerprint of this resource. A hash of the contents stored in this object. This field is used in optimistic locking. This field will be ignored when inserting a BackendService. An up-to-date fingerprint must be provided in order to update the BackendService.

port_name

Name of backend port. The same name should appear in the resource views referenced by this service. Required.

health_checks

The list of URLs to the HttpHealthCheck or HttpsHealthCheck resource for health checking this BackendService. Currently at most one health check can be specified, and a health check is required.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

port

Deprecated in favor of portName. The TCP port to connect on the backend. The default value is 80.

self_link

[Output Only] Server-defined URL for the resource.

name

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.