Struct google_compute1::Compute

pub struct Compute<C, A> {
    // some fields omitted
}

Central instance to access all Compute related resource activities

Examples

Instantiate a new hub

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
use compute1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.instances().set_disk_auto_delete("project", "zone", "instance", true, "deviceName")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Methods

impl<'a, C, A> Compute<C, A> where C: BorrowMut<Client>, A: GetToken

fn new(client: C, authenticator: A) -> Compute<C, A>

fn addresses(&'a self) -> AddresseMethods<'a, C, A>

fn autoscalers(&'a self) -> AutoscalerMethods<'a, C, A>

fn backend_services(&'a self) -> BackendServiceMethods<'a, C, A>

fn disk_types(&'a self) -> DiskTypeMethods<'a, C, A>

fn disks(&'a self) -> DiskMethods<'a, C, A>

fn firewalls(&'a self) -> FirewallMethods<'a, C, A>

fn forwarding_rules(&'a self) -> ForwardingRuleMethods<'a, C, A>

fn global_addresses(&'a self) -> GlobalAddresseMethods<'a, C, A>

fn global_forwarding_rules(&'a self) -> GlobalForwardingRuleMethods<'a, C, A>

fn global_operations(&'a self) -> GlobalOperationMethods<'a, C, A>

fn http_health_checks(&'a self) -> HttpHealthCheckMethods<'a, C, A>

fn https_health_checks(&'a self) -> HttpsHealthCheckMethods<'a, C, A>

fn images(&'a self) -> ImageMethods<'a, C, A>

fn instance_group_managers(&'a self) -> InstanceGroupManagerMethods<'a, C, A>

fn instance_groups(&'a self) -> InstanceGroupMethods<'a, C, A>

fn instance_templates(&'a self) -> InstanceTemplateMethods<'a, C, A>

fn instances(&'a self) -> InstanceMethods<'a, C, A>

fn licenses(&'a self) -> LicenseMethods<'a, C, A>

fn machine_types(&'a self) -> MachineTypeMethods<'a, C, A>

fn networks(&'a self) -> NetworkMethods<'a, C, A>

fn projects(&'a self) -> ProjectMethods<'a, C, A>

fn region_operations(&'a self) -> RegionOperationMethods<'a, C, A>

fn regions(&'a self) -> RegionMethods<'a, C, A>

fn routes(&'a self) -> RouteMethods<'a, C, A>

fn snapshots(&'a self) -> SnapshotMethods<'a, C, A>

fn ssl_certificates(&'a self) -> SslCertificateMethods<'a, C, A>

fn subnetworks(&'a self) -> SubnetworkMethods<'a, C, A>

fn target_http_proxies(&'a self) -> TargetHttpProxyMethods<'a, C, A>

fn target_https_proxies(&'a self) -> TargetHttpsProxyMethods<'a, C, A>

fn target_instances(&'a self) -> TargetInstanceMethods<'a, C, A>

fn target_pools(&'a self) -> TargetPoolMethods<'a, C, A>

fn target_vpn_gateways(&'a self) -> TargetVpnGatewayMethods<'a, C, A>

fn url_maps(&'a self) -> UrlMapMethods<'a, C, A>

fn vpn_tunnels(&'a self) -> VpnTunnelMethods<'a, C, A>

fn zone_operations(&'a self) -> ZoneOperationMethods<'a, C, A>

fn zones(&'a self) -> ZoneMethods<'a, C, A>

fn user_agent(&mut self, agent_name: String) -> String

Set the user-agent header field to use in all requests to the server. It defaults to google-api-rust-client/0.1.14.

Returns the previously set user-agent.