Struct google_compute1::Disk

pub struct Disk {
    pub status: Option<String>,
    pub source_snapshot: Option<String>,
    pub source_image_encryption_key: Option<CustomerEncryptionKey>,
    pub last_detach_timestamp: Option<String>,
    pub description: Option<String>,
    pub size_gb: Option<String>,
    pub label_fingerprint: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub guest_os_features: Option<Vec<GuestOsFeature>>,
    pub source_snapshot_id: Option<String>,
    pub resource_policies: Option<Vec<String>>,
    pub replica_zones: Option<Vec<String>>,
    pub licenses: Option<Vec<String>>,
    pub users: Option<Vec<String>>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub license_codes: Option<Vec<String>>,
    pub disk_encryption_key: Option<CustomerEncryptionKey>,
    pub source_image_id: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub zone: Option<String>,
    pub region: Option<String>,
    pub physical_block_size_bytes: Option<String>,
    pub last_attach_timestamp: Option<String>,
    pub source_snapshot_encryption_key: Option<CustomerEncryptionKey>,
    pub type_: Option<String>,
    pub options: Option<String>,
    pub self_link: Option<String>,
    pub source_image: Option<String>,
}

Represents a Persistent Disk resource.

Persistent disks are required for running your VM instances. Create both boot and non-boot (data) persistent disks. For more information, read Persistent Disks. For more storage options, read Storage options.

The disks resource represents a zonal persistent disk. For more information, read Zonal persistent disks.

The regionDisks resource represents a regional persistent disk. For more information, read Regional resources. (== resource_for beta.disks ==) (== resource_for v1.disks ==) (== resource_for v1.regionDisks ==) (== resource_for beta.regionDisks ==)

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert disks (request)

set iam policy disks (none)

resize disks (none)

delete disks (none)

add resource policies disks (none)

remove resource policies disks (none)

test iam permissions disks (none)

get disks (response)

get iam policy disks (none)

insert region disks (request)

create snapshot disks (none)

set labels disks (none)

aggregated list disks (none)

list disks (none)

get region disks (response)

Fields
status: Option<String>

[Output Only] The status of disk creation.

source_snapshot: Option<String>

The source snapshot used to create this disk. You can provide this as a partial or full URL to the resource. For example, the following are valid values:

https://www.googleapis.com/compute/v1/projects/project/global/snapshots/snapshot

projects/project/global/snapshots/snapshot

global/snapshots/snapshot

source_image_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source image. Required if the source image is protected by a customer-supplied encryption key.

last_detach_timestamp: Option<String>

[Output Only] Last detach timestamp in RFC3339 text format.

description: Option<String>

An optional description of this resource. Provide this property when you create the resource.

size_gb: Option<String>

Size of the persistent disk, specified in GB. You can specify this field when creating a persistent disk using the sourceImage or sourceSnapshot parameter, or specify it alone to create an empty persistent disk.

If you specify this field along with sourceImage or sourceSnapshot, the value of sizeGb must not be less than the size of the sourceImage or the size of the snapshot. Acceptable values are 1 to 65536, inclusive.

label_fingerprint: Option<String>

A fingerprint for the labels being applied to this disk, which is essentially a hash of the labels set used for optimistic locking. The fingerprint is initially generated by Compute Engine and changes after every request to modify or update labels. You must always provide an up-to-date fingerprint hash in order to update or change labels, otherwise the request will fail with error 412 conditionNotMet.

To see the latest fingerprint, make a get() request to retrieve a disk.

labels: Option<HashMap<String, String>>

Labels to apply to this disk. These can be later modified by the setLabels method.

guest_os_features: Option<Vec<GuestOsFeature>>

A list of features to enable on the guest operating system. Applicable only for bootable images. Read Enabling guest operating system features to see a list of available options.

source_snapshot_id: Option<String>

[Output Only] The unique ID of the snapshot used to create this disk. This value identifies the exact snapshot that was used to create this persistent disk. For example, if you created the persistent disk from a snapshot that was later deleted and recreated under the same name, the source snapshot ID would identify the exact version of the snapshot that was used.

resource_policies: Option<Vec<String>>

Resource policies applied to this disk for automatic snapshot creations.

replica_zones: Option<Vec<String>>

URLs of the zones where the disk should be replicated to. Only applicable for regional resources.

licenses: Option<Vec<String>>

A list of publicly visible licenses. Reserved for Google's use.

users: Option<Vec<String>>

[Output Only] Links to the users of the disk (attached instances) in form: projects/project/zones/zone/instances/instance

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

id: Option<String>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

license_codes: Option<Vec<String>>

Integer license codes indicating which licenses are attached to this disk.

disk_encryption_key: Option<CustomerEncryptionKey>

Encrypts the disk using a customer-supplied encryption key.

After you encrypt a disk with a customer-supplied key, you must provide the same key if you use the disk later (e.g. to create a disk snapshot, to create a disk image, to create a machine image, or to attach the disk to a virtual machine).

Customer-supplied encryption keys do not protect access to metadata of the disk.

If you do not provide an encryption key when creating the disk, then the disk will be encrypted using an automatically generated key and you do not need to provide a key to use the disk later.

source_image_id: Option<String>

[Output Only] The ID value of the image used to create this disk. This value identifies the exact image that was used to create this persistent disk. For example, if you created the persistent disk from an image that was later deleted and recreated under the same name, the source image ID would identify the exact version of the image that was used.

kind: Option<String>

[Output Only] Type of the resource. Always compute#disk for disks.

name: Option<String>

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression [a-z]([-a-z0-9]*[a-z0-9])? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

zone: Option<String>

[Output Only] URL of the zone where the disk resides. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

region: Option<String>

[Output Only] URL of the region where the disk resides. Only applicable for regional resources. You must specify this field as part of the HTTP request URL. It is not settable as a field in the request body.

physical_block_size_bytes: Option<String>

Physical block size of the persistent disk, in bytes. If not present in a request, a default value is used. Currently supported sizes are 4096 and 16384, other sizes may be added in the future. If an unsupported value is requested, the error message will list the supported values for the caller's project.

last_attach_timestamp: Option<String>

[Output Only] Last attach timestamp in RFC3339 text format.

source_snapshot_encryption_key: Option<CustomerEncryptionKey>

The customer-supplied encryption key of the source snapshot. Required if the source snapshot is protected by a customer-supplied encryption key.

type_: Option<String>

URL of the disk type resource describing which disk type to use to create the disk. Provide this when creating the disk. For example: projects/project/zones/zone/diskTypes/pd-standard or pd-ssd

options: Option<String>

Internal use only.

self_link: Option<String>

[Output Only] Server-defined fully-qualified URL for this resource.

source_image: Option<String>

The source image used to create this disk. If the source image is deleted, this field will not be set.

To create a disk with one of the public operating system images, specify the image by its family name. For example, specify family/debian-9 to use the latest Debian 9 image: projects/debian-cloud/global/images/family/debian-9

Alternatively, use a specific version of a public operating system image: projects/debian-cloud/global/images/debian-9-stretch-vYYYYMMDD

To create a disk with a custom image that you created, specify the image name in the following format: global/images/my-custom-image

You can also specify a custom image by its image family, which returns the latest version of the image in that family. Replace the image name with family/family-name: global/images/family/my-image-family