Struct google_compute1::GlobalForwardingRuleMethods

pub struct GlobalForwardingRuleMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on globalForwardingRule resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)` and `set_target(...)`
// to build up your call.
let rb = hub.global_forwarding_rules();

Methods

impl<'a, C, A> GlobalForwardingRuleMethods<'a, C, A>

fn delete(&self, project: &str, forwarding_rule: &str) -> GlobalForwardingRuleDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified ForwardingRule resource.

Arguments

project - Project ID for this request.

forwardingRule - Name of the ForwardingRule resource to delete.

fn set_target(&self, request: TargetReference, project: &str, forwarding_rule: &str) -> GlobalForwardingRuleSetTargetCall<'a, C, A>

Create a builder to help you perform the following task:

Changes target URL for forwarding rule. The new target should be of the same type as the old target.

Arguments

request - No description provided.

project - Project ID for this request.

forwardingRule - Name of the ForwardingRule resource in which target is to be set.

fn insert(&self, request: ForwardingRule, project: &str) -> GlobalForwardingRuleInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a ForwardingRule resource in the specified project and region using the data included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

fn list(&self, project: &str) -> GlobalForwardingRuleListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of ForwardingRule resources available to the specified project.

Arguments

project - Project ID for this request.

fn get(&self, project: &str, forwarding_rule: &str) -> GlobalForwardingRuleGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the specified ForwardingRule resource. Get a list of available forwarding rules by making a list() request.

Arguments

project - Project ID for this request.

forwardingRule - Name of the ForwardingRule resource to return.