Struct google_compute1::Instance

pub struct Instance {
    pub cpu_platform: Option<String>,
    pub status: Option<String>,
    pub can_ip_forward: Option<bool>,
    pub description: Option<String>,
    pub tags: Option<Tags>,
    pub scheduling: Option<Scheduling>,
    pub machine_type: Option<String>,
    pub service_accounts: Option<Vec<ServiceAccount>>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub network_interfaces: Option<Vec<NetworkInterface>>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub zone: Option<String>,
    pub disks: Option<Vec<AttachedDisk>>,
    pub self_link: Option<String>,
    pub status_message: Option<String>,
    pub metadata: Option<Metadata>,
}

An Instance resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get serial port output instances (none)

attach disk instances (none)

insert instances (request)

stop instances (none)

set metadata instances (none)

start instances (none)

detach disk instances (none)

delete access config instances (none)

set scheduling instances (none)

get instances (response)

set tags instances (none)

set machine type instances (none)

list instances (none)

delete instances (none)

reset instances (none)

aggregated list instances (none)

add access config instances (none)

set disk auto delete instances (none)

Fields

cpu_platform

[Output Only] The CPU platform used by this instance.

status

[Output Only] The status of the instance. One of the following values: PROVISIONING, STAGING, RUNNING, STOPPING, and TERMINATED.

can_ip_forward

Allows this instance to send and receive packets with non-matching destination or source IPs. This is required if you plan to use this instance to forward routes. For more information, see Enabling IP Forwarding.

description

An optional description of this resource. Provide this property when you create the resource.

tags

A list of tags to apply to this instance. Tags are used to identify valid sources or targets for network firewalls and are specified by the client during instance creation. The tags can be later modified by the setTags method. Each tag within the list must comply with RFC1035.

scheduling

Scheduling options for this instance.

machine_type

Full or partial URL of the machine type resource to use for this instance, in the format: zones/zone/machineTypes/machine-type. This is provided by the client when the instance is created. For example, the following is a valid partial url to a predefined machine type:

zones/us-central1-f/machineTypes/n1-standard-1

To create a custom machine type, provide a URL to a machine type in the following format, where CPUS is 1 or an even number up to 32 (2, 4, 6, ... 24, etc), and MEMORY is the total memory for this instance. Memory must be a multiple of 256 MB and must be supplied in MB (e.g. 5 GB of memory is 5120 MB):

zones/zone/machineTypes/custom-CPUS-MEMORY

For example: zones/us-central1-f/machineTypes/custom-4-5120

For a full list of restrictions, read the Specifications for custom machine types.

service_accounts

A list of service accounts, with their specified scopes, authorized for this instance. Service accounts generate access tokens that can be accessed through the metadata server and used to authenticate applications on the instance. See Authenticating from Google Compute Engine for more information.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

network_interfaces

An array of configurations for this interface. This specifies how this interface is configured to interact with other network services, such as connecting to the internet.

kind

[Output Only] Type of the resource. Always compute#instance for instances.

name

The name of the resource, provided by the client when initially creating the resource. The resource name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

zone

[Output Only] URL of the zone where the instance resides.

disks

Array of disks associated with this instance. Persistent disks must be created before you can assign them.

self_link

[Output Only] Server-defined URL for this resource.

status_message

[Output Only] An optional, human-readable explanation of the status.

metadata

The metadata key/value pairs assigned to this instance. This includes custom metadata and predefined keys.