Struct google_compute1::InstanceGroupManagerMethods

pub struct InstanceGroupManagerMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on instanceGroupManager resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `abandon_instances(...)`, `aggregated_list(...)`, `delete(...)`, `delete_instances(...)`, `get(...)`, `insert(...)`, `list(...)`, `list_managed_instances(...)`, `recreate_instances(...)`, `resize(...)`, `set_instance_template(...)` and `set_target_pools(...)`
// to build up your call.
let rb = hub.instance_group_managers();

Methods

impl<'a, C, A> InstanceGroupManagerMethods<'a, C, A>

fn resize(&self, project: &str, zone: &str, instance_group_manager: &str, size: i32) -> InstanceGroupManagerResizeCall<'a, C, A>

Create a builder to help you perform the following task:

Resizes the managed instance group. If you increase the size, the group creates new instances using the current instance template. If you decrease the size, the group deletes instances. The resize operation is marked DONE when the resize actions are scheduled even if the group has not yet added or deleted any instances. You must separately verify the status of the creating or deleting actions with the listmanagedinstances method.

Arguments

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

size - The number of running instances that the managed instance group should maintain at any given time. The group automatically adds or removes instances to maintain the number of instances specified by this parameter.

fn list(&self, project: &str, zone: &str) -> InstanceGroupManagerListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of managed instance groups that are contained within the specified project and zone.

Arguments

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

fn list_managed_instances(&self, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerListManagedInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all of the instances in the managed instance group. Each instance in the list has a currentAction, which indicates the action that the managed instance group is performing on the instance. For example, if the group is still creating an instance, the currentAction is CREATING. If a previous action failed, the list displays the errors for that failed action.

Arguments

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

fn insert(&self, request: InstanceGroupManager, project: &str, zone: &str) -> InstanceGroupManagerInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a managed instance group using the information that you specify in the request. After the group is created, it schedules an action to create instances in the group using the specified instance template. This operation is marked as DONE when the group is created even if the instances in the group have not yet been created. You must separately verify the status of the individual instances with the listmanagedinstances method.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where you want to create the managed instance group.

fn aggregated_list(&self, project: &str) -> InstanceGroupManagerAggregatedListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the list of managed instance groups and groups them by zone.

Arguments

project - Project ID for this request.

fn delete_instances(&self, request: InstanceGroupManagersDeleteInstancesRequest, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerDeleteInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Schedules a group action to delete the specified instances in the managed instance group. The instances are also removed from any target pools of which they were a member. This method reduces the targetSize of the managed instance group by the number of instances that you delete. This operation is marked as DONE when the action is scheduled even if the instances are still being deleted. You must separately verify the status of the deleting action with the listmanagedinstances method.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

fn set_instance_template(&self, request: InstanceGroupManagersSetInstanceTemplateRequest, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerSetInstanceTemplateCall<'a, C, A>

Create a builder to help you perform the following task:

Specifies the instance template to use when creating new instances in this group. The templates for existing instances in the group do not change unless you recreate them.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

fn set_target_pools(&self, request: InstanceGroupManagersSetTargetPoolsRequest, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerSetTargetPoolCall<'a, C, A>

Create a builder to help you perform the following task:

Modifies the target pools to which all instances in this managed instance group are assigned. The target pools automatically apply to all of the instances in the managed instance group. This operation is marked DONE when you make the request even if the instances have not yet been added to their target pools. The change might take some time to apply to all of the instances in the group depending on the size of the group.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

fn delete(&self, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified managed instance group and all of the instances in that group. Note that the instance group must not belong to a backend service. Read Deleting an instance group for more information.

Arguments

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group to delete.

fn get(&self, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns all of the details about the specified managed instance group. Get a list of available managed instance groups by making a list() request.

Arguments

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

fn abandon_instances(&self, request: InstanceGroupManagersAbandonInstancesRequest, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerAbandonInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Schedules a group action to remove the specified instances from the managed instance group. Abandoning an instance does not delete the instance, but it does remove the instance from any target pools that are applied by the managed instance group. This method reduces the targetSize of the managed instance group by the number of instances that you abandon. This operation is marked as DONE when the action is scheduled even if the instances have not yet been removed from the group. You must separately verify the status of the abandoning action with the listmanagedinstances method.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.

fn recreate_instances(&self, request: InstanceGroupManagersRecreateInstancesRequest, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerRecreateInstanceCall<'a, C, A>

Create a builder to help you perform the following task:

Schedules a group action to recreate the specified instances in the managed instance group. The instances are deleted and recreated using the current instance template for the managed instance group. This operation is marked as DONE when the action is scheduled even if the instances have not yet been recreated. You must separately verify the status of the recreating action with the listmanagedinstances method.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone where the managed instance group is located.

instanceGroupManager - The name of the managed instance group.