Struct google_compute1::InstanceMethods

pub struct InstanceMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on instance resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_compute1 as compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `add_access_config(...)`, `aggregated_list(...)`, `attach_disk(...)`, `delete(...)`, `delete_access_config(...)`, `detach_disk(...)`, `get(...)`, `get_serial_port_output(...)`, `insert(...)`, `list(...)`, `reset(...)`, `set_disk_auto_delete(...)`, `set_machine_type(...)`, `set_metadata(...)`, `set_scheduling(...)`, `set_tags(...)`, `start(...)` and `stop(...)`
// to build up your call.
let rb = hub.instances();

Methods

impl<'a, C, A> InstanceMethods<'a, C, A>

fn set_machine_type(&self, request: InstancesSetMachineTypeRequest, project: &str, zone: &str, instance: &str) -> InstanceSetMachineTypeCall<'a, C, A>

Create a builder to help you perform the following task:

Changes the machine type for a stopped instance to the machine type specified in the request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

instance - Name of the instance scoping this request.

fn set_scheduling(&self, request: Scheduling, project: &str, zone: &str, instance: &str) -> InstanceSetSchedulingCall<'a, C, A>

Create a builder to help you perform the following task:

Sets an instance's scheduling options.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

instance - Instance name.

fn delete(&self, project: &str, zone: &str, instance: &str) -> InstanceDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified Instance resource. For more information, see Stopping or Deleting an Instance.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

instance - Name of the instance resource to delete.

fn get_serial_port_output(&self, project: &str, zone: &str, instance: &str) -> InstanceGetSerialPortOutputCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the specified instance's serial port output.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

instance - Name of the instance scoping this request.

fn set_disk_auto_delete(&self, project: &str, zone: &str, instance: &str, auto_delete: bool, device_name: &str) -> InstanceSetDiskAutoDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the auto-delete flag for a disk attached to an instance.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

instance - The instance name.

autoDelete - Whether to auto-delete the disk when the instance is deleted.

deviceName - The device name of the disk to modify.

fn set_tags(&self, request: Tags, project: &str, zone: &str, instance: &str) -> InstanceSetTagCall<'a, C, A>

Create a builder to help you perform the following task:

Sets tags for the specified instance to the data included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

instance - Name of the instance scoping this request.

fn add_access_config(&self, request: AccessConfig, project: &str, zone: &str, instance: &str, network_interface: &str) -> InstanceAddAccessConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Adds an access config to an instance's network interface.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

instance - The instance name for this request.

networkInterface - The name of the network interface to add to this instance.

fn set_metadata(&self, request: Metadata, project: &str, zone: &str, instance: &str) -> InstanceSetMetadataCall<'a, C, A>

Create a builder to help you perform the following task:

Sets metadata for the specified instance to the data included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

instance - Name of the instance scoping this request.

fn start(&self, project: &str, zone: &str, instance: &str) -> InstanceStartCall<'a, C, A>

Create a builder to help you perform the following task:

Starts an instance that was stopped using the using the instances().stop method. For more information, see Restart an instance.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

instance - Name of the instance resource to start.

fn get(&self, project: &str, zone: &str, instance: &str) -> InstanceGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the specified Instance resource. Get a list of available instances by making a list() request.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

instance - Name of the instance resource to return.

fn stop(&self, project: &str, zone: &str, instance: &str) -> InstanceStopCall<'a, C, A>

Create a builder to help you perform the following task:

Stops a running instance, shutting it down cleanly, and allows you to restart the instance at a later time. Stopped instances do not incur per-minute, virtual machine usage charges while they are stopped, but any resources that the virtual machine is using, such as persistent disks and static IP addresses, will continue to be charged until they are deleted. For more information, see Stopping an instance.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

instance - Name of the instance resource to stop.

fn detach_disk(&self, project: &str, zone: &str, instance: &str, device_name: &str) -> InstanceDetachDiskCall<'a, C, A>

Create a builder to help you perform the following task:

Detaches a disk from an instance.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

instance - Instance name.

deviceName - Disk device name to detach.

fn attach_disk(&self, request: AttachedDisk, project: &str, zone: &str, instance: &str) -> InstanceAttachDiskCall<'a, C, A>

Create a builder to help you perform the following task:

Attaches a Disk resource to an instance.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

instance - The instance name for this request.

fn delete_access_config(&self, project: &str, zone: &str, instance: &str, access_config: &str, network_interface: &str) -> InstanceDeleteAccessConfigCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an access config from an instance's network interface.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

instance - The instance name for this request.

accessConfig - The name of the access config to delete.

networkInterface - The name of the network interface.

fn insert(&self, request: Instance, project: &str, zone: &str) -> InstanceInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an instance resource in the specified project using the data included in the request.

Arguments

request - No description provided.

project - Project ID for this request.

zone - The name of the zone for this request.

fn list(&self, project: &str, zone: &str) -> InstanceListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the list of instances contained within the specified zone.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

fn reset(&self, project: &str, zone: &str, instance: &str) -> InstanceResetCall<'a, C, A>

Create a builder to help you perform the following task:

Performs a hard reset on the instance.

Arguments

project - Project ID for this request.

zone - The name of the zone for this request.

instance - Name of the instance scoping this request.

fn aggregated_list(&self, project: &str) -> InstanceAggregatedListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves aggregated list of instances.

Arguments

project - Project ID for this request.