Struct google_compute1::InstanceProperties

pub struct InstanceProperties {
    pub can_ip_forward: Option<bool>,
    pub description: Option<String>,
    pub tags: Option<Tags>,
    pub disks: Option<Vec<AttachedDisk>>,
    pub scheduling: Option<Scheduling>,
    pub machine_type: Option<String>,
    pub service_accounts: Option<Vec<ServiceAccount>>,
    pub network_interfaces: Option<Vec<NetworkInterface>>,
    pub metadata: Option<Metadata>,
}

This type is not used in any activity, and only used as part of another schema.

Fields

can_ip_forward

Enables instances created based on this template to send packets with source IP addresses other than their own and receive packets with destination IP addresses other than their own. If these instances will be used as an IP gateway or it will be set as the next-hop in a Route resource, specify true. If unsure, leave this set to false. See the canIpForward documentation for more information.

description

An optional text description for the instances that are created from this instance template.

tags

A list of tags to apply to the instances that are created from this template. The tags identify valid sources or targets for network firewalls. The setTags method can modify this list of tags. Each tag within the list must comply with RFC1035.

disks

An array of disks that are associated with the instances that are created from this template.

scheduling

Specifies the scheduling options for the instances that are created from this template.

machine_type

The machine type to use for instances that are created from this template.

service_accounts

A list of service accounts with specified scopes. Access tokens for these service accounts are available to the instances that are created from this template. Use metadata queries to obtain the access tokens for these instances.

network_interfaces

An array of network access configurations for this interface.

metadata

The metadata key/value pairs to assign to instances that are created from this template. These pairs can consist of custom metadata or predefined keys. See Project and instance metadata for more information.