Struct google_compute1::MachineType

pub struct MachineType {
    pub guest_cpus: Option<i32>,
    pub image_space_gb: Option<i32>,
    pub kind: Option<String>,
    pub description: Option<String>,
    pub zone: Option<String>,
    pub maximum_persistent_disks_size_gb: Option<String>,
    pub deprecated: Option<DeprecationStatus>,
    pub maximum_persistent_disks: Option<i32>,
    pub memory_mb: Option<i32>,
    pub scratch_disks: Option<Vec<MachineTypeScratchDisks>>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

A Machine Type resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list machine types (none)

aggregated list machine types (none)

get machine types (response)

Fields

guest_cpus

[Output Only] The number of virtual CPUs that are available to the instance.

image_space_gb

[Deprecated] This property is deprecated and will never be populated with any relevant values.

kind

[Output Only] The type of the resource. Always compute#machineType for machine types.

description

[Output Only] An optional textual description of the resource.

zone

[Output Only] The name of the zone where the machine type resides, such as us-central1-a.

maximum_persistent_disks_size_gb

[Output Only] Maximum total persistent disks size (GB) allowed.

deprecated

[Output Only] The deprecation status associated with this machine type.

maximum_persistent_disks

[Output Only] Maximum persistent disks allowed.

memory_mb

[Output Only] The amount of physical memory available to the instance, defined in MB.

scratch_disks

[Output Only] List of extended scratch disks assigned to the instance.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

self_link

[Output Only] Server-defined URL for the resource.

name

[Output Only] Name of the resource.