Struct google_compute1::Operation

pub struct Operation {
    pub status: Option<String>,
    pub description: Option<String>,
    pub warnings: Option<Vec<OperationWarnings>>,
    pub error: Option<OperationError>,
    pub target_id: Option<String>,
    pub user: Option<String>,
    pub start_time: Option<String>,
    pub client_operation_id: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub zone: Option<String>,
    pub region: Option<String>,
    pub operation_type: Option<String>,
    pub self_link: Option<String>,
    pub insert_time: Option<String>,
    pub http_error_message: Option<String>,
    pub progress: Option<i32>,
    pub end_time: Option<String>,
    pub http_error_status_code: Option<i32>,
    pub status_message: Option<String>,
    pub target_link: Option<String>,
}

Represents an Operation resource.

You can use an operation resource to manage asynchronous API requests. For more information, read Handling API responses.

Operations can be global, regional or zonal.

For global operations, use the globalOperations resource.

For regional operations, use the regionOperations resource.

For zonal operations, use the zonalOperations resource.

For more information, read Global, Regional, and Zonal Resources. (== resource_for v1.globalOperations ==) (== resource_for beta.globalOperations ==) (== resource_for v1.regionOperations ==) (== resource_for beta.regionOperations ==) (== resource_for v1.zoneOperations ==) (== resource_for beta.zoneOperations ==)

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete backend buckets (response)

set ssl policy target ssl proxies (response)

delete ssl certificates (response)

resize region disks (response)

patch autoscalers (response)

resize instance group managers (response)

delete instance templates (response)

patch region instance group managers (response)

update url maps (response)

delete region backend services (response)

delete interconnects (response)

update shielded instance config instances (response)

add resource policies region disks (response)

delete vpn tunnels (response)

set labels region disks (response)

set machine type instances (response)

delete https health checks (response)

delete node groups (response)

delete ssl policies (response)

insert ssl policies (response)

disable xpn host projects (response)

set service account instances (response)

delete global forwarding rules (response)

insert vpn tunnels (response)

delete disks (response)

insert networks (response)

add resource policies disks (response)

patch health checks (response)

set scheduling instances (response)

delete instances (response)

add health check target pools (response)

insert instance group managers (response)

insert global addresses (response)

set ssl certificates target ssl proxies (response)

set target global forwarding rules (response)

insert autoscalers (response)

add rule security policies (response)

insert instance groups (response)

insert subnetworks (response)

insert routes (response)

resize disks (response)

delete target tcp proxies (response)

set backend service target ssl proxies (response)

set default network tier projects (response)

set instance template instance group managers (response)

set node template node groups (response)

delete resource policies (response)

delete region autoscalers (response)

remove peering networks (response)

patch ssl policies (response)

delete interconnect attachments (response)

patch firewalls (response)

insert target pools (response)

insert node templates (response)

disable xpn resource projects (response)

set disk auto delete instances (response)

recreate instances region instance group managers (response)

set target pools instance group managers (response)

insert disks (response)

delete target ssl proxies (response)

insert region backend services (response)

insert https health checks (response)

simulate maintenance event instances (response)

remove health check target pools (response)

add signed url key backend buckets (response)

set url map target https proxies (response)

create snapshot region disks (response)

delete backend services (response)

update network interface instances (response)

insert url maps (response)

add access config instances (response)

set machine resources instances (response)

update access config instances (response)

insert security policies (response)

insert addresses (response)

insert health checks (response)

insert global forwarding rules (response)

delete target https proxies (response)

insert target tcp proxies (response)

get global operations (response)

delete addresses (response)

insert target vpn gateways (response)

set backend service target tcp proxies (response)

create snapshot disks (response)

update health checks (response)

remove instances instance groups (response)

set metadata instances (response)

insert region commitments (response)

add signed url key backend services (response)

insert forwarding rules (response)

start instances (response)

delete instance group managers (response)

set deletion protection instances (response)

expand ip cidr range subnetworks (response)

set quic override target https proxies (response)

move disk projects (response)

patch interconnects (response)

insert firewalls (response)

set labels disks (response)

set backup target pools (response)

delete instances instance group managers (response)

set security policy backend services (response)

delete instances region instance group managers (response)

detach disk instances (response)

insert region disks (response)

delete instance groups (response)

delete access config instances (response)

set ssl certificates target https proxies (response)

add instances instance groups (response)

abandon instances instance group managers (response)

set tags instances (response)

attach network endpoints network endpoint groups (response)

resize reservations (response)

delete snapshots (response)

insert target https proxies (response)

insert reservations (response)

set proxy header target tcp proxies (response)

update backend services (response)

remove resource policies disks (response)

delete nodes node groups (response)

delete licenses (response)

remove resource policies region disks (response)

patch instance group managers (response)

patch subnetworks (response)

delete reservations (response)

set ssl policy target https proxies (response)

attach disk instances (response)

set common instance metadata projects (response)

delete node templates (response)

get region operations (response)

add nodes node groups (response)

insert backend buckets (response)

insert node groups (response)

insert licenses (response)

patch backend services (response)

set url map target http proxies (response)

delete network endpoint groups (response)

delete region instance group managers (response)

deprecate images (response)

patch http health checks (response)

insert region autoscalers (response)

patch rule security policies (response)

patch interconnect attachments (response)

insert images (response)

delete signed url key backend buckets (response)

invalidate cache url maps (response)

set named ports instance groups (response)

insert ssl certificates (response)

insert interconnects (response)

switch to custom mode networks (response)

enable xpn resource projects (response)

move instance projects (response)

patch security policies (response)

delete autoscalers (response)

set labels snapshots (response)

recreate instances instance group managers (response)

update https health checks (response)

patch region autoscalers (response)

patch url maps (response)

delete subnetworks (response)

delete region disks (response)

stop instances (response)

add instance target pools (response)

remove instance target pools (response)

detach network endpoints network endpoint groups (response)

delete target pools (response)

update firewalls (response)

insert instances (response)

set usage export bucket projects (response)

patch region backend services (response)

set labels images (response)

insert target instances (response)

insert target http proxies (response)

insert interconnect attachments (response)

patch networks (response)

insert resource policies (response)

remove rule security policies (response)

set named ports region instance groups (response)

insert instance templates (response)

delete target instances (response)

set instance template region instance group managers (response)

enable xpn host projects (response)

set target forwarding rules (response)

delete routers (response)

resize region instance group managers (response)

delete target vpn gateways (response)

delete health checks (response)

insert backend services (response)

set labels instances (response)

update region backend services (response)

set shielded instance integrity policy instances (response)

delete target http proxies (response)

patch backend buckets (response)

delete global addresses (response)

patch https health checks (response)

delete networks (response)

delete url maps (response)

insert routers (response)

reset instances (response)

update routers (response)

patch routers (response)

add peering networks (response)

delete security policies (response)

set proxy header target ssl proxies (response)

delete http health checks (response)

insert http health checks (response)

delete images (response)

abandon instances region instance group managers (response)

set private ip google access subnetworks (response)

update autoscalers (response)

set target pools region instance group managers (response)

update region autoscalers (response)

insert target ssl proxies (response)

start with encryption key instances (response)

set min cpu platform instances (response)

delete forwarding rules (response)

insert network endpoint groups (response)

update backend buckets (response)

insert region instance group managers (response)

delete firewalls (response)

get zone operations (response)

update http health checks (response)

delete routes (response)

delete signed url key backend services (response)

Fields
status: Option<String>

[Output Only] The status of the operation, which can be one of the following: PENDING, RUNNING, or DONE.

description: Option<String>

[Output Only] A textual description of the operation, which is set when the operation is created.

warnings: Option<Vec<OperationWarnings>>

[Output Only] If warning messages are generated during processing of the operation, this field will be populated.

error: Option<OperationError>

[Output Only] If errors are generated during processing of the operation, this field will be populated.

target_id: Option<String>

[Output Only] The unique target ID, which identifies a specific incarnation of the target resource.

user: Option<String>

[Output Only] User who requested the operation, for example: user@example.com.

start_time: Option<String>

[Output Only] The time that this operation was started by the server. This value is in RFC3339 text format.

client_operation_id: Option<String>

[Output Only] The value of requestId if you provided it in the request. Not present otherwise.

creation_timestamp: Option<String>

[Deprecated] This field is deprecated.

id: Option<String>

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

kind: Option<String>

[Output Only] Type of the resource. Always compute#operation for Operation resources.

name: Option<String>

[Output Only] Name of the resource.

zone: Option<String>

[Output Only] The URL of the zone where the operation resides. Only applicable when performing per-zone operations.

region: Option<String>

[Output Only] The URL of the region where the operation resides. Only applicable when performing regional operations.

operation_type: Option<String>

[Output Only] The type of operation, such as insert, update, or delete, and so on.

self_link: Option<String>

[Output Only] Server-defined URL for the resource.

insert_time: Option<String>

[Output Only] The time that this operation was requested. This value is in RFC3339 text format.

http_error_message: Option<String>

[Output Only] If the operation fails, this field contains the HTTP error message that was returned, such as NOT FOUND.

progress: Option<i32>

[Output Only] An optional progress indicator that ranges from 0 to 100. There is no requirement that this be linear or support any granularity of operations. This should not be used to guess when the operation will be complete. This number should monotonically increase as the operation progresses.

end_time: Option<String>

[Output Only] The time that this operation was completed. This value is in RFC3339 text format.

http_error_status_code: Option<i32>

[Output Only] If the operation fails, this field contains the HTTP error status code that was returned. For example, a 404 means the resource was not found.

status_message: Option<String>

[Output Only] An optional textual description of the current status of the operation.

target_link: Option<String>

[Output Only] The URL of the resource that the operation modifies. For operations related to creating a snapshot, this points to the persistent disk that the snapshot was created from.