Struct google_compute1::Project

pub struct Project {
    pub kind: Option<String>,
    pub description: Option<String>,
    pub common_instance_metadata: Option<Metadata>,
    pub quotas: Option<Vec<Quota>>,
    pub usage_export_location: Option<UsageExportLocation>,
    pub enabled_features: Option<Vec<String>>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

A Project resource. Projects can only be created in the Google Cloud Platform Console. Unless marked otherwise, values can only be modified in the console.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

set common instance metadata projects (none)

set usage export bucket projects (none)

move instance projects (none)

get projects (response)

move disk projects (none)

Fields

kind

[Output Only] Type of the resource. Always compute#project for projects.

description

An optional textual description of the resource.

common_instance_metadata

Metadata key/value pairs available to all instances contained in this project. See Custom metadata for more information.

quotas

[Output Only] Quotas assigned to this project.

usage_export_location

The naming prefix for daily usage reports and the Google Cloud Storage bucket where they are stored.

enabled_features

Restricted features enabled for use on this project.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server. This is not the project ID, and is just a unique ID used by Compute Engine to identify resources.

self_link

[Output Only] Server-defined URL for the resource.

name

The project ID. For example: my-example-project. Use the project ID to make requests to Compute Engine.