Struct google_compute1::Route

pub struct Route {
    pub next_hop_gateway: Option<String>,
    pub description: Option<String>,
    pub tags: Option<Vec<String>>,
    pub next_hop_network: Option<String>,
    pub dest_range: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub next_hop_ip: Option<String>,
    pub network: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub next_hop_instance: Option<String>,
    pub warnings: Option<Vec<RouteWarnings>>,
    pub priority: Option<u32>,
    pub next_hop_vpn_tunnel: Option<String>,
    pub self_link: Option<String>,
}

Represents a Route resource. A route specifies how certain packets should be handled by the network. Routes are associated with instances by tags and the set of routes for a particular instance is called its routing table.

For each packet leaving a instance, the system searches that instance's routing table for a single best matching route. Routes match packets by destination IP address, preferring smaller or more specific ranges over larger ones. If there is a tie, the system selects the route with the smallest priority value. If there is still a tie, it uses the layer three and four packet headers to select just one of the remaining matching routes. The packet is then forwarded as specified by the nextHop field of the winning route - either to another instance destination, a instance gateway or a Google Compute Engine-operated gateway.

Packets that do not match any route in the sending instance's routing table are dropped.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list routes (none)

insert routes (request)

delete routes (none)

get routes (response)

Fields

next_hop_gateway

The URL to a gateway that should handle matching packets. You can only specify the internet gateway using a full or partial valid URL: projects//global/gateways/default-internet-gateway

description

An optional description of this resource. Provide this property when you create the resource.

tags

A list of instance tags to which this route applies.

next_hop_network

The URL of the local network if it should handle matching packets.

dest_range

The destination range of outgoing packets that this route applies to.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

next_hop_ip

The network IP address of an instance that should handle matching packets.

network

Fully-qualified URL of the network that this route applies to.

kind

[Output Only] Type of this resource. Always compute#routes for Route resources.

name

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.

next_hop_instance

The URL to an instance that should handle matching packets. You can specify this as a full or partial URL. For example: https://www.googleapis.com/compute/v1/projects/project/zones/zone/instances/

warnings

[Output Only] If potential misconfigurations are detected for this route, this field will be populated with warning messages.

priority

The priority of this route. Priority is used to break ties in cases where there is more than one matching route of equal prefix length. In the case of two routes with equal prefix length, the one with the lowest-numbered priority value wins. Default value is 1000. Valid range is 0 through 65535.

next_hop_vpn_tunnel

The URL to a VpnTunnel that should handle matching packets.

self_link

[Output Only] Server-defined fully-qualified URL for this resource.