Struct google_compute1::Snapshot

pub struct Snapshot {
    pub status: Option<String>,
    pub kind: Option<String>,
    pub storage_bytes: Option<String>,
    pub description: Option<String>,
    pub source_disk: Option<String>,
    pub creation_timestamp: Option<String>,
    pub source_disk_id: Option<String>,
    pub disk_size_gb: Option<String>,
    pub licenses: Option<Vec<String>>,
    pub storage_bytes_status: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

A persistent disk snapshot resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list snapshots (none)

create snapshot disks (request)

delete snapshots (none)

get snapshots (response)

Fields

status

[Output Only] The status of the snapshot. This can be CREATING, DELETING, FAILED, READY, or UPLOADING.

kind

[Output Only] Type of the resource. Always compute#snapshot for Snapshot resources.

storage_bytes

[Output Only] A size of the the storage used by the snapshot. As snapshots share storage, this number is expected to change with snapshot creation/deletion.

description

An optional description of this resource. Provide this property when you create the resource.

source_disk

[Output Only] The source disk used to create this snapshot.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

source_disk_id

[Output Only] The ID value of the disk used to create this snapshot. This value may be used to determine whether the snapshot was taken from the current or a previous instance of a given disk name.

disk_size_gb

[Output Only] Size of the snapshot, specified in GB.

licenses

[Output Only] A list of public visible licenses that apply to this snapshot. This can be because the original image had licenses attached (such as a Windows image).

storage_bytes_status

[Output Only] An indicator whether storageBytes is in a stable state or it is being adjusted as a result of shared storage reallocation. This status can either be UPDATING, meaning the size of the snapshot is being updated, or UP_TO_DATE, meaning the size of the snapshot is up-to-date.

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

self_link

[Output Only] Server-defined URL for the resource.

name

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.