Struct google_compute1::TargetInstance

pub struct TargetInstance {
    pub nat_policy: Option<String>,
    pub description: Option<String>,
    pub zone: Option<String>,
    pub kind: Option<String>,
    pub instance: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

A TargetInstance resource. This resource defines an endpoint VM that terminates traffic of certain protocols.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete target instances (none)

aggregated list target instances (none)

get target instances (response)

list target instances (none)

insert target instances (request)

Fields

nat_policy

NAT option controlling how IPs are NAT'ed to the VM. Currently only NO_NAT (default value) is supported.

description

An optional textual description of the resource; provided by the client when the resource is created.

zone

URL of the zone where the target instance resides (output only).

kind

Type of the resource.

instance

The URL to the instance that terminates the relevant traffic.

creation_timestamp

Creation timestamp in RFC3339 text format (output only).

id

Unique identifier for the resource; defined by the server (output only).

self_link

Server defined URL for the resource (output only).

name

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035.