Struct google_compute1::TargetPool

pub struct TargetPool {
    pub failover_ratio: Option<f32>,
    pub description: Option<String>,
    pub region: Option<String>,
    pub kind: Option<String>,
    pub session_affinity: Option<String>,
    pub instances: Option<Vec<String>>,
    pub backup_pool: Option<String>,
    pub health_checks: Option<Vec<String>>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

A TargetPool resource. This resource defines a pool of VMs, associated HttpHealthCheck resources, and the fallback TargetPool.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get health target pools (none)

add instance target pools (none)

remove instance target pools (none)

remove health check target pools (none)

delete target pools (none)

insert target pools (request)

aggregated list target pools (none)

list target pools (none)

add health check target pools (none)

get target pools (response)

set backup target pools (none)

Fields

failover_ratio

This field is applicable only when the containing target pool is serving a forwarding rule as the primary pool (i.e., not as a backup pool to some other target pool). The value of the field must be in [0, 1].

If set, 'backupPool' must also be set. They together define the fallback behavior of the primary target pool: if the ratio of the healthy VMs in the primary pool is at or below this number, traffic arriving at the load-balanced IP will be directed to the backup pool.

In case where 'failoverRatio' is not set or all the VMs in the backup pool are unhealthy, the traffic will be directed back to the primary pool in the "force" mode, where traffic will be spread to the healthy VMs with the best effort, or to all VMs when no VM is healthy.

description

An optional textual description of the resource; provided by the client when the resource is created.

region

URL of the region where the target pool resides (output only).

kind

Type of the resource.

session_affinity

Sesssion affinity option, must be one of the following values: 'NONE': Connections from the same client IP may go to any VM in the pool; 'CLIENT_IP': Connections from the same client IP will go to the same VM in the pool while that VM remains healthy. 'CLIENT_IP_PROTO': Connections from the same client IP with the same IP protocol will go to the same VM in the pool while that VM remains healthy.

instances

A list of resource URLs to the member VMs serving this pool. They must live in zones contained in the same region as this pool.

backup_pool

This field is applicable only when the containing target pool is serving a forwarding rule as the primary pool, and its 'failoverRatio' field is properly set to a value between [0, 1].

'backupPool' and 'failoverRatio' together define the fallback behavior of the primary target pool: if the ratio of the healthy VMs in the primary pool is at or below 'failoverRatio', traffic arriving at the load-balanced IP will be directed to the backup pool.

In case where 'failoverRatio' and 'backupPool' are not set, or all the VMs in the backup pool are unhealthy, the traffic will be directed back to the primary pool in the "force" mode, where traffic will be spread to the healthy VMs with the best effort, or to all VMs when no VM is healthy.

health_checks

A list of URLs to the HttpHealthCheck resource. A member VM in this pool is considered healthy if and only if all specified health checks pass. An empty list means all member VMs will be considered healthy at all times.

creation_timestamp

Creation timestamp in RFC3339 text format (output only).

id

Unique identifier for the resource; defined by the server (output only).

self_link

Server defined URL for the resource (output only).

name

Name of the resource; provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035.