Struct google_compute1::TargetPool

pub struct TargetPool {
    pub failover_ratio: Option<f32>,
    pub description: Option<String>,
    pub region: Option<String>,
    pub kind: Option<String>,
    pub session_affinity: Option<String>,
    pub instances: Option<Vec<String>>,
    pub backup_pool: Option<String>,
    pub health_checks: Option<Vec<String>>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

A TargetPool resource. This resource defines a pool of instances, associated HttpHealthCheck resources, and the fallback target pool.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get health target pools (none)

add instance target pools (none)

remove instance target pools (none)

remove health check target pools (none)

delete target pools (none)

insert target pools (request)

aggregated list target pools (none)

list target pools (none)

add health check target pools (none)

get target pools (response)

set backup target pools (none)

Fields

failover_ratio

This field is applicable only when the containing target pool is serving a forwarding rule as the primary pool (i.e., not as a backup pool to some other target pool). The value of the field must be in [0, 1].

If set, backupPool must also be set. They together define the fallback behavior of the primary target pool: if the ratio of the healthy instances in the primary pool is at or below this number, traffic arriving at the load-balanced IP will be directed to the backup pool.

In case where failoverRatio is not set or all the instances in the backup pool are unhealthy, the traffic will be directed back to the primary pool in the "force" mode, where traffic will be spread to the healthy instances with the best effort, or to all instances when no instance is healthy.

description

An optional description of this resource. Provide this property when you create the resource.

region

[Output Only] URL of the region where the target pool resides.

kind

[Output Only] Type of the resource. Always compute#targetPool for target pools.

session_affinity

Sesssion affinity option, must be one of the following values: NONE: Connections from the same client IP may go to any instance in the pool. CLIENT_IP: Connections from the same client IP will go to the same instance in the pool while that instance remains healthy. CLIENT_IP_PROTO: Connections from the same client IP with the same IP protocol will go to the same instance in the pool while that instance remains healthy.

instances

A list of resource URLs to the virtual machine instances serving this pool. They must live in zones contained in the same region as this pool.

backup_pool

This field is applicable only when the containing target pool is serving a forwarding rule as the primary pool, and its failoverRatio field is properly set to a value between [0, 1].

backupPool and failoverRatio together define the fallback behavior of the primary target pool: if the ratio of the healthy instances in the primary pool is at or below failoverRatio, traffic arriving at the load-balanced IP will be directed to the backup pool.

In case where failoverRatio and backupPool are not set, or all the instances in the backup pool are unhealthy, the traffic will be directed back to the primary pool in the "force" mode, where traffic will be spread to the healthy instances with the best effort, or to all instances when no instance is healthy.

health_checks

A list of URLs to the HttpHealthCheck resource. A member instance in this pool is considered healthy if and only if all specified health checks pass. An empty list means all member instances will be considered healthy at all times.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] The unique identifier for the resource. This identifier is defined by the server.

self_link

[Output Only] Server-defined URL for the resource.

name

Name of the resource. Provided by the client when the resource is created. The name must be 1-63 characters long, and comply with RFC1035. Specifically, the name must be 1-63 characters long and match the regular expression a-z? which means the first character must be a lowercase letter, and all following characters must be a dash, lowercase letter, or digit, except the last character, which cannot be a dash.