Struct google_compute1::Zone

pub struct Zone {
    pub status: Option<String>,
    pub kind: Option<String>,
    pub description: Option<String>,
    pub maintenance_windows: Option<Vec<ZoneMaintenanceWindows>>,
    pub deprecated: Option<DeprecationStatus>,
    pub region: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
}

A Zone resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list zones (none)

get zones (response)

Fields

status

[Output Only] Status of the zone, either UP or DOWN.

kind

[Output Only] Type of the resource. Always kind#zone for zones.

description

[Output Only] Textual description of the resource.

maintenance_windows

[Output Only] Any scheduled maintenance windows for this zone. When the zone is in a maintenance window, all resources which reside in the zone will be unavailable. For more information, see Maintenance Windows

deprecated

[Output Only] The deprecation status associated with this zone.

region

[Output Only] Full URL reference to the region which hosts the zone.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] Unique identifier for the resource; defined by the server.

self_link

[Output Only] Server defined URL for the resource.

name

[Output Only] Name of the resource.